Top F&O gainers and losers

Top futures and options gainers and losers for the last trading session.

nse_fo_top_gainers(clean_names = TRUE)

nse_fo_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_fo_top_gainers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 RAMCO~ EQ     2019-07-31       Annual General ~      755        793  
#>  2 SHREE~ EQ     2019-07-31       Dividend - Rs 3~    21097.     22050  
#>  3 VEDL   EQ     2019-07-04       Annual General ~      153.       157. 
#>  4 PFC    EQ     2019-08-20       Annual General ~      114        119. 
#>  5 L&TFH  EQ     2019-07-19       Annual General ~      114        117  
#>  6 MINDT~ EQ     2019-10-24       Interim Dividen~      823.       854  
#>  7 INDIGO EQ     2019-08-19       Annual General ~     1327.      1384  
#>  8 MOTHE~ EQ     2019-08-08       Annual General ~      143        147. 
#>  9 GMRIN~ EQ     2019-09-05       Annual General ~       22.8       23.8
#> 10 ZEEL   EQ     2019-07-15       Annual General ~      265        271. 
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_fo_top_gainers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 RAMCO~ EQ     2019-07-31       Annual General ~     755       793      753. 
#>  2 SHREE~ EQ     2019-07-31       Dividend - Rs 3~   21097.    22050    20951  
#>  3 VEDL   EQ     2019-07-04       Annual General ~     153.      157.     153. 
#>  4 PFC    EQ     2019-08-20       Annual General ~     114       119.     114  
#>  5 L&TFH  EQ     2019-07-19       Annual General ~     114       117      114. 
#>  6 MINDT~ EQ     2019-10-24       Interim Dividen~     823.      854      822  
#>  7 INDIGO EQ     2019-08-19       Annual General ~    1327.     1384     1327. 
#>  8 MOTHE~ EQ     2019-08-08       Annual General ~     143       147.     142. 
#>  9 GMRIN~ EQ     2019-09-05       Annual General ~      22.8      23.8     22.8
#> 10 ZEEL   EQ     2019-07-15       Annual General ~     265       271.     263. 
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# top losers
nse_fo_top_losers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 INFRA~ EQ     2019-12-19       Interim Dividen~       242.       247.
#>  2 BIOCON EQ     2019-07-18       Dividend - Rs 0~       293.       295.
#>  3 BPCL   EQ     2019-08-21       Dividend - Rs 8~       468.       475.
#>  4 NMDC   EQ     2019-08-22       Annual General ~       130        131.
#>  5 INFY   EQ     2019-10-23       Interim Dividen~       738        743.
#>  6 HAVEL~ EQ     2019-07-18       Annual General ~       648.       657.
#>  7 TORNT~ EQ     2019-06-13       Dividend Rs 4 P~      1886.      1910 
#>  8 CONCOR EQ     2019-08-19       Annual General ~       565.       568.
#>  9 BHART~ EQ     2019-04-23       Rights 19:67 @ ~       452        455.
#> 10 MCDOW~ EQ     2019-08-13       Annual General ~       572.       574.
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_fo_top_losers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 INFRA~ EQ     2019-12-19       Interim Dividen~      242.      247.     237.
#>  2 BIOCON EQ     2019-07-18       Dividend - Rs 0~      293.      295.     275.
#>  3 BPCL   EQ     2019-08-21       Dividend - Rs 8~      468.      475.     460.
#>  4 NMDC   EQ     2019-08-22       Annual General ~      130       131.     124.
#>  5 INFY   EQ     2019-10-23       Interim Dividen~      738       743.     725.
#>  6 HAVEL~ EQ     2019-07-18       Annual General ~      648.      657.     639 
#>  7 TORNT~ EQ     2019-06-13       Dividend Rs 4 P~     1886.     1910     1866.
#>  8 CONCOR EQ     2019-08-19       Annual General ~      565.      568.     555.
#>  9 BHART~ EQ     2019-04-23       Rights 19:67 @ ~      452       455.     443.
#> 10 MCDOW~ EQ     2019-08-13       Annual General ~      572.      574.     562.
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# }