Most actively traded stocks

Fetch most actively traded stocks in a month on NSE.

nse_stock_most_traded(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

security

Name of the firm.

share_turnover

Share in total turnover (percentage).

traded_quantity

Total traded quantity (in lakhs)

no_of_trades

Number of trades.

avg_daily_turnover

Average daily turnover (in crores).

turnover

Turnover (in crores.)

Examples

# \donttest{
nse_stock_most_traded()

#>                       security share_turnover traded_quantity no_of_trades
#> 1      Reliance Industries Ltd           6.82            4620      9245543
#> 2       Hindustan Unilever Ltd           6.25            3185      5752204
#> 3            Bajaj Finance Ltd           4.14            2057      8232909
#> 4                HDFC Bank Ltd           3.43            3832      7054716
#> 5               ICICI Bank Ltd           3.36           10513      9288909
#> 6                AXIS Bank Ltd           3.15            8265      8308649
#> 7            Bharti Airtel Ltd           3.00            5347      5774781
#> 8                     HDFC Ltd           2.46            1518      4924321
#> 9      Maruti Suzuki India Ltd           2.14             426      3320762
#> 10         State Bank Of India           2.03           12470      7371875
#> 11 TOTAL of Top Ten securities          36.78           52234     69274669
#> 12                       TOTAL         100.00          493359    341796702
#>    avg_daily_turnonver   turnover
#> 1              3591.40   68236.63
#> 2              3292.50   62557.46
#> 3              2179.74   41415.04
#> 4              1808.33   34358.27
#> 5              1767.89   33589.97
#> 6              1658.21   31506.01
#> 7              1577.88   29979.79
#> 8              1293.73   24580.89
#> 9              1125.83   21390.76
#> 10             1069.64   20323.16
#> 11            19365.16  367937.97
#> 12            52655.61 1000456.50

# retain original column names as returned by NSE
nse_stock_most_traded(clean_names = FALSE)

#>                        skipped sharetotal trdQty nooftrades avgdailyturn
#> 1      Reliance Industries Ltd       6.82   4620    9245543      3591.40
#> 2       Hindustan Unilever Ltd       6.25   3185    5752204      3292.50
#> 3            Bajaj Finance Ltd       4.14   2057    8232909      2179.74
#> 4                HDFC Bank Ltd       3.43   3832    7054716      1808.33
#> 5               ICICI Bank Ltd       3.36  10513    9288909      1767.89
#> 6                AXIS Bank Ltd       3.15   8265    8308649      1658.21
#> 7            Bharti Airtel Ltd       3.00   5347    5774781      1577.88
#> 8                     HDFC Ltd       2.46   1518    4924321      1293.73
#> 9      Maruti Suzuki India Ltd       2.14    426    3320762      1125.83
#> 10         State Bank Of India       2.03  12470    7371875      1069.64
#> 11 TOTAL of Top Ten securities      36.78  52234   69274669     19365.16
#> 12                       TOTAL     100.00 493359  341796702     52655.61
#>      turnover
#> 1    68236.63
#> 2    62557.46
#> 3    41415.04
#> 4    34358.27
#> 5    33589.97
#> 6    31506.01
#> 7    29979.79
#> 8    24580.89
#> 9    21390.76
#> 10   20323.16
#> 11  367937.97
#> 12 1000456.50

# }