NSE top gainers & losers

Fetch top gainers and losers for the last trading session.

nse_stock_top_gainers(clean_names = TRUE)

nse_stock_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_stock_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1       CIPLA     EQ                  2020-08-11
#> 2         M&M     EQ                  2020-07-16
#> 3        BPCL     EQ                  2021-03-25
#> 4  INDUSINDBK     EQ                  2019-08-08
#> 5       TITAN     EQ                  2020-08-03
#> 6        SBIN     EQ                  2018-06-15
#> 7          LT     EQ                  2020-11-04
#> 8        INFY     EQ                  2021-05-31
#> 9  BAJAJFINSV     EQ                  2020-03-03
#> 10   HDFCLIFE     EQ                  2019-07-11
#>                                            last_corp_announcement open_price
#> 1                                          Annual General Meeting     906.00
#> 2             Annual General Meeting/Dividend - Rs 2.35 Per Share     784.05
#> 3                               Interim Dividend - Rs 5 Per Share     448.00
#> 4             Annual General Meeting/Dividend - Rs 7.50 Per Share     963.00
#> 5                Annual General Meeting/Dividend - Rs 4 Per Share    1523.00
#> 6  Annual General Meeting/ Change In Registrar And Transfer Agent     387.00
#> 7                              Special Dividend - Rs 18 Per Share    1410.80
#> 8               Annual General Meeting/Dividend - Rs 15 Per Share    1348.40
#> 9                               Interim Dividend - Rs 5 Per Share   11260.00
#> 10                                         Annual General Meeting     669.00
#>    high_price low_price last_traded_price prev_close_price percent_change
#> 1      927.90    905.00            925.50           903.95           2.38
#> 2      809.55    782.00            803.00           784.80           2.32
#> 3      457.85    441.50            455.80           446.45           2.09
#> 4      987.90    963.00            976.50           967.00           0.98
#> 5     1566.25   1522.50           1535.50          1521.15           0.94
#> 6      387.80    382.00            386.80           384.40           0.62
#> 7     1426.95   1401.65           1409.35          1403.45           0.42
#> 8     1349.10   1335.00           1341.95          1337.00           0.37
#> 9    11299.95  11130.70          11250.05         11215.15           0.31
#> 10     679.00    665.70            667.85           666.15           0.26
#>    traded_quantity turnover_in_lakhs
#> 1         12378690         113846.81
#> 2         10092542          80881.63
#> 3          9399907          42447.16
#> 4          5835440          57058.35
#> 5          3032029          46950.67
#> 6         42898608         164910.83
#> 7          1996254          28255.38
#> 8          4490092          60203.15
#> 9           177897          19964.56
#> 10         4316905          29073.92

# retain original column names as returned by NSE
nse_stock_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1       CIPLA     EQ               2020-08-11
#> 2         M&M     EQ               2020-07-16
#> 3        BPCL     EQ               2021-03-25
#> 4  INDUSINDBK     EQ               2019-08-08
#> 5       TITAN     EQ               2020-08-03
#> 6        SBIN     EQ               2018-06-15
#> 7          LT     EQ               2020-11-04
#> 8        INFY     EQ               2021-05-31
#> 9  BAJAJFINSV     EQ               2020-03-03
#> 10   HDFCLIFE     EQ               2019-07-11
#>                                              lastCorpAnnouncement openPrice
#> 1                                          Annual General Meeting    906.00
#> 2             Annual General Meeting/Dividend - Rs 2.35 Per Share    784.05
#> 3                               Interim Dividend - Rs 5 Per Share    448.00
#> 4             Annual General Meeting/Dividend - Rs 7.50 Per Share    963.00
#> 5                Annual General Meeting/Dividend - Rs 4 Per Share   1523.00
#> 6  Annual General Meeting/ Change In Registrar And Transfer Agent    387.00
#> 7                              Special Dividend - Rs 18 Per Share   1410.80
#> 8               Annual General Meeting/Dividend - Rs 15 Per Share   1348.40
#> 9                               Interim Dividend - Rs 5 Per Share  11260.00
#> 10                                         Annual General Meeting    669.00
#>    highPrice lowPrice      ltp previousPrice netPrice tradedQuantity
#> 1     927.90   905.00   925.50        903.95     2.38       12378690
#> 2     809.55   782.00   803.00        784.80     2.32       10092542
#> 3     457.85   441.50   455.80        446.45     2.09        9399907
#> 4     987.90   963.00   976.50        967.00     0.98        5835440
#> 5    1566.25  1522.50  1535.50       1521.15     0.94        3032029
#> 6     387.80   382.00   386.80        384.40     0.62       42898608
#> 7    1426.95  1401.65  1409.35       1403.45     0.42        1996254
#> 8    1349.10  1335.00  1341.95       1337.00     0.37        4490092
#> 9   11299.95 11130.70 11250.05      11215.15     0.31         177897
#> 10    679.00   665.70   667.85        666.15     0.26        4316905
#>    turnoverInLakhs
#> 1        113846.81
#> 2         80881.63
#> 3         42447.16
#> 4         57058.35
#> 5         46950.67
#> 6        164910.83
#> 7         28255.38
#> 8         60203.15
#> 9         19964.56
#> 10        29073.92

# top losers
nse_stock_top_losers()

#>        symbol series last_corp_announcement_date
#> 1   TATASTEEL     EQ                  2021-06-17
#> 2    HINDALCO     EQ                  2020-09-03
#> 3   COALINDIA     EQ                  2021-03-15
#> 4   BRITANNIA     EQ                  2021-05-25
#> 5        ONGC     EQ                  2021-02-17
#> 6         IOC     EQ                  2021-03-23
#> 7   SUNPHARMA     EQ                  2021-02-09
#> 8  TATAMOTORS     EQ                  2016-07-18
#> 9   POWERGRID     EQ                  2021-03-08
#> 10 HEROMOTOCO     EQ                  2021-02-15
#>                                                  last_corp_announcement
#> 1                     Annual General Meeting/Dividend - Rs 25 Per Share
#> 2                      Annual General Meeting/Dividend - Re 1 Per Share
#> 3                                     Interim Dividend - Rs 5 Per Share
#> 4     Scheme Of Arangement- Bonus - 1 Debenture For 1 Equity Share Held
#> 5                                  Interim Dividend - Rs 1.75 Per Share
#> 6                                     Interim Dividend - Rs 3 Per Share
#> 7                                  Interim Dividend - Rs 5.50 Per Share
#> 8                                        Dividend - Re 0.20/- Per Share
#> 9                                     Interim Dividend - Rs 4 Per Share
#> 10 Interim Dividend - Rs 65 Per Share Special Dividend - Rs 5 Per Share
#>    open_price high_price low_price last_traded_price prev_close_price
#> 1     1131.10    1135.50   1098.05           1104.90          1164.25
#> 2      390.00     393.20    379.45            386.20           403.30
#> 3      152.40     152.40    146.05            146.65           151.85
#> 4     3500.00    3517.70   3415.00           3428.90          3534.40
#> 5      113.45     114.70    111.20            111.60           114.90
#> 6      109.70     109.70    104.50            104.75           107.20
#> 7      700.00     701.65    685.00            688.30           703.40
#> 8      316.00     316.60    306.10            308.00           314.45
#> 9      234.00     234.95    228.40            229.45           233.75
#> 10    2911.35    2916.80   2861.75           2867.90          2915.00
#>    percent_change traded_quantity turnover_in_lakhs
#> 1           -5.10        28446253         316697.82
#> 2           -4.24        30689219         118810.24
#> 3           -3.42        22983457          34192.49
#> 4           -2.98         1046540          36059.48
#> 5           -2.87        21131045          23738.62
#> 6           -2.29        56482450          59888.34
#> 7           -2.15         6226376          43044.81
#> 8           -2.05        54043899         167784.69
#> 9           -1.84         6022766          13926.44
#> 10          -1.62          637388          18376.41

# retain original column names as returned by NSE
nse_stock_top_losers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1   TATASTEEL     EQ               2021-06-17
#> 2    HINDALCO     EQ               2020-09-03
#> 3   COALINDIA     EQ               2021-03-15
#> 4   BRITANNIA     EQ               2021-05-25
#> 5        ONGC     EQ               2021-02-17
#> 6         IOC     EQ               2021-03-23
#> 7   SUNPHARMA     EQ               2021-02-09
#> 8  TATAMOTORS     EQ               2016-07-18
#> 9   POWERGRID     EQ               2021-03-08
#> 10 HEROMOTOCO     EQ               2021-02-15
#>                                                    lastCorpAnnouncement
#> 1                     Annual General Meeting/Dividend - Rs 25 Per Share
#> 2                      Annual General Meeting/Dividend - Re 1 Per Share
#> 3                                     Interim Dividend - Rs 5 Per Share
#> 4     Scheme Of Arangement- Bonus - 1 Debenture For 1 Equity Share Held
#> 5                                  Interim Dividend - Rs 1.75 Per Share
#> 6                                     Interim Dividend - Rs 3 Per Share
#> 7                                  Interim Dividend - Rs 5.50 Per Share
#> 8                                        Dividend - Re 0.20/- Per Share
#> 9                                     Interim Dividend - Rs 4 Per Share
#> 10 Interim Dividend - Rs 65 Per Share Special Dividend - Rs 5 Per Share
#>    openPrice highPrice lowPrice     ltp previousPrice netPrice tradedQuantity
#> 1    1131.10   1135.50  1098.05 1104.90       1164.25    -5.10       28446253
#> 2     390.00    393.20   379.45  386.20        403.30    -4.24       30689219
#> 3     152.40    152.40   146.05  146.65        151.85    -3.42       22983457
#> 4    3500.00   3517.70  3415.00 3428.90       3534.40    -2.98        1046540
#> 5     113.45    114.70   111.20  111.60        114.90    -2.87       21131045
#> 6     109.70    109.70   104.50  104.75        107.20    -2.29       56482450
#> 7     700.00    701.65   685.00  688.30        703.40    -2.15        6226376
#> 8     316.00    316.60   306.10  308.00        314.45    -2.05       54043899
#> 9     234.00    234.95   228.40  229.45        233.75    -1.84        6022766
#> 10   2911.35   2916.80  2861.75 2867.90       2915.00    -1.62         637388
#>    turnoverInLakhs
#> 1        316697.82
#> 2        118810.24
#> 3         34192.49
#> 4         36059.48
#> 5         23738.62
#> 6         59888.34
#> 7         43044.81
#> 8        167784.69
#> 9         13926.44
#> 10        18376.41

# }