Pre Open market data

Fetch data of pre open session of Nifty & Nifty Bank.

nse_preopen_nifty(clean_names = TRUE)

nse_preopen_nifty_bank(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ)

corp_action_date

Corporate action date.

corp_action

Corporate Action

price

Price

change

Change in price

percent_change

Percentage change in price.

prev_close

Previous close.

quantity

Quantity

value

Value (in lakhs),

mkt_cap

Free float market capitalization (in crores).

year_high

Normal market 52 week high.

year_low

Normal market 52 week low.

Examples

# \donttest{
# nifty
nse_preopen_nifty()

#>        symbol series corp_action_date
#> 1    HINDALCO     EQ             <NA>
#> 2   TATASTEEL     EQ             <NA>
#> 3        ONGC     EQ             <NA>
#> 4    AXISBANK     EQ             <NA>
#> 5    JSWSTEEL     EQ             <NA>
#> 6         IOC     EQ       2021-03-23
#> 7   BRITANNIA     EQ             <NA>
#> 8   SUNPHARMA     EQ             <NA>
#> 9         TCS     EQ             <NA>
#> 10 INDUSINDBK     EQ             <NA>
#> 11 HEROMOTOCO     EQ             <NA>
#> 12        M&M     EQ             <NA>
#> 13 ASIANPAINT     EQ             <NA>
#> 14 BAJFINANCE     EQ             <NA>
#> 15     GRASIM     EQ             <NA>
#> 16       INFY     EQ             <NA>
#> 17      WIPRO     EQ             <NA>
#> 18      TECHM     EQ             <NA>
#> 19       SBIN     EQ             <NA>
#> 20   SHREECEM     EQ             <NA>
#> 21 ULTRACEMCO     EQ             <NA>
#> 22        UPL     EQ             <NA>
#> 23 ADANIPORTS     EQ             <NA>
#> 24 BHARTIARTL     EQ             <NA>
#> 25         LT     EQ             <NA>
#> 26 TATAMOTORS     EQ             <NA>
#> 27 TATACONSUM     EQ             <NA>
#> 28   HDFCLIFE     EQ             <NA>
#> 29  ICICIBANK     EQ             <NA>
#> 30 BAJAJFINSV     EQ             <NA>
#> 31  COALINDIA     EQ       2021-03-15
#> 32    SBILIFE     EQ       2021-04-05
#> 33       BPCL     EQ       2021-03-25
#> 34        ITC     EQ       2021-02-22
#> 35    DRREDDY     EQ             <NA>
#> 36       HDFC     EQ             <NA>
#> 37   DIVISLAB     EQ             <NA>
#> 38      CIPLA     EQ             <NA>
#> 39      TITAN     EQ             <NA>
#> 40  POWERGRID     EQ       2021-03-08
#> 41  KOTAKBANK     EQ             <NA>
#> 42     MARUTI     EQ             <NA>
#> 43    HCLTECH     EQ       2021-04-29
#> 44  NESTLEIND     EQ       2021-04-29
#> 45   RELIANCE     EQ             <NA>
#> 46       NTPC     EQ             <NA>
#> 47 HINDUNILVR     EQ             <NA>
#> 48  EICHERMOT     EQ             <NA>
#> 49   HDFCBANK     EQ             <NA>
#> 50 BAJAJ-AUTO     EQ             <NA>
#>                                                                     corp_action
#> 1                                                                          <NA>
#> 2                                                                          <NA>
#> 3                                                                          <NA>
#> 4                                                                          <NA>
#> 5                                                                          <NA>
#> 6                                             INTERIM DIVIDEND - RS 3 PER SHARE
#> 7                                                                          <NA>
#> 8                                                                          <NA>
#> 9                                                                          <NA>
#> 10                                                                         <NA>
#> 11                                                                         <NA>
#> 12                                                                         <NA>
#> 13                                                                         <NA>
#> 14                                                                         <NA>
#> 15                                                                         <NA>
#> 16                                                                         <NA>
#> 17                                                                         <NA>
#> 18                                                                         <NA>
#> 19                                                                         <NA>
#> 20                                                                         <NA>
#> 21                                                                         <NA>
#> 22                                                                         <NA>
#> 23                                                                         <NA>
#> 24                                                                         <NA>
#> 25                                                                         <NA>
#> 26                                                                         <NA>
#> 27                                                                         <NA>
#> 28                                                                         <NA>
#> 29                                                                         <NA>
#> 30                                                                         <NA>
#> 31                                            INTERIM DIVIDEND - RS 5 PER SHARE
#> 32                                          INTERIM DIVIDEND - RS 2.5 PER SHARE
#> 33                                            INTERIM DIVIDEND - RS 5 PER SHARE
#> 34                                            INTERIM DIVIDEND - RS 5 PER SHARE
#> 35                                                                         <NA>
#> 36                                                                         <NA>
#> 37                                                                         <NA>
#> 38                                                                         <NA>
#> 39                                                                         <NA>
#> 40                                            INTERIM DIVIDEND - RS 4 PER SHARE
#> 41                                                                         <NA>
#> 42                                                                         <NA>
#> 43 INTERIM DIVIDEND - RS 6 PER SHARE SPECIAL INTERIM DIVIDEND - RS 10 PER SHARE
#> 44                            ANNUAL GENERAL MEETING/DIVIDEND - RS 65 PER SHARE
#> 45                                                                         <NA>
#> 46                                                                         <NA>
#> 47                                                                         <NA>
#> 48                                                                         <NA>
#> 49                                                                         <NA>
#> 50                                                                         <NA>
#>       price change percent_change prev_close quantity   value   mkt_cap
#> 1    390.00 -13.30          -3.30     403.30    96605  376.76  56252.48
#> 2   1131.10 -33.15          -2.85    1164.25    90968 1028.94  85948.80
#> 3    113.45  -1.45          -1.26     114.90    72462   82.21  40787.78
#> 4    708.60  -8.20          -1.14     716.80   263562 1867.60 177320.19
#> 5    698.00  -7.85          -1.11     705.85    30902  215.70  67150.38
#> 6    109.70   2.50           2.33     107.20   450551  494.25  26612.89
#> 7   3500.00 -34.40          -0.97    3534.40     3141  109.94  40394.81
#> 8    700.00  -3.40          -0.48     703.40    21292  149.04  74159.25
#> 9   3067.10 -14.90          -0.48    3082.00    56110 1720.95 316934.72
#> 10   963.00  -4.00          -0.41     967.00     6893   66.38  55057.54
#> 11  2911.35  -3.65          -0.13    2915.00     1166   33.95  37205.37
#> 12   784.05  -0.75          -0.10     784.80     3828   30.01  76891.77
#> 13  2814.00  -1.85          -0.07    2815.85     2478   69.73 125964.44
#> 14  5628.00  -3.20          -0.06    5631.20     3990  224.56 148179.24
#> 15  1389.00  -0.70          -0.05    1389.70     1232   17.11  52481.13
#> 16  1348.40  11.40           0.85    1337.00    31039  418.53 496358.32
#> 17   516.00   4.35           0.85     511.65    18484   95.38  75188.46
#> 18   984.80   7.15           0.73     977.65     6011   59.20  60301.69
#> 19   387.00   2.60           0.68     384.40   136445  528.04 147574.25
#> 20 27700.00 167.30           0.61   27532.70      106   29.36  36725.98
#> 21  6600.00  38.80           0.59    6561.20      687   45.34  75899.42
#> 22   786.40   4.25           0.54     782.15    17419  136.98  42221.15
#> 23   776.50   4.10           0.53     772.40     9489   73.68  56005.64
#> 24   531.00   2.80           0.53     528.20    23795  126.35 125814.66
#> 25  1410.80   7.35           0.52    1403.45     1816   25.62 170271.61
#> 26   316.00   1.55           0.49     314.45   132679  419.27  55169.57
#> 27   656.00   2.80           0.43     653.20     3895   25.55  38630.07
#> 28   669.00   2.85           0.43     666.15     2177   14.56  55308.18
#> 29   628.05   2.60           0.42     625.45    38689  242.99 427790.01
#> 30 11260.00  44.85           0.40   11215.15      226   25.45  67897.61
#> 31   152.40   0.55           0.36     151.85    46137   70.31  30727.98
#> 32   981.45   3.40           0.35     978.05     2530   24.83  37799.49
#> 33   448.00   1.55           0.35     446.45     9260   41.48  42422.77
#> 34   208.45   0.70           0.34     207.75    40535   84.50 180888.76
#> 35  5255.00  17.20           0.33    5237.80     2345  123.23  63568.30
#> 36  2467.00   7.15           0.29    2459.85    14921  368.10 440396.04
#> 37  4049.65  10.00           0.25    4039.65      433   17.53  51489.86
#> 38   906.00   2.05           0.23     903.95    16921  153.30  46955.84
#> 39  1523.00   1.85           0.12    1521.15     1778   27.08  63897.03
#> 40   234.00   0.25           0.11     233.75    16059   37.58  58690.85
#> 41  1732.00   1.45           0.08    1730.55     5421   93.89 250490.47
#> 42  6780.80   4.90           0.07    6775.90      571   38.72  89556.29
#> 43   932.00   0.60           0.06     931.40    10617   98.95 100340.48
#> 44 17460.00  10.65           0.06   17449.35      141   24.62  61663.44
#> 45  1997.90   0.85           0.04    1997.05    12886  257.45 641902.73
#> 46   112.00   0.05           0.04     111.95    18792   21.05  53096.52
#> 47  2360.00   0.25           0.01    2359.75      881   20.79 209293.81
#> 48  2565.00   0.30           0.01    2564.70     1255   32.19  35276.90
#> 49  1458.35   0.15           0.01    1458.20    13685  199.58 623997.79
#> 50  4095.95   0.00           0.00    4095.95     2480  101.58  53013.20
#>    year_high year_low  sum_val sum_quantity fin_quantity sum_fin_quantity
#> 1     427.50   116.25 10590.19      1090006        96605          1745785
#> 2    1246.85   267.50 10590.19      1090006        90968          1745785
#> 3     122.35    64.10 10590.19      1090006        72462          1745785
#> 4     799.00   332.85 10590.19      1090006       263562          1745785
#> 5     773.00   163.50 10590.19      1090006        30902          1745785
#> 6     109.70    71.65 10590.19      1090006       450551          1745785
#> 7    4010.00  3080.00 10590.19      1090006         3141          1745785
#> 8     721.85   437.25 10590.19      1090006        21292          1745785
#> 9    3354.35  1903.00 10590.19      1090006        56110          1745785
#> 10   1119.50   330.00 10590.19      1090006         6893          1745785
#> 11   3629.05  1991.35 10590.19      1090006         1166          1745785
#> 12    952.05   371.20 10590.19      1090006         3828          1745785
#> 13   2873.45  1482.95 10590.19      1090006         2478          1745785
#> 14   5822.20  1783.00 10590.19      1090006         3990          1745785
#> 15   1475.00   495.05 10590.19      1090006         1232          1745785
#> 16   1477.55   655.00 10590.19      1090006        31039          1745785
#> 17    528.50   178.00 10590.19      1090006        18484          1745785
#> 18   1081.55   503.30 10590.19      1090006         6011          1745785
#> 19    427.70   149.45 10590.19      1090006       136445          1745785
#> 20  32048.00 18075.00 10590.19      1090006          106          1745785
#> 21   7055.95  3335.25 10590.19      1090006          687          1745785
#> 22    805.30   336.00 10590.19      1090006        17419          1745785
#> 23    885.00   297.35 10590.19      1090006         9489          1745785
#> 24    623.00   394.00 10590.19      1090006        23795          1745785
#> 25   1593.00   791.70 10590.19      1090006         1816          1745785
#> 26    357.00    80.70 10590.19      1090006       132679          1745785
#> 27    698.00   349.25 10590.19      1090006         3895          1745785
#> 28    746.00   465.25 10590.19      1090006         2177          1745785
#> 29    679.40   285.55 10590.19      1090006        38689          1745785
#> 30  11589.00  3985.30 10590.19      1090006          226          1745785
#> 31    163.00   109.55 10590.19      1090006        46137          1745785
#> 32   1044.70   698.00 10590.19      1090006         2530          1745785
#> 33    482.40   291.80 10590.19      1090006         9260          1745785
#> 34    239.20   163.35 10590.19      1090006        40535          1745785
#> 35   5512.65  3660.10 10590.19      1090006         2345          1745785
#> 36   2896.00  1486.00 10590.19      1090006        14921          1745785
#> 37   4205.00  2090.05 10590.19      1090006          433          1745785
#> 38    966.35   590.20 10590.19      1090006        16921          1745785
#> 39   1621.35   818.15 10590.19      1090006         1778          1745785
#> 40    239.85   154.00 10590.19      1090006        16059          1745785
#> 41   2049.00  1115.00 10590.19      1090006         5421          1745785
#> 42   8329.00  4745.50 10590.19      1090006          571          1745785
#> 43   1067.00   506.20 10590.19      1090006        10617          1745785
#> 44  18844.00 15100.00 10590.19      1090006          141          1745785
#> 45   2369.35  1393.00 10590.19      1090006        12886          1745785
#> 46    116.75    78.10 10590.19      1090006        18792          1745785
#> 47   2505.90  1955.20 10590.19      1090006          881          1745785
#> 48   3037.00  1286.80 10590.19      1090006         1255          1745785
#> 49   1641.00   826.10 10590.19      1090006        13685          1745785
#> 50   4361.40  2450.00 10590.19      1090006         2480          1745785

# retain original column names as returned by NSE
nse_preopen_nifty(clean_names = FALSE)

#>        symbol series        xDt
#> 1    HINDALCO     EQ       <NA>
#> 2   TATASTEEL     EQ       <NA>
#> 3        ONGC     EQ       <NA>
#> 4    AXISBANK     EQ       <NA>
#> 5    JSWSTEEL     EQ       <NA>
#> 6         IOC     EQ 2021-03-23
#> 7   BRITANNIA     EQ       <NA>
#> 8   SUNPHARMA     EQ       <NA>
#> 9         TCS     EQ       <NA>
#> 10 INDUSINDBK     EQ       <NA>
#> 11 HEROMOTOCO     EQ       <NA>
#> 12        M&M     EQ       <NA>
#> 13 ASIANPAINT     EQ       <NA>
#> 14 BAJFINANCE     EQ       <NA>
#> 15     GRASIM     EQ       <NA>
#> 16       INFY     EQ       <NA>
#> 17      WIPRO     EQ       <NA>
#> 18      TECHM     EQ       <NA>
#> 19       SBIN     EQ       <NA>
#> 20   SHREECEM     EQ       <NA>
#> 21 ULTRACEMCO     EQ       <NA>
#> 22        UPL     EQ       <NA>
#> 23 ADANIPORTS     EQ       <NA>
#> 24 BHARTIARTL     EQ       <NA>
#> 25         LT     EQ       <NA>
#> 26 TATAMOTORS     EQ       <NA>
#> 27 TATACONSUM     EQ       <NA>
#> 28   HDFCLIFE     EQ       <NA>
#> 29  ICICIBANK     EQ       <NA>
#> 30 BAJAJFINSV     EQ       <NA>
#> 31  COALINDIA     EQ 2021-03-15
#> 32    SBILIFE     EQ 2021-04-05
#> 33       BPCL     EQ 2021-03-25
#> 34        ITC     EQ 2021-02-22
#> 35    DRREDDY     EQ       <NA>
#> 36       HDFC     EQ       <NA>
#> 37   DIVISLAB     EQ       <NA>
#> 38      CIPLA     EQ       <NA>
#> 39      TITAN     EQ       <NA>
#> 40  POWERGRID     EQ 2021-03-08
#> 41  KOTAKBANK     EQ       <NA>
#> 42     MARUTI     EQ       <NA>
#> 43    HCLTECH     EQ 2021-04-29
#> 44  NESTLEIND     EQ 2021-04-29
#> 45   RELIANCE     EQ       <NA>
#> 46       NTPC     EQ       <NA>
#> 47 HINDUNILVR     EQ       <NA>
#> 48  EICHERMOT     EQ       <NA>
#> 49   HDFCBANK     EQ       <NA>
#> 50 BAJAJ-AUTO     EQ       <NA>
#>                                                                           caAct
#> 1                                                                          <NA>
#> 2                                                                          <NA>
#> 3                                                                          <NA>
#> 4                                                                          <NA>
#> 5                                                                          <NA>
#> 6                                             INTERIM DIVIDEND - RS 3 PER SHARE
#> 7                                                                          <NA>
#> 8                                                                          <NA>
#> 9                                                                          <NA>
#> 10                                                                         <NA>
#> 11                                                                         <NA>
#> 12                                                                         <NA>
#> 13                                                                         <NA>
#> 14                                                                         <NA>
#> 15                                                                         <NA>
#> 16                                                                         <NA>
#> 17                                                                         <NA>
#> 18                                                                         <NA>
#> 19                                                                         <NA>
#> 20                                                                         <NA>
#> 21                                                                         <NA>
#> 22                                                                         <NA>
#> 23                                                                         <NA>
#> 24                                                                         <NA>
#> 25                                                                         <NA>
#> 26                                                                         <NA>
#> 27                                                                         <NA>
#> 28                                                                         <NA>
#> 29                                                                         <NA>
#> 30                                                                         <NA>
#> 31                                            INTERIM DIVIDEND - RS 5 PER SHARE
#> 32                                          INTERIM DIVIDEND - RS 2.5 PER SHARE
#> 33                                            INTERIM DIVIDEND - RS 5 PER SHARE
#> 34                                            INTERIM DIVIDEND - RS 5 PER SHARE
#> 35                                                                         <NA>
#> 36                                                                         <NA>
#> 37                                                                         <NA>
#> 38                                                                         <NA>
#> 39                                                                         <NA>
#> 40                                            INTERIM DIVIDEND - RS 4 PER SHARE
#> 41                                                                         <NA>
#> 42                                                                         <NA>
#> 43 INTERIM DIVIDEND - RS 6 PER SHARE SPECIAL INTERIM DIVIDEND - RS 10 PER SHARE
#> 44                            ANNUAL GENERAL MEETING/DIVIDEND - RS 65 PER SHARE
#> 45                                                                         <NA>
#> 46                                                                         <NA>
#> 47                                                                         <NA>
#> 48                                                                         <NA>
#> 49                                                                         <NA>
#> 50                                                                         <NA>
#>         iep    chn perChn     pCls trdQnty    iVal    mktCap    yHigh     yLow
#> 1    390.00 -13.30  -3.30   403.30   96605  376.76  56252.48   427.50   116.25
#> 2   1131.10 -33.15  -2.85  1164.25   90968 1028.94  85948.80  1246.85   267.50
#> 3    113.45  -1.45  -1.26   114.90   72462   82.21  40787.78   122.35    64.10
#> 4    708.60  -8.20  -1.14   716.80  263562 1867.60 177320.19   799.00   332.85
#> 5    698.00  -7.85  -1.11   705.85   30902  215.70  67150.38   773.00   163.50
#> 6    109.70   2.50   2.33   107.20  450551  494.25  26612.89   109.70    71.65
#> 7   3500.00 -34.40  -0.97  3534.40    3141  109.94  40394.81  4010.00  3080.00
#> 8    700.00  -3.40  -0.48   703.40   21292  149.04  74159.25   721.85   437.25
#> 9   3067.10 -14.90  -0.48  3082.00   56110 1720.95 316934.72  3354.35  1903.00
#> 10   963.00  -4.00  -0.41   967.00    6893   66.38  55057.54  1119.50   330.00
#> 11  2911.35  -3.65  -0.13  2915.00    1166   33.95  37205.37  3629.05  1991.35
#> 12   784.05  -0.75  -0.10   784.80    3828   30.01  76891.77   952.05   371.20
#> 13  2814.00  -1.85  -0.07  2815.85    2478   69.73 125964.44  2873.45  1482.95
#> 14  5628.00  -3.20  -0.06  5631.20    3990  224.56 148179.24  5822.20  1783.00
#> 15  1389.00  -0.70  -0.05  1389.70    1232   17.11  52481.13  1475.00   495.05
#> 16  1348.40  11.40   0.85  1337.00   31039  418.53 496358.32  1477.55   655.00
#> 17   516.00   4.35   0.85   511.65   18484   95.38  75188.46   528.50   178.00
#> 18   984.80   7.15   0.73   977.65    6011   59.20  60301.69  1081.55   503.30
#> 19   387.00   2.60   0.68   384.40  136445  528.04 147574.25   427.70   149.45
#> 20 27700.00 167.30   0.61 27532.70     106   29.36  36725.98 32048.00 18075.00
#> 21  6600.00  38.80   0.59  6561.20     687   45.34  75899.42  7055.95  3335.25
#> 22   786.40   4.25   0.54   782.15   17419  136.98  42221.15   805.30   336.00
#> 23   776.50   4.10   0.53   772.40    9489   73.68  56005.64   885.00   297.35
#> 24   531.00   2.80   0.53   528.20   23795  126.35 125814.66   623.00   394.00
#> 25  1410.80   7.35   0.52  1403.45    1816   25.62 170271.61  1593.00   791.70
#> 26   316.00   1.55   0.49   314.45  132679  419.27  55169.57   357.00    80.70
#> 27   656.00   2.80   0.43   653.20    3895   25.55  38630.07   698.00   349.25
#> 28   669.00   2.85   0.43   666.15    2177   14.56  55308.18   746.00   465.25
#> 29   628.05   2.60   0.42   625.45   38689  242.99 427790.01   679.40   285.55
#> 30 11260.00  44.85   0.40 11215.15     226   25.45  67897.61 11589.00  3985.30
#> 31   152.40   0.55   0.36   151.85   46137   70.31  30727.98   163.00   109.55
#> 32   981.45   3.40   0.35   978.05    2530   24.83  37799.49  1044.70   698.00
#> 33   448.00   1.55   0.35   446.45    9260   41.48  42422.77   482.40   291.80
#> 34   208.45   0.70   0.34   207.75   40535   84.50 180888.76   239.20   163.35
#> 35  5255.00  17.20   0.33  5237.80    2345  123.23  63568.30  5512.65  3660.10
#> 36  2467.00   7.15   0.29  2459.85   14921  368.10 440396.04  2896.00  1486.00
#> 37  4049.65  10.00   0.25  4039.65     433   17.53  51489.86  4205.00  2090.05
#> 38   906.00   2.05   0.23   903.95   16921  153.30  46955.84   966.35   590.20
#> 39  1523.00   1.85   0.12  1521.15    1778   27.08  63897.03  1621.35   818.15
#> 40   234.00   0.25   0.11   233.75   16059   37.58  58690.85   239.85   154.00
#> 41  1732.00   1.45   0.08  1730.55    5421   93.89 250490.47  2049.00  1115.00
#> 42  6780.80   4.90   0.07  6775.90     571   38.72  89556.29  8329.00  4745.50
#> 43   932.00   0.60   0.06   931.40   10617   98.95 100340.48  1067.00   506.20
#> 44 17460.00  10.65   0.06 17449.35     141   24.62  61663.44 18844.00 15100.00
#> 45  1997.90   0.85   0.04  1997.05   12886  257.45 641902.73  2369.35  1393.00
#> 46   112.00   0.05   0.04   111.95   18792   21.05  53096.52   116.75    78.10
#> 47  2360.00   0.25   0.01  2359.75     881   20.79 209293.81  2505.90  1955.20
#> 48  2565.00   0.30   0.01  2564.70    1255   32.19  35276.90  3037.00  1286.80
#> 49  1458.35   0.15   0.01  1458.20   13685  199.58 623997.79  1641.00   826.10
#> 50  4095.95   0.00   0.00  4095.95    2480  101.58  53013.20  4361.40  2450.00
#>      sumVal sumQnty finQnty sumfinQnty
#> 1  10590.19 1090006   96605    1745785
#> 2  10590.19 1090006   90968    1745785
#> 3  10590.19 1090006   72462    1745785
#> 4  10590.19 1090006  263562    1745785
#> 5  10590.19 1090006   30902    1745785
#> 6  10590.19 1090006  450551    1745785
#> 7  10590.19 1090006    3141    1745785
#> 8  10590.19 1090006   21292    1745785
#> 9  10590.19 1090006   56110    1745785
#> 10 10590.19 1090006    6893    1745785
#> 11 10590.19 1090006    1166    1745785
#> 12 10590.19 1090006    3828    1745785
#> 13 10590.19 1090006    2478    1745785
#> 14 10590.19 1090006    3990    1745785
#> 15 10590.19 1090006    1232    1745785
#> 16 10590.19 1090006   31039    1745785
#> 17 10590.19 1090006   18484    1745785
#> 18 10590.19 1090006    6011    1745785
#> 19 10590.19 1090006  136445    1745785
#> 20 10590.19 1090006     106    1745785
#> 21 10590.19 1090006     687    1745785
#> 22 10590.19 1090006   17419    1745785
#> 23 10590.19 1090006    9489    1745785
#> 24 10590.19 1090006   23795    1745785
#> 25 10590.19 1090006    1816    1745785
#> 26 10590.19 1090006  132679    1745785
#> 27 10590.19 1090006    3895    1745785
#> 28 10590.19 1090006    2177    1745785
#> 29 10590.19 1090006   38689    1745785
#> 30 10590.19 1090006     226    1745785
#> 31 10590.19 1090006   46137    1745785
#> 32 10590.19 1090006    2530    1745785
#> 33 10590.19 1090006    9260    1745785
#> 34 10590.19 1090006   40535    1745785
#> 35 10590.19 1090006    2345    1745785
#> 36 10590.19 1090006   14921    1745785
#> 37 10590.19 1090006     433    1745785
#> 38 10590.19 1090006   16921    1745785
#> 39 10590.19 1090006    1778    1745785
#> 40 10590.19 1090006   16059    1745785
#> 41 10590.19 1090006    5421    1745785
#> 42 10590.19 1090006     571    1745785
#> 43 10590.19 1090006   10617    1745785
#> 44 10590.19 1090006     141    1745785
#> 45 10590.19 1090006   12886    1745785
#> 46 10590.19 1090006   18792    1745785
#> 47 10590.19 1090006     881    1745785
#> 48 10590.19 1090006    1255    1745785
#> 49 10590.19 1090006   13685    1745785
#> 50 10590.19 1090006    2480    1745785

# nifty bank
nse_preopen_nifty_bank()

#>        symbol series corp_action_date corp_action   price change percent_change
#> 1    AXISBANK     EQ             <NA>        <NA>  708.60  -8.20          -1.14
#> 2  INDUSINDBK     EQ             <NA>        <NA>  963.00  -4.00          -0.41
#> 3      AUBANK     EQ             <NA>        <NA>  980.00  -3.45          -0.35
#> 4         PNB     EQ             <NA>        <NA>   36.50   0.30           0.83
#> 5        SBIN     EQ             <NA>        <NA>  387.00   2.60           0.68
#> 6     RBLBANK     EQ             <NA>        <NA>  201.00   1.05           0.53
#> 7   ICICIBANK     EQ             <NA>        <NA>  628.05   2.60           0.42
#> 8  FEDERALBNK     EQ             <NA>        <NA>   84.70   0.30           0.36
#> 9  BANDHANBNK     EQ             <NA>        <NA>  290.10   0.55           0.19
#> 10 IDFCFIRSTB     EQ             <NA>        <NA>   55.65   0.10           0.18
#> 11  KOTAKBANK     EQ             <NA>        <NA> 1732.00   1.45           0.08
#> 12   HDFCBANK     EQ             <NA>        <NA> 1458.35   0.15           0.01
#>    prev_close quantity   value   mkt_cap year_high year_low sum_val
#> 1      716.80   263562 1867.60 177320.19    799.00   332.85 3251.22
#> 2      967.00     6893   66.38  55057.54   1119.50   330.00 3251.22
#> 3      983.45      354    3.47  20375.36   1354.40   366.00 3251.22
#> 4       36.20   400652  146.24   9037.00     46.35    26.30 3251.22
#> 5      384.40   136445  528.04 147574.25    427.70   149.45 3251.22
#> 6      199.95    20728   41.66   9160.75    274.30   105.50 3251.22
#> 7      625.45    38689  242.99 427790.01    679.40   285.55 3251.22
#> 8       84.40    23537   19.94  16665.09     92.50    36.80 3251.22
#> 9      289.55     8492   24.64  19894.25    430.70   195.40 3251.22
#> 10      55.55    30199   16.81  14700.51     69.30    18.45 3251.22
#> 11    1730.55     5421   93.89 250490.47   2049.00  1115.00 3251.22
#> 12    1458.20    13685  199.58 623997.79   1641.00   826.10 3251.22
#>    sum_quantity fin_quantity sum_fin_quantity
#> 1        664332       263562           948657
#> 2        664332         6893           948657
#> 3        664332          354           948657
#> 4        664332       400652           948657
#> 5        664332       136445           948657
#> 6        664332        20728           948657
#> 7        664332        38689           948657
#> 8        664332        23537           948657
#> 9        664332         8492           948657
#> 10       664332        30199           948657
#> 11       664332         5421           948657
#> 12       664332        13685           948657

# retain original column names as returned by NSE
nse_preopen_nifty_bank(clean_names = FALSE)

#>        symbol series  xDt caAct     iep   chn perChn    pCls trdQnty    iVal
#> 1    AXISBANK     EQ <NA>  <NA>  708.60 -8.20  -1.14  716.80  263562 1867.60
#> 2  INDUSINDBK     EQ <NA>  <NA>  963.00 -4.00  -0.41  967.00    6893   66.38
#> 3      AUBANK     EQ <NA>  <NA>  980.00 -3.45  -0.35  983.45     354    3.47
#> 4         PNB     EQ <NA>  <NA>   36.50  0.30   0.83   36.20  400652  146.24
#> 5        SBIN     EQ <NA>  <NA>  387.00  2.60   0.68  384.40  136445  528.04
#> 6     RBLBANK     EQ <NA>  <NA>  201.00  1.05   0.53  199.95   20728   41.66
#> 7   ICICIBANK     EQ <NA>  <NA>  628.05  2.60   0.42  625.45   38689  242.99
#> 8  FEDERALBNK     EQ <NA>  <NA>   84.70  0.30   0.36   84.40   23537   19.94
#> 9  BANDHANBNK     EQ <NA>  <NA>  290.10  0.55   0.19  289.55    8492   24.64
#> 10 IDFCFIRSTB     EQ <NA>  <NA>   55.65  0.10   0.18   55.55   30199   16.81
#> 11  KOTAKBANK     EQ <NA>  <NA> 1732.00  1.45   0.08 1730.55    5421   93.89
#> 12   HDFCBANK     EQ <NA>  <NA> 1458.35  0.15   0.01 1458.20   13685  199.58
#>       mktCap   yHigh    yLow  sumVal sumQnty finQnty sumfinQnty
#> 1  177320.19  799.00  332.85 3251.22  664332  263562     948657
#> 2   55057.54 1119.50  330.00 3251.22  664332    6893     948657
#> 3   20375.36 1354.40  366.00 3251.22  664332     354     948657
#> 4    9037.00   46.35   26.30 3251.22  664332  400652     948657
#> 5  147574.25  427.70  149.45 3251.22  664332  136445     948657
#> 6    9160.75  274.30  105.50 3251.22  664332   20728     948657
#> 7  427790.01  679.40  285.55 3251.22  664332   38689     948657
#> 8   16665.09   92.50   36.80 3251.22  664332   23537     948657
#> 9   19894.25  430.70  195.40 3251.22  664332    8492     948657
#> 10  14700.51   69.30   18.45 3251.22  664332   30199     948657
#> 11 250490.47 2049.00 1115.00 3251.22  664332    5421     948657
#> 12 623997.79 1641.00  826.10 3251.22  664332   13685     948657

# }