Top F&O gainers and losers

Top futures and options gainers and losers for the last trading session.

nse_fo_top_gainers(clean_names = TRUE)

nse_fo_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_fo_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1      CONCOR     EQ                  2020-11-18
#> 2        BHEL     EQ                  2020-09-18
#> 3      AUBANK     EQ                  2019-07-18
#> 4    GMRINFRA     EQ                  2020-09-10
#> 5  APOLLOTYRE     EQ                  2020-08-12
#> 6         PVR     EQ                  2020-07-09
#> 7  ULTRACEMCO     EQ                  2020-07-29
#> 8   AMBUJACEM     EQ                  2021-03-19
#> 9   TATAPOWER     EQ                  2020-07-14
#> 10      IRCTC     EQ                  2020-10-19
#>                                 last_corp_announcement open_price high_price
#> 1                    Interim Dividend - Rs 3 Per Share     577.00     643.95
#> 2                               Annual General Meeting      50.35      56.10
#> 3  Annual General Meeting/Dividend - Rs 0.75 Per Share    1187.65    1290.00
#> 4                               Annual General Meeting      27.30      29.95
#> 5                               Annual General Meeting     243.00     261.25
#> 6                       Rights 7:94 @ Premium Rs 774/-    1410.00    1501.95
#> 7    Annual General Meeting/Dividend - Rs 13 Per Share    6429.00    6830.00
#> 8                            Dividend - Re 1 Per Share     281.50     294.30
#> 9  Annual General Meeting/Dividend - Rs 1.55 Per Share     103.80     114.00
#> 10                              Annual General Meeting    1885.00    2015.80
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1     575.05            615.95           578.35           6.50        15871139
#> 2      49.90             54.30            51.00           6.47       213396105
#> 3    1176.80           1274.00          1200.90           6.09         2247523
#> 4      27.05             29.00            27.75           4.50        56000402
#> 5     242.70            260.00           250.00           4.00        12623097
#> 6    1393.90           1474.00          1417.65           3.97         2267515
#> 7    6394.45           6755.00          6500.50           3.92         1303144
#> 8     278.75            292.90           282.10           3.83        10478749
#> 9     102.75            109.25           105.25           3.80       147425687
#> 10   1870.00           1957.15          1886.10           3.77         9592188
#>    turnover_in_lakhs
#> 1           98707.37
#> 2          115383.27
#> 3           28271.59
#> 4           16139.32
#> 5           31928.86
#> 6           33295.96
#> 7           86771.93
#> 8           30241.67
#> 9          161416.38
#> 10         189152.19

# retain original column names as returned by NSE
nse_fo_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1      CONCOR     EQ               2020-11-18
#> 2        BHEL     EQ               2020-09-18
#> 3      AUBANK     EQ               2019-07-18
#> 4    GMRINFRA     EQ               2020-09-10
#> 5  APOLLOTYRE     EQ               2020-08-12
#> 6         PVR     EQ               2020-07-09
#> 7  ULTRACEMCO     EQ               2020-07-29
#> 8   AMBUJACEM     EQ               2021-03-19
#> 9   TATAPOWER     EQ               2020-07-14
#> 10      IRCTC     EQ               2020-10-19
#>                                   lastCorpAnnouncement openPrice highPrice
#> 1                    Interim Dividend - Rs 3 Per Share    577.00    643.95
#> 2                               Annual General Meeting     50.35     56.10
#> 3  Annual General Meeting/Dividend - Rs 0.75 Per Share   1187.65   1290.00
#> 4                               Annual General Meeting     27.30     29.95
#> 5                               Annual General Meeting    243.00    261.25
#> 6                       Rights 7:94 @ Premium Rs 774/-   1410.00   1501.95
#> 7    Annual General Meeting/Dividend - Rs 13 Per Share   6429.00   6830.00
#> 8                            Dividend - Re 1 Per Share    281.50    294.30
#> 9  Annual General Meeting/Dividend - Rs 1.55 Per Share    103.80    114.00
#> 10                              Annual General Meeting   1885.00   2015.80
#>    lowPrice     ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1    575.05  615.95        578.35     6.50       15871139        98707.37
#> 2     49.90   54.30         51.00     6.47      213396105       115383.27
#> 3   1176.80 1274.00       1200.90     6.09        2247523        28271.59
#> 4     27.05   29.00         27.75     4.50       56000402        16139.32
#> 5    242.70  260.00        250.00     4.00       12623097        31928.86
#> 6   1393.90 1474.00       1417.65     3.97        2267515        33295.96
#> 7   6394.45 6755.00       6500.50     3.92        1303144        86771.93
#> 8    278.75  292.90        282.10     3.83       10478749        30241.67
#> 9    102.75  109.25        105.25     3.80      147425687       161416.38
#> 10  1870.00 1957.15       1886.10     3.77        9592188       189152.19

# top losers
nse_fo_top_losers()

#>        symbol series last_corp_announcement_date
#> 1  JINDALSTEL     EQ                  2020-09-23
#> 2    JSWSTEEL     EQ                  2020-07-06
#> 3        SAIL     EQ                  2021-02-04
#> 4        HDFC     EQ                  2020-07-09
#> 5    HINDALCO     EQ                  2020-09-03
#> 6   TATASTEEL     EQ                  2020-08-06
#> 7        VEDL     EQ                  2020-10-28
#> 8       L&TFH     EQ                  2021-01-21
#> 9  TATAMOTORS     EQ                  2016-07-18
#> 10 BAJAJFINSV     EQ                  2020-03-03
#>                               last_corp_announcement open_price high_price
#> 1                             Annual General Meeting     338.65     344.00
#> 2                          Dividend - Rs 2 Per Share     415.10     421.40
#> 3              Interim Dividend - Rs  1.00 Per Share      76.90      77.75
#> 4  Annual General Meeting/Dividend - Rs 21 Per Share    2610.00    2610.00
#> 5   Annual General Meeting/Dividend - Re 1 Per Share     352.00     357.70
#> 6                         Dividend - Rs 10 Per Share     751.00     766.90
#> 7               Interim Dividend - Rs 9.50 Per Share     215.80     217.05
#> 8              Rights 17:74@ Premium Rs 55 Per Share     109.80     111.35
#> 9                     Dividend - Re 0.20/- Per Share     339.00     345.40
#> 10                 Interim Dividend - Rs 5 Per Share   10150.00   10280.00
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1     335.20            337.70           349.55          -3.39         8664486
#> 2     411.05            415.50           428.15          -2.95         8655068
#> 3      75.00             76.20            78.45          -2.87        46203742
#> 4    2568.00           2584.90          2654.90          -2.64         5315682
#> 5     347.00            349.90           359.35          -2.63        13623173
#> 6     743.65            757.60           777.15          -2.52        22873822
#> 7     212.00            213.80           219.30          -2.51        13750833
#> 8     109.00            109.55           112.30          -2.45        13265491
#> 9     337.30            340.00           348.50          -2.44        59863169
#> 10  10077.75          10148.00         10395.60          -2.38          412117
#>    turnover_in_lakhs
#> 1           29374.34
#> 2           35973.06
#> 3           35276.56
#> 4          137372.11
#> 5           47971.28
#> 6          173209.73
#> 7           29532.66
#> 8           14609.29
#> 9          204432.72
#> 10          41921.98

# retain original column names as returned by NSE
nse_fo_top_losers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1  JINDALSTEL     EQ               2020-09-23
#> 2    JSWSTEEL     EQ               2020-07-06
#> 3        SAIL     EQ               2021-02-04
#> 4        HDFC     EQ               2020-07-09
#> 5    HINDALCO     EQ               2020-09-03
#> 6   TATASTEEL     EQ               2020-08-06
#> 7        VEDL     EQ               2020-10-28
#> 8       L&TFH     EQ               2021-01-21
#> 9  TATAMOTORS     EQ               2016-07-18
#> 10 BAJAJFINSV     EQ               2020-03-03
#>                                 lastCorpAnnouncement openPrice highPrice
#> 1                             Annual General Meeting    338.65    344.00
#> 2                          Dividend - Rs 2 Per Share    415.10    421.40
#> 3              Interim Dividend - Rs  1.00 Per Share     76.90     77.75
#> 4  Annual General Meeting/Dividend - Rs 21 Per Share   2610.00   2610.00
#> 5   Annual General Meeting/Dividend - Re 1 Per Share    352.00    357.70
#> 6                         Dividend - Rs 10 Per Share    751.00    766.90
#> 7               Interim Dividend - Rs 9.50 Per Share    215.80    217.05
#> 8              Rights 17:74@ Premium Rs 55 Per Share    109.80    111.35
#> 9                     Dividend - Re 0.20/- Per Share    339.00    345.40
#> 10                 Interim Dividend - Rs 5 Per Share  10150.00  10280.00
#>    lowPrice      ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1    335.20   337.70        349.55    -3.39        8664486        29374.34
#> 2    411.05   415.50        428.15    -2.95        8655068        35973.06
#> 3     75.00    76.20         78.45    -2.87       46203742        35276.56
#> 4   2568.00  2584.90       2654.90    -2.64        5315682       137372.11
#> 5    347.00   349.90        359.35    -2.63       13623173        47971.28
#> 6    743.65   757.60        777.15    -2.52       22873822       173209.73
#> 7    212.00   213.80        219.30    -2.51       13750833        29532.66
#> 8    109.00   109.55        112.30    -2.45       13265491        14609.29
#> 9    337.30   340.00        348.50    -2.44       59863169       204432.72
#> 10 10077.75 10148.00      10395.60    -2.38         412117        41921.98

# }