Top F&O gainers and losers

Top futures and options gainers and losers for the last trading session.

nse_fo_top_gainers(clean_names = TRUE)

nse_fo_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_fo_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1        BHEL     EQ                  2020-09-18
#> 2      CONCOR     EQ                  2020-11-18
#> 3    GMRINFRA     EQ                  2020-09-10
#> 4      AUBANK     EQ                  2019-07-18
#> 5   TATAPOWER     EQ                  2020-07-14
#> 6  TORNTPOWER     EQ                  2021-02-16
#> 7  ULTRACEMCO     EQ                  2020-07-29
#> 8       IRCTC     EQ                  2020-10-19
#> 9   AMBUJACEM     EQ                  2021-03-19
#> 10        PVR     EQ                  2020-07-09
#>                                 last_corp_announcement open_price high_price
#> 1                               Annual General Meeting      50.35      56.10
#> 2                    Interim Dividend - Rs 3 Per Share     577.00     643.95
#> 3                               Annual General Meeting      27.30      29.95
#> 4  Annual General Meeting/Dividend - Rs 0.75 Per Share    1187.65    1290.00
#> 5  Annual General Meeting/Dividend - Rs 1.55 Per Share     103.80     114.00
#> 6                 Interim Dividend - Rs 5.50 Per Share     397.00     426.50
#> 7    Annual General Meeting/Dividend - Rs 13 Per Share    6429.00    6830.00
#> 8                               Annual General Meeting    1885.00    2015.80
#> 9                            Dividend - Re 1 Per Share     281.50     294.30
#> 10                      Rights 7:94 @ Premium Rs 774/-    1410.00    1501.95
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1      49.90             54.55            51.00           6.96       193520016
#> 2     575.05            616.40           578.35           6.58        14798928
#> 3      27.05             29.30            27.75           5.59        50780006
#> 4    1176.80           1262.10          1200.90           5.10         2018744
#> 5     102.75            110.00           105.25           4.51       133555506
#> 6     395.40            423.00           404.75           4.51         4466225
#> 7    6394.45           6793.40          6500.50           4.51         1117090
#> 8    1870.00           1963.70          1886.10           4.11         9049994
#> 9     278.75            293.45           282.10           4.02         8596755
#> 10   1393.90           1470.70          1417.65           3.74         2088678
#>    turnover_in_lakhs
#> 1          104597.57
#> 2           92108.53
#> 3           14624.64
#> 4           25370.77
#> 5          146256.63
#> 6           18620.59
#> 7           74167.96
#> 8          178520.18
#> 9           24747.48
#> 10          30666.39

# retain original column names as returned by NSE
nse_fo_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1        BHEL     EQ               2020-09-18
#> 2      CONCOR     EQ               2020-11-18
#> 3    GMRINFRA     EQ               2020-09-10
#> 4      AUBANK     EQ               2019-07-18
#> 5   TATAPOWER     EQ               2020-07-14
#> 6  TORNTPOWER     EQ               2021-02-16
#> 7  ULTRACEMCO     EQ               2020-07-29
#> 8       IRCTC     EQ               2020-10-19
#> 9   AMBUJACEM     EQ               2021-03-19
#> 10        PVR     EQ               2020-07-09
#>                                   lastCorpAnnouncement openPrice highPrice
#> 1                               Annual General Meeting     50.35     56.10
#> 2                    Interim Dividend - Rs 3 Per Share    577.00    643.95
#> 3                               Annual General Meeting     27.30     29.95
#> 4  Annual General Meeting/Dividend - Rs 0.75 Per Share   1187.65   1290.00
#> 5  Annual General Meeting/Dividend - Rs 1.55 Per Share    103.80    114.00
#> 6                 Interim Dividend - Rs 5.50 Per Share    397.00    426.50
#> 7    Annual General Meeting/Dividend - Rs 13 Per Share   6429.00   6830.00
#> 8                               Annual General Meeting   1885.00   2015.80
#> 9                            Dividend - Re 1 Per Share    281.50    294.30
#> 10                      Rights 7:94 @ Premium Rs 774/-   1410.00   1501.95
#>    lowPrice     ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1     49.90   54.55         51.00     6.96      193520016       104597.57
#> 2    575.05  616.40        578.35     6.58       14798928        92108.53
#> 3     27.05   29.30         27.75     5.59       50780006        14624.64
#> 4   1176.80 1262.10       1200.90     5.10        2018744        25370.77
#> 5    102.75  110.00        105.25     4.51      133555506       146256.63
#> 6    395.40  423.00        404.75     4.51        4466225        18620.59
#> 7   6394.45 6793.40       6500.50     4.51        1117090        74167.96
#> 8   1870.00 1963.70       1886.10     4.11        9049994       178520.18
#> 9    278.75  293.45        282.10     4.02        8596755        24747.48
#> 10  1393.90 1470.70       1417.65     3.74        2088678        30666.39

# top losers
nse_fo_top_losers()

#>        symbol series last_corp_announcement_date
#> 1  JINDALSTEL     EQ                  2020-09-23
#> 2    JSWSTEEL     EQ                  2020-07-06
#> 3        HDFC     EQ                  2020-07-09
#> 4    HDFCBANK     EQ                  2019-09-19
#> 5  TATAMOTORS     EQ                  2016-07-18
#> 6        VEDL     EQ                  2020-10-28
#> 7        SAIL     EQ                  2021-02-04
#> 8   LICHSGFIN     EQ                  2020-09-17
#> 9  BAJAJFINSV     EQ                  2020-03-03
#> 10       SBIN     EQ                  2018-06-15
#>                                                     last_corp_announcement
#> 1                                                   Annual General Meeting
#> 2                                                Dividend - Rs 2 Per Share
#> 3                        Annual General Meeting/Dividend - Rs 21 Per Share
#> 4  Face Value Split (Sub-Division) - From Rs 2 Per Share To Rs 1 Per Share
#> 5                                           Dividend - Re 0.20/- Per Share
#> 6                                     Interim Dividend - Rs 9.50 Per Share
#> 7                                    Interim Dividend - Rs  1.00 Per Share
#> 8                         Annual General Meeting/Dividend - Rs 8 Per Share
#> 9                                        Interim Dividend - Rs 5 Per Share
#> 10          Annual General Meeting/ Change In Registrar And Transfer Agent
#>    open_price high_price low_price last_traded_price prev_close_price
#> 1      338.65     344.00    335.20            337.25           349.55
#> 2      415.10     421.40    411.05            415.15           428.15
#> 3     2610.00    2610.00   2568.00           2591.85          2654.90
#> 4     1548.55    1571.00   1539.10           1551.35          1586.85
#> 5      339.00     345.40    337.30            340.90           348.50
#> 6      215.80     217.05    213.50            214.55           219.30
#> 7       76.90      77.75     75.00             76.80            78.45
#> 8      459.95     463.00    452.70            454.85           464.30
#> 9    10150.00   10280.00  10077.75          10186.65         10395.60
#> 10     398.70     403.50    396.60            397.05           405.05
#>    percent_change traded_quantity turnover_in_lakhs
#> 1           -3.52         7734345          26244.18
#> 2           -3.04         7686998          31964.84
#> 3           -2.37         4151100         107290.58
#> 4           -2.24         6983094         108404.85
#> 5           -2.18        51676924         176662.73
#> 6           -2.17        11034932          23740.55
#> 7           -2.10        41229823          31474.85
#> 8           -2.04         3449188          15775.21
#> 9           -2.01          356285          36265.82
#> 10          -1.98        35534387         142258.36

# retain original column names as returned by NSE
nse_fo_top_losers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1  JINDALSTEL     EQ               2020-09-23
#> 2    JSWSTEEL     EQ               2020-07-06
#> 3        HDFC     EQ               2020-07-09
#> 4    HDFCBANK     EQ               2019-09-19
#> 5  TATAMOTORS     EQ               2016-07-18
#> 6        VEDL     EQ               2020-10-28
#> 7        SAIL     EQ               2021-02-04
#> 8   LICHSGFIN     EQ               2020-09-17
#> 9  BAJAJFINSV     EQ               2020-03-03
#> 10       SBIN     EQ               2018-06-15
#>                                                       lastCorpAnnouncement
#> 1                                                   Annual General Meeting
#> 2                                                Dividend - Rs 2 Per Share
#> 3                        Annual General Meeting/Dividend - Rs 21 Per Share
#> 4  Face Value Split (Sub-Division) - From Rs 2 Per Share To Rs 1 Per Share
#> 5                                           Dividend - Re 0.20/- Per Share
#> 6                                     Interim Dividend - Rs 9.50 Per Share
#> 7                                    Interim Dividend - Rs  1.00 Per Share
#> 8                         Annual General Meeting/Dividend - Rs 8 Per Share
#> 9                                        Interim Dividend - Rs 5 Per Share
#> 10          Annual General Meeting/ Change In Registrar And Transfer Agent
#>    openPrice highPrice lowPrice      ltp previousPrice netPrice tradedQuantity
#> 1     338.65    344.00   335.20   337.25        349.55    -3.52        7734345
#> 2     415.10    421.40   411.05   415.15        428.15    -3.04        7686998
#> 3    2610.00   2610.00  2568.00  2591.85       2654.90    -2.37        4151100
#> 4    1548.55   1571.00  1539.10  1551.35       1586.85    -2.24        6983094
#> 5     339.00    345.40   337.30   340.90        348.50    -2.18       51676924
#> 6     215.80    217.05   213.50   214.55        219.30    -2.17       11034932
#> 7      76.90     77.75    75.00    76.80         78.45    -2.10       41229823
#> 8     459.95    463.00   452.70   454.85        464.30    -2.04        3449188
#> 9   10150.00  10280.00 10077.75 10186.65      10395.60    -2.01         356285
#> 10    398.70    403.50   396.60   397.05        405.05    -1.98       35534387
#>    turnoverInLakhs
#> 1         26244.18
#> 2         31964.84
#> 3        107290.58
#> 4        108404.85
#> 5        176662.73
#> 6         23740.55
#> 7         31474.85
#> 8         15775.21
#> 9         36265.82
#> 10       142258.36

# }