Top F&O gainers and losers

Top futures and options gainers and losers for the last trading session.

nse_fo_top_gainers(clean_names = TRUE)

nse_fo_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_fo_top_gainers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 JINDA~ EQ     2019-09-19       Annual General ~      138.       146. 
#>  2 NIITT~ EQ     2019-11-04       Interim Dividen~     1503       1562  
#>  3 VEDL   EQ     2019-07-04       Annual General ~      148.       154. 
#>  4 TATAS~ EQ     2019-07-04       Annual General ~      424.       441. 
#>  5 CASTR~ EQ     2019-08-08       Interim Dividen~      131.       136. 
#>  6 HINDA~ EQ     2019-08-14       Annual General ~      208.       215. 
#>  7 SAIL   EQ     2019-08-22       Annual General ~       39.6       41.2
#>  8 JSWST~ EQ     2019-07-08       Annual General ~      254.       262. 
#>  9 AMARA~ EQ     2019-11-21       Interim Dividen~      727.       745. 
#> 10 JUSTD~ EQ     2018-10-11       Buyback               572        588. 
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_fo_top_gainers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 JINDA~ EQ     2019-09-19       Annual General ~     138.      146.     137  
#>  2 NIITT~ EQ     2019-11-04       Interim Dividen~    1503      1562     1492. 
#>  3 VEDL   EQ     2019-07-04       Annual General ~     148.      154.     148  
#>  4 TATAS~ EQ     2019-07-04       Annual General ~     424.      441.     423. 
#>  5 CASTR~ EQ     2019-08-08       Interim Dividen~     131.      136.     131. 
#>  6 HINDA~ EQ     2019-08-14       Annual General ~     208.      215.     207. 
#>  7 SAIL   EQ     2019-08-22       Annual General ~      39.6      41.2     39.4
#>  8 JSWST~ EQ     2019-07-08       Annual General ~     254.      262.     254. 
#>  9 AMARA~ EQ     2019-11-21       Interim Dividen~     727.      745.     724. 
#> 10 JUSTD~ EQ     2018-10-11       Buyback              572       588.     570  
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# top losers
nse_fo_top_losers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 IDEA   EQ     2019-08-16       Annual General ~        6.6       6.65
#>  2 GAIL   EQ     2019-08-08       Dividend - Rs 0~      121.      121.  
#>  3 SRTRA~ EQ     2019-11-07       Interim Dividen~     1084.     1090   
#>  4 BIOCON EQ     2019-07-18       Dividend - Rs 0~      296.      296.  
#>  5 SUNPH~ EQ     2019-08-20       Annual General ~      434       438.  
#>  6 TORNT~ EQ     2019-06-13       Dividend Rs 4 P~     1865.     1870   
#>  7 AUROP~ EQ     2019-11-21       Interim Dividen~      455       457   
#>  8 MANAP~ EQ     2019-11-19       Interim Dividen~      173.      173   
#>  9 BATAI~ EQ     2019-07-22       Annual General ~     1696      1698.  
#> 10 SRF    EQ     2019-08-13       Interim Dividen~     3270.     3285   
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_fo_top_losers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 IDEA   EQ     2019-08-16       Annual General ~       6.6      6.65     6.35
#>  2 GAIL   EQ     2019-08-08       Dividend - Rs 0~     121.     121.     118.  
#>  3 SRTRA~ EQ     2019-11-07       Interim Dividen~    1084.    1090     1068   
#>  4 BIOCON EQ     2019-07-18       Dividend - Rs 0~     296.     296.     289.  
#>  5 SUNPH~ EQ     2019-08-20       Annual General ~     434      438.     428.  
#>  6 TORNT~ EQ     2019-06-13       Dividend Rs 4 P~    1865.    1870     1841   
#>  7 AUROP~ EQ     2019-11-21       Interim Dividen~     455      457      449.  
#>  8 MANAP~ EQ     2019-11-19       Interim Dividen~     173.     173      170.  
#>  9 BATAI~ EQ     2019-07-22       Annual General ~    1696     1698.    1678   
#> 10 SRF    EQ     2019-08-13       Interim Dividen~    3270.    3285     3237.  
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# }