Top F&O gainers and losers

Top futures and options gainers and losers for the last trading session.

nse_fo_top_gainers(clean_names = TRUE)

nse_fo_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_fo_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1    GRANULES     EQ                  2021-02-08
#> 2  IBULHSGFIN     EQ                  2020-09-03
#> 3     HDFCAMC     EQ                  2020-07-09
#> 4  CUMMINSIND     EQ                  2021-02-09
#> 5         PNB     EQ                  2020-07-27
#> 6  FEDERALBNK     EQ                  2020-07-09
#> 7       L&TFH     EQ                  2021-01-21
#> 8      AUBANK     EQ                  2019-07-18
#> 9    TATACHEM     EQ                  2020-06-22
#> 10     INDIGO     EQ                  2020-08-27
#>                                 last_corp_announcement open_price high_price
#> 1                Interim Dividend - Rs 0.25  Per Share     338.25     369.00
#> 2                               Annual General Meeting     231.10     251.00
#> 3       Annual General Meeting/Dividend - Rs 28 Per Sh    2950.00    3220.00
#> 4                    Interim Dividend - Rs 7 Per Share     822.50     882.95
#> 5                               Annual General Meeting      41.40      43.90
#> 6                               Annual General Meeting      83.70      88.50
#> 7                Rights 17:74@ Premium Rs 55 Per Share     107.50     113.40
#> 8  Annual General Meeting/Dividend - Rs 0.75 Per Share    1162.40    1225.00
#> 9    Annual General Meeting/Dividend - Rs 11 Per Share     738.00     782.00
#> 10                              Annual General Meeting    1704.00    1823.85
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1     338.25            367.50           336.25           9.29        10446684
#> 2     229.50            249.50           229.00           8.95        45419012
#> 3    2934.00           3189.50          2934.10           8.70         1461732
#> 4     822.50            869.80           818.45           6.27         5093443
#> 5      41.25             43.55            41.00           6.22       238140740
#> 6      83.50             88.45            83.30           6.18        38195349
#> 7     107.05            113.10           107.00           5.70        26327611
#> 8    1136.55           1215.00          1151.50           5.51         2078700
#> 9     735.45            774.00           733.65           5.50        12732034
#> 10   1700.25           1783.00          1691.40           5.42         2461629
#>    turnover_in_lakhs
#> 1           37719.84
#> 2          109291.77
#> 3           45834.95
#> 4           44003.78
#> 5          102090.94
#> 6           32974.04
#> 7           29039.35
#> 8           24143.27
#> 9           97734.91
#> 10          43531.45

# retain original column names as returned by NSE
nse_fo_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1    GRANULES     EQ               2021-02-08
#> 2  IBULHSGFIN     EQ               2020-09-03
#> 3     HDFCAMC     EQ               2020-07-09
#> 4  CUMMINSIND     EQ               2021-02-09
#> 5         PNB     EQ               2020-07-27
#> 6  FEDERALBNK     EQ               2020-07-09
#> 7       L&TFH     EQ               2021-01-21
#> 8      AUBANK     EQ               2019-07-18
#> 9    TATACHEM     EQ               2020-06-22
#> 10     INDIGO     EQ               2020-08-27
#>                                   lastCorpAnnouncement openPrice highPrice
#> 1                Interim Dividend - Rs 0.25  Per Share    338.25    369.00
#> 2                               Annual General Meeting    231.10    251.00
#> 3       Annual General Meeting/Dividend - Rs 28 Per Sh   2950.00   3220.00
#> 4                    Interim Dividend - Rs 7 Per Share    822.50    882.95
#> 5                               Annual General Meeting     41.40     43.90
#> 6                               Annual General Meeting     83.70     88.50
#> 7                Rights 17:74@ Premium Rs 55 Per Share    107.50    113.40
#> 8  Annual General Meeting/Dividend - Rs 0.75 Per Share   1162.40   1225.00
#> 9    Annual General Meeting/Dividend - Rs 11 Per Share    738.00    782.00
#> 10                              Annual General Meeting   1704.00   1823.85
#>    lowPrice     ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1    338.25  367.50        336.25     9.29       10446684        37719.84
#> 2    229.50  249.50        229.00     8.95       45419012       109291.77
#> 3   2934.00 3189.50       2934.10     8.70        1461732        45834.95
#> 4    822.50  869.80        818.45     6.27        5093443        44003.78
#> 5     41.25   43.55         41.00     6.22      238140740       102090.94
#> 6     83.50   88.45         83.30     6.18       38195349        32974.04
#> 7    107.05  113.10        107.00     5.70       26327611        29039.35
#> 8   1136.55 1215.00       1151.50     5.51        2078700        24143.27
#> 9    735.45  774.00        733.65     5.50       12732034        97734.91
#> 10  1700.25 1783.00       1691.40     5.42        2461629        43531.45

# top losers
nse_fo_top_losers()

#>        symbol series last_corp_announcement_date
#> 1       ALKEM     EQ                  2021-02-11
#> 2        IDEA     EQ                  2020-09-21
#> 3        BHEL     EQ                  2020-09-18
#> 4  HEROMOTOCO     EQ                  2021-02-15
#> 5       SUNTV     EQ                  2021-02-17
#> 6    BOSCHLTD     EQ                  2020-08-03
#> 7        LTTS     EQ                  2020-10-26
#> 8      MARUTI     EQ                  2020-08-13
#> 9  BAJAJ-AUTO     EQ                  2020-03-03
#> 10    PAGEIND     EQ                  2021-02-17
#>                                                  last_corp_announcement
#> 1                                    Interim Dividend - Rs 25 Per Share
#> 2                                                Annual General Meeting
#> 3                                                Annual General Meeting
#> 4  Interim Dividend - Rs 65 Per Share Special Dividend - Rs 5 Per Share
#> 5                                     Interim Dividend - Rs 5 Per Share
#> 6                    Annual General Meeting/Dividend - Rs 105 Per Share
#> 7                                  Interim Dividend - Rs 7.50 Per Share
#> 8                                            Dividend - Rs 60 Per Share
#> 9                                   Interim Dividend - Rs 120 Per Share
#> 10                                  Interim Dividend - Rs 150 Per Share
#>    open_price high_price low_price last_traded_price prev_close_price
#> 1     2932.00    2942.00   2835.00           2850.00          2914.65
#> 2       11.45      11.65     11.10             11.10            11.35
#> 3       52.20      52.65     50.60             50.95            51.75
#> 4     3510.00    3510.90   3394.00           3422.90          3476.40
#> 5      515.10     520.15    505.25            506.50           513.30
#> 6    15639.00   15639.00  15190.40          15251.00         15449.90
#> 7     2730.00    2807.60   2682.60           2695.00          2729.80
#> 8     7245.60    7262.30   7101.20           7129.25          7214.10
#> 9     3985.00    3985.00   3863.60           3897.95          3944.05
#> 10   29381.20   29596.00  28813.10          29161.00         29463.25
#>    percent_change traded_quantity turnover_in_lakhs
#> 1           -2.22          224139           6405.09
#> 2           -2.20       216754057          24449.86
#> 3           -1.55        65329850          33481.55
#> 4           -1.54          965062          33036.58
#> 5           -1.32         2301946          11739.46
#> 6           -1.29           39743           6086.30
#> 7           -1.27          551997          15227.94
#> 8           -1.18          796258          56794.54
#> 9           -1.17          615104          23956.82
#> 10          -1.03           36650          10685.66

# retain original column names as returned by NSE
nse_fo_top_losers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1       ALKEM     EQ               2021-02-11
#> 2        IDEA     EQ               2020-09-21
#> 3        BHEL     EQ               2020-09-18
#> 4  HEROMOTOCO     EQ               2021-02-15
#> 5       SUNTV     EQ               2021-02-17
#> 6    BOSCHLTD     EQ               2020-08-03
#> 7        LTTS     EQ               2020-10-26
#> 8      MARUTI     EQ               2020-08-13
#> 9  BAJAJ-AUTO     EQ               2020-03-03
#> 10    PAGEIND     EQ               2021-02-17
#>                                                    lastCorpAnnouncement
#> 1                                    Interim Dividend - Rs 25 Per Share
#> 2                                                Annual General Meeting
#> 3                                                Annual General Meeting
#> 4  Interim Dividend - Rs 65 Per Share Special Dividend - Rs 5 Per Share
#> 5                                     Interim Dividend - Rs 5 Per Share
#> 6                    Annual General Meeting/Dividend - Rs 105 Per Share
#> 7                                  Interim Dividend - Rs 7.50 Per Share
#> 8                                            Dividend - Rs 60 Per Share
#> 9                                   Interim Dividend - Rs 120 Per Share
#> 10                                  Interim Dividend - Rs 150 Per Share
#>    openPrice highPrice lowPrice      ltp previousPrice netPrice tradedQuantity
#> 1    2932.00   2942.00  2835.00  2850.00       2914.65    -2.22         224139
#> 2      11.45     11.65    11.10    11.10         11.35    -2.20      216754057
#> 3      52.20     52.65    50.60    50.95         51.75    -1.55       65329850
#> 4    3510.00   3510.90  3394.00  3422.90       3476.40    -1.54         965062
#> 5     515.10    520.15   505.25   506.50        513.30    -1.32        2301946
#> 6   15639.00  15639.00 15190.40 15251.00      15449.90    -1.29          39743
#> 7    2730.00   2807.60  2682.60  2695.00       2729.80    -1.27         551997
#> 8    7245.60   7262.30  7101.20  7129.25       7214.10    -1.18         796258
#> 9    3985.00   3985.00  3863.60  3897.95       3944.05    -1.17         615104
#> 10  29381.20  29596.00 28813.10 29161.00      29463.25    -1.03          36650
#>    turnoverInLakhs
#> 1          6405.09
#> 2         24449.86
#> 3         33481.55
#> 4         33036.58
#> 5         11739.46
#> 6          6086.30
#> 7         15227.94
#> 8         56794.54
#> 9         23956.82
#> 10        10685.66

# }