Top F&O gainers and losers

Top futures and options gainers and losers for the last trading session.

nse_fo_top_gainers(clean_names = TRUE)

nse_fo_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_fo_top_gainers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 IDEA   EQ     2019-08-16       Annual General ~       5.95       6.55
#>  2 SHREE~ EQ     2019-07-31       Dividend - Rs 3~   21849      22692.  
#>  3 NIITT~ EQ     2019-11-04       Interim Dividen~    1653.      1692.  
#>  4 AMBUJ~ EQ     2019-02-27       Dividend - Rs 1~     196        205.  
#>  5 TCS    EQ     2019-10-17       Interim Dividen~    2205       2240   
#>  6 YESBA~ EQ     2019-06-03       Annual General ~      46.2       47.2 
#>  7 ULTRA~ EQ     2019-07-10       Annual General ~    4225.      4310   
#>  8 MCDOW~ EQ     2019-08-13       Annual General ~     558.       570.  
#>  9 BHART~ EQ     2019-04-23       Rights 19:67 @ ~     443.       449.  
#> 10 HCLTE~ EQ     2020-01-24       Interim Dividend     583        588.  
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_fo_top_gainers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 IDEA   EQ     2019-08-16       Annual General ~      5.95      6.55      5.7
#>  2 SHREE~ EQ     2019-07-31       Dividend - Rs 3~  21849     22692.    21710. 
#>  3 NIITT~ EQ     2019-11-04       Interim Dividen~   1653.     1692.     1645  
#>  4 AMBUJ~ EQ     2019-02-27       Dividend - Rs 1~    196       205.      196. 
#>  5 TCS    EQ     2019-10-17       Interim Dividen~   2205      2240      2202. 
#>  6 YESBA~ EQ     2019-06-03       Annual General ~     46.2      47.2      45.3
#>  7 ULTRA~ EQ     2019-07-10       Annual General ~   4225.     4310      4184. 
#>  8 MCDOW~ EQ     2019-08-13       Annual General ~    558.      570.      558. 
#>  9 BHART~ EQ     2019-04-23       Rights 19:67 @ ~    443.      449.      440. 
#> 10 HCLTE~ EQ     2020-01-24       Interim Dividend    583       588.      578  
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# top losers
nse_fo_top_losers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 EICHE~ EQ     2019-07-24       Annual General ~    20700      20700  
#>  2 NMDC   EQ     2019-08-22       Annual General ~      127.       127. 
#>  3 JSWST~ EQ     2019-07-08       Annual General ~      261.       264. 
#>  4 NCC    EQ     2019-08-08       Dividend - Rs 1~       54.6       55.2
#>  5 CANBK  EQ     2019-07-17       Annual General ~      202.       205. 
#>  6 PFC    EQ     2019-08-20       Annual General ~      114.       116. 
#>  7 RECLTD EQ     2019-08-21       Annual General ~      139.       140. 
#>  8 LT     EQ     2019-07-24       Annual General ~     1302       1309. 
#>  9 ASHOK~ EQ     2019-07-23       Annual General ~       80.3       81.5
#> 10 TATAM~ EQ     2016-07-18       Dividend - Re 0~       74.9       75.8
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_fo_top_losers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 EICHE~ EQ     2019-07-24       Annual General ~   20700     20700    20094. 
#>  2 NMDC   EQ     2019-08-22       Annual General ~     127.      127.     124. 
#>  3 JSWST~ EQ     2019-07-08       Annual General ~     261.      264.     257  
#>  4 NCC    EQ     2019-08-08       Dividend - Rs 1~      54.6      55.2     53.6
#>  5 CANBK  EQ     2019-07-17       Annual General ~     202.      205.     200. 
#>  6 PFC    EQ     2019-08-20       Annual General ~     114.      116.     114. 
#>  7 RECLTD EQ     2019-08-21       Annual General ~     139.      140.     137. 
#>  8 LT     EQ     2019-07-24       Annual General ~    1302      1309.    1283. 
#>  9 ASHOK~ EQ     2019-07-23       Annual General ~      80.3      81.5     80  
#> 10 TATAM~ EQ     2016-07-18       Dividend - Re 0~      74.9      75.8     74.0
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# }