Top F&O gainers and losers

Top futures and options gainers and losers for the last trading session.

nse_fo_top_gainers(clean_names = TRUE)

nse_fo_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_fo_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1        BHEL     EQ                  2019-09-11
#> 2  MUTHOOTFIN     EQ                  2020-03-23
#> 3  BAJAJFINSV     EQ                  2020-03-03
#> 4         BEL     EQ                  2020-02-11
#> 5   COALINDIA     EQ                  2020-03-19
#> 6    CHOLAFIN     EQ                  2020-03-04
#> 7  BAJFINANCE     EQ                  2020-03-03
#> 8        ZEEL     EQ                  2019-07-15
#> 9         NCC     EQ                  2019-08-08
#> 10       VEDL     EQ                  2020-03-05
#>                                  last_corp_announcement open_price high_price
#> 1   Annual General Meeting/Dividend - Rs 1.20 Per Share      28.65      32.15
#> 2                    Interim Dividend - Rs 15 Per Share    1036.90    1199.55
#> 3                     Interim Dividend - Rs 5 Per Share    5036.95    5499.00
#> 4                  Interim Dividend - Rs 1.40 Per Share      71.95      77.25
#> 5                    Interim Dividend - Rs 12 Per Share     129.00     139.70
#> 6                    Interim Dividend-Re 0.70 Per Share     168.20     178.80
#> 7                    Interim Dividend - Rs 10 Per Share    2375.00    2541.70
#> 8  Annual General Meeting/ Dividend - Rs 3.50 Per Share     166.50     175.45
#> 9                          Dividend - Rs 1.50 Per Share      27.80      29.70
#> 10                  Interim Dividend - Rs 3.9 Per Share     102.80     107.80
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1      28.50             32.00            27.45          16.58       250949789
#> 2    1036.90           1160.30           999.65          16.07        18024858
#> 3    5021.55           5450.75          5036.95           8.22         1806208
#> 4      71.65             76.95            71.65           7.40        33392838
#> 5     128.15            137.65           129.45           6.33        34587382
#> 6     168.20            177.95           168.20           5.80        26257427
#> 7    2373.00           2527.20          2396.00           5.48        12738830
#> 8     165.50            173.95           165.00           5.42        37917230
#> 9      27.55             29.30            27.85           5.21        26610887
#> 10    102.00            106.90           102.10           4.70        27586795
#>    turnover_in_lakhs
#> 1           77443.10
#> 2          203545.71
#> 3           95403.73
#> 4           25204.91
#> 5           47031.92
#> 6           45821.84
#> 7          315647.83
#> 8           65103.88
#> 9            7743.77
#> 10          29197.86

# retain original column names as returned by NSE
nse_fo_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1        BHEL     EQ               2019-09-11
#> 2  MUTHOOTFIN     EQ               2020-03-23
#> 3  BAJAJFINSV     EQ               2020-03-03
#> 4         BEL     EQ               2020-02-11
#> 5   COALINDIA     EQ               2020-03-19
#> 6    CHOLAFIN     EQ               2020-03-04
#> 7  BAJFINANCE     EQ               2020-03-03
#> 8        ZEEL     EQ               2019-07-15
#> 9         NCC     EQ               2019-08-08
#> 10       VEDL     EQ               2020-03-05
#>                                    lastCorpAnnouncement openPrice highPrice
#> 1   Annual General Meeting/Dividend - Rs 1.20 Per Share     28.65     32.15
#> 2                    Interim Dividend - Rs 15 Per Share   1036.90   1199.55
#> 3                     Interim Dividend - Rs 5 Per Share   5036.95   5499.00
#> 4                  Interim Dividend - Rs 1.40 Per Share     71.95     77.25
#> 5                    Interim Dividend - Rs 12 Per Share    129.00    139.70
#> 6                    Interim Dividend-Re 0.70 Per Share    168.20    178.80
#> 7                    Interim Dividend - Rs 10 Per Share   2375.00   2541.70
#> 8  Annual General Meeting/ Dividend - Rs 3.50 Per Share    166.50    175.45
#> 9                          Dividend - Rs 1.50 Per Share     27.80     29.70
#> 10                  Interim Dividend - Rs 3.9 Per Share    102.80    107.80
#>    lowPrice     ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1     28.50   32.00         27.45    16.58      250949789        77443.10
#> 2   1036.90 1160.30        999.65    16.07       18024858       203545.71
#> 3   5021.55 5450.75       5036.95     8.22        1806208        95403.73
#> 4     71.65   76.95         71.65     7.40       33392838        25204.91
#> 5    128.15  137.65        129.45     6.33       34587382        47031.92
#> 6    168.20  177.95        168.20     5.80       26257427        45821.84
#> 7   2373.00 2527.20       2396.00     5.48       12738830       315647.83
#> 8    165.50  173.95        165.00     5.42       37917230        65103.88
#> 9     27.55   29.30         27.85     5.21       26610887         7743.77
#> 10   102.00  106.90        102.10     4.70       27586795        29197.86

# top losers
nse_fo_top_losers()

#>        symbol series last_corp_announcement_date
#> 1         IGL     EQ                  2019-09-12
#> 2  CUMMINSIND     EQ                  2020-02-07
#> 3        NMDC     EQ                  2020-02-18
#> 4   LICHSGFIN     EQ                  2019-08-16
#> 5  PIDILITIND     EQ                  2020-03-13
#> 6      NAUKRI     EQ                  2020-02-18
#> 7    JUSTDIAL     EQ                  2018-10-11
#> 8     ESCORTS     EQ                  2019-07-11
#> 9         PEL     EQ                  2020-07-16
#> 10    HAVELLS     EQ                  2020-03-16
#>                                  last_corp_announcement open_price high_price
#> 1   Annual General Meeting/Dividend - Rs 2.40 Per Share     474.40     474.50
#> 2                     Interim Dividend - Rs 7 Per Share     383.00     383.00
#> 3                  Interim Dividend - Rs 5.29 Per Share      85.55      85.75
#> 4  Annual General Meeting/ Dividend - Rs 7.60 Per Share     292.00     292.00
#> 5                     Interim Dividend - Rs 7 Per Share    1432.15    1444.95
#> 6                  Interim Dividend - Rs 3.50 Per Share    2788.00    2799.00
#> 7                                               Buyback     372.00     372.00
#> 8                          Dividend Rs - 2.50 Per Share     995.00    1014.90
#> 9     Annual General Meeting/Dividend - Rs 14 Per Share    1035.00    1044.95
#> 10                    Interim Dividend - Rs 4 Per Share     553.00     553.90
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1     448.00            449.35           474.15          -5.23        10581901
#> 2     368.35            369.90           385.90          -4.15         2461538
#> 3      81.60             82.65            85.55          -3.39         9479762
#> 4     282.30            283.50           292.15          -2.96         5343997
#> 5    1383.10           1394.00          1432.15          -2.66         1798258
#> 6    2738.40           2745.10          2798.90          -1.92          263110
#> 7     364.00            367.00           373.95          -1.86         1690500
#> 8     978.00            983.45           999.85          -1.64         2663456
#> 9    1015.05           1037.75          1051.30          -1.29         1461988
#> 10    542.00            546.65           553.50          -1.24         2210632
#>    turnover_in_lakhs
#> 1           48041.83
#> 2            9194.09
#> 3            7859.67
#> 4           15231.46
#> 5           25147.56
#> 6            7270.28
#> 7            6205.99
#> 8           26529.89
#> 9           15058.33
#> 10          12100.78

# retain original column names as returned by NSE
nse_fo_top_losers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1         IGL     EQ               2019-09-12
#> 2  CUMMINSIND     EQ               2020-02-07
#> 3        NMDC     EQ               2020-02-18
#> 4   LICHSGFIN     EQ               2019-08-16
#> 5  PIDILITIND     EQ               2020-03-13
#> 6      NAUKRI     EQ               2020-02-18
#> 7    JUSTDIAL     EQ               2018-10-11
#> 8     ESCORTS     EQ               2019-07-11
#> 9         PEL     EQ               2020-07-16
#> 10    HAVELLS     EQ               2020-03-16
#>                                    lastCorpAnnouncement openPrice highPrice
#> 1   Annual General Meeting/Dividend - Rs 2.40 Per Share    474.40    474.50
#> 2                     Interim Dividend - Rs 7 Per Share    383.00    383.00
#> 3                  Interim Dividend - Rs 5.29 Per Share     85.55     85.75
#> 4  Annual General Meeting/ Dividend - Rs 7.60 Per Share    292.00    292.00
#> 5                     Interim Dividend - Rs 7 Per Share   1432.15   1444.95
#> 6                  Interim Dividend - Rs 3.50 Per Share   2788.00   2799.00
#> 7                                               Buyback    372.00    372.00
#> 8                          Dividend Rs - 2.50 Per Share    995.00   1014.90
#> 9     Annual General Meeting/Dividend - Rs 14 Per Share   1035.00   1044.95
#> 10                    Interim Dividend - Rs 4 Per Share    553.00    553.90
#>    lowPrice     ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1    448.00  449.35        474.15    -5.23       10581901        48041.83
#> 2    368.35  369.90        385.90    -4.15        2461538         9194.09
#> 3     81.60   82.65         85.55    -3.39        9479762         7859.67
#> 4    282.30  283.50        292.15    -2.96        5343997        15231.46
#> 5   1383.10 1394.00       1432.15    -2.66        1798258        25147.56
#> 6   2738.40 2745.10       2798.90    -1.92         263110         7270.28
#> 7    364.00  367.00        373.95    -1.86        1690500         6205.99
#> 8    978.00  983.45        999.85    -1.64        2663456        26529.89
#> 9   1015.05 1037.75       1051.30    -1.29        1461988        15058.33
#> 10   542.00  546.65        553.50    -1.24        2210632        12100.78

# }