Top F&O gainers and losers

Top futures and options gainers and losers for the last trading session.

nse_fo_top_gainers(clean_names = TRUE)

nse_fo_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_fo_top_gainers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 DISHTV EQ     2018-11-05       Interim Dividen~       12.6       14.4
#>  2 M&MFIN EQ     2019-07-15       Annual General ~      328.       339. 
#>  3 SAIL   EQ     2019-08-22       Annual General ~       38.5       40.0
#>  4 CHOLA~ EQ     2019-07-22       Annual General ~      302.       312. 
#>  5 JINDA~ EQ     2019-09-19       Annual General ~      141.       145. 
#>  6 CANBK  EQ     2019-07-17       Annual General ~      208        216. 
#>  7 MGL    EQ     2019-08-29       Annual General ~     1025.      1046  
#>  8 ADANI~ EQ     2019-07-26       Annual General ~      368.       378. 
#>  9 JUBLF~ EQ     2019-09-16       Annual General ~     1579       1617  
#> 10 PFC    EQ     2019-08-20       Annual General ~      109.       113. 
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_fo_top_gainers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 DISHTV EQ     2018-11-05       Interim Dividen~      12.6      14.4     12.5
#>  2 M&MFIN EQ     2019-07-15       Annual General ~     328.      339.     326  
#>  3 SAIL   EQ     2019-08-22       Annual General ~      38.5      40.0     38.3
#>  4 CHOLA~ EQ     2019-07-22       Annual General ~     302.      312.     297. 
#>  5 JINDA~ EQ     2019-09-19       Annual General ~     141.      145.     139. 
#>  6 CANBK  EQ     2019-07-17       Annual General ~     208       216.     206. 
#>  7 MGL    EQ     2019-08-29       Annual General ~    1025.     1046     1017. 
#>  8 ADANI~ EQ     2019-07-26       Annual General ~     368.      378.     365. 
#>  9 JUBLF~ EQ     2019-09-16       Annual General ~    1579      1617     1572  
#> 10 PFC    EQ     2019-08-20       Annual General ~     109.      113.     108  
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# top losers
nse_fo_top_losers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 IDEA   EQ     2019-08-16       Annual General ~        6.5        6.7
#>  2 IBULH~ EQ     2019-11-18       Interim Dividen~      275        277. 
#>  3 APOLL~ EQ     2019-09-12       Annual General ~     1466       1474. 
#>  4 TCS    EQ     2019-10-17       Interim Dividen~     2128.      2128. 
#>  5 RBLBA~ EQ     2019-07-01       Annual General ~      339.       346. 
#>  6 CIPLA  EQ     2019-07-31       Annual General ~      457.       458. 
#>  7 APOLL~ EQ     2019-07-23       Annual General ~      174.       174. 
#>  8 MUTHO~ EQ     2019-09-19       Annual General ~      693        694. 
#>  9 HCLTE~ EQ     2019-12-05       Bonus 1:1             563.       563. 
#> 10 UJJIV~ EQ     2019-07-25       Annual General ~      352.       353. 
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_fo_top_losers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 IDEA   EQ     2019-08-16       Annual General ~       6.5       6.7      5.9
#>  2 IBULH~ EQ     2019-11-18       Interim Dividen~     275       277.     254  
#>  3 APOLL~ EQ     2019-09-12       Annual General ~    1466      1474.    1403. 
#>  4 TCS    EQ     2019-10-17       Interim Dividen~    2128.     2128.    2070. 
#>  5 RBLBA~ EQ     2019-07-01       Annual General ~     339.      346.     325. 
#>  6 CIPLA  EQ     2019-07-31       Annual General ~     457.      458.     445. 
#>  7 APOLL~ EQ     2019-07-23       Annual General ~     174.      174.     170. 
#>  8 MUTHO~ EQ     2019-09-19       Annual General ~     693       694.     680. 
#>  9 HCLTE~ EQ     2019-12-05       Bonus 1:1            563.      563.     547  
#> 10 UJJIV~ EQ     2019-07-25       Annual General ~     352.      353.     344  
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# }