Top F&O gainers and losers

Top futures and options gainers and losers for the last trading session.

nse_fo_top_gainers(clean_names = TRUE)

nse_fo_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_fo_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1  IBULHSGFIN     EQ                  2020-09-03
#> 2    BOSCHLTD     EQ                  2020-08-03
#> 3    CHOLAFIN     EQ                  2021-02-09
#> 4         PNB     EQ                  2020-07-27
#> 5      M&MFIN     EQ                  2020-07-22
#> 6     RBLBANK     EQ                  2020-07-09
#> 7       CIPLA     EQ                  2020-08-11
#> 8         M&M     EQ                  2020-07-16
#> 9    RAMCOCEM     EQ                  2021-03-22
#> 10       BPCL     EQ                  2021-03-25
#>                                 last_corp_announcement open_price high_price
#> 1                               Annual General Meeting     197.45     216.45
#> 2   Annual General Meeting/Dividend - Rs 105 Per Share   14770.00   16250.00
#> 3                 Interimdividend - Rs 1.30  Per Share     547.95     573.70
#> 4                               Annual General Meeting      36.50      37.90
#> 5                         Rights 1:1 @ Premium Rs 48/-     154.25     161.00
#> 6                               Annual General Meeting     201.00     207.85
#> 7                               Annual General Meeting     906.00     927.90
#> 8  Annual General Meeting/Dividend - Rs 2.35 Per Share     784.05     809.55
#> 9                   Interim Dividend - Rs  3 Per Share     985.00    1005.00
#> 10                   Interim Dividend - Rs 5 Per Share     448.00     457.85
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1     192.50            210.90           193.25           9.13        78804626
#> 2   14550.20          15845.00         14780.55           7.20          534380
#> 3     545.00            570.70           544.05           4.90         9211100
#> 4      35.95             37.90            36.20           4.70       339156585
#> 5     152.60            159.95           154.20           3.73        21575696
#> 6     200.35            204.95           199.95           2.50        19971105
#> 7     905.00            925.50           903.95           2.38        12378690
#> 8     782.00            803.00           784.80           2.32        10092542
#> 9     970.10            999.00           977.95           2.15          647692
#> 10    441.50            455.80           446.45           2.09         9399907
#>    turnover_in_lakhs
#> 1          162353.29
#> 2           84012.34
#> 3           51728.62
#> 4          126098.42
#> 5           34206.11
#> 6           40808.96
#> 7          113846.81
#> 8           80881.63
#> 9            6410.34
#> 10          42447.16

# retain original column names as returned by NSE
nse_fo_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1  IBULHSGFIN     EQ               2020-09-03
#> 2    BOSCHLTD     EQ               2020-08-03
#> 3    CHOLAFIN     EQ               2021-02-09
#> 4         PNB     EQ               2020-07-27
#> 5      M&MFIN     EQ               2020-07-22
#> 6     RBLBANK     EQ               2020-07-09
#> 7       CIPLA     EQ               2020-08-11
#> 8         M&M     EQ               2020-07-16
#> 9    RAMCOCEM     EQ               2021-03-22
#> 10       BPCL     EQ               2021-03-25
#>                                   lastCorpAnnouncement openPrice highPrice
#> 1                               Annual General Meeting    197.45    216.45
#> 2   Annual General Meeting/Dividend - Rs 105 Per Share  14770.00  16250.00
#> 3                 Interimdividend - Rs 1.30  Per Share    547.95    573.70
#> 4                               Annual General Meeting     36.50     37.90
#> 5                         Rights 1:1 @ Premium Rs 48/-    154.25    161.00
#> 6                               Annual General Meeting    201.00    207.85
#> 7                               Annual General Meeting    906.00    927.90
#> 8  Annual General Meeting/Dividend - Rs 2.35 Per Share    784.05    809.55
#> 9                   Interim Dividend - Rs  3 Per Share    985.00   1005.00
#> 10                   Interim Dividend - Rs 5 Per Share    448.00    457.85
#>    lowPrice      ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1    192.50   210.90        193.25     9.13       78804626       162353.29
#> 2  14550.20 15845.00      14780.55     7.20         534380        84012.34
#> 3    545.00   570.70        544.05     4.90        9211100        51728.62
#> 4     35.95    37.90         36.20     4.70      339156585       126098.42
#> 5    152.60   159.95        154.20     3.73       21575696        34206.11
#> 6    200.35   204.95        199.95     2.50       19971105        40808.96
#> 7    905.00   925.50        903.95     2.38       12378690       113846.81
#> 8    782.00   803.00        784.80     2.32       10092542        80881.63
#> 9    970.10   999.00        977.95     2.15         647692         6410.34
#> 10   441.50   455.80        446.45     2.09        9399907        42447.16

# top losers
nse_fo_top_losers()

#>        symbol series last_corp_announcement_date
#> 1        SAIL     EQ                  2021-02-04
#> 2   TATASTEEL     EQ                  2021-06-17
#> 3  JINDALSTEL     EQ                  2020-09-23
#> 4    HINDALCO     EQ                  2020-09-03
#> 5   COALINDIA     EQ                  2021-03-15
#> 6        GAIL     EQ                  2021-03-22
#> 7        NMDC     EQ                  2021-03-22
#> 8  BHARATFORG     EQ                  2020-03-03
#> 9    GMRINFRA     EQ                  2020-09-10
#> 10  BRITANNIA     EQ                  2021-05-25
#>                                               last_corp_announcement open_price
#> 1                              Interim Dividend - Rs  1.00 Per Share     125.00
#> 2                  Annual General Meeting/Dividend - Rs 25 Per Share    1131.10
#> 3                                             Annual General Meeting     408.50
#> 4                   Annual General Meeting/Dividend - Re 1 Per Share     390.00
#> 5                                  Interim Dividend - Rs 5 Per Share     152.40
#> 6                                Interim Dividend - Rs 2.5 Per Share     151.95
#> 7                              Interim Dividend - Rs 7.76  Per Share     180.00
#> 8                                  Interim Dividend - Rs 2 Per Share     681.00
#> 9                                             Annual General Meeting      26.50
#> 10 Scheme Of Arangement- Bonus - 1 Debenture For 1 Equity Share Held    3500.00
#>    high_price low_price last_traded_price prev_close_price percent_change
#> 1      125.40    121.00            121.30           128.60          -5.68
#> 2     1135.50   1098.05           1104.90          1164.25          -5.10
#> 3      418.25    390.00            401.40           421.85          -4.85
#> 4      393.20    379.45            386.20           403.30          -4.24
#> 5      152.40    146.05            146.65           151.85          -3.42
#> 6      152.30    146.75            147.70           152.80          -3.34
#> 7      183.70    176.70            178.45           184.35          -3.20
#> 8      684.35    654.00            658.00           679.35          -3.14
#> 9       26.85     25.90             26.00            26.80          -2.99
#> 10    3517.70   3415.00           3428.90          3534.40          -2.98
#>    traded_quantity turnover_in_lakhs
#> 1         59942600          73969.17
#> 2         28446253         316697.82
#> 3         35292844         142777.20
#> 4         30689219         118810.24
#> 5         22983457          34192.49
#> 6         14905120          22122.18
#> 7         19475577          35036.56
#> 8          1943758          12917.24
#> 9         12506758           3304.29
#> 10         1046540          36059.48

# retain original column names as returned by NSE
nse_fo_top_losers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1        SAIL     EQ               2021-02-04
#> 2   TATASTEEL     EQ               2021-06-17
#> 3  JINDALSTEL     EQ               2020-09-23
#> 4    HINDALCO     EQ               2020-09-03
#> 5   COALINDIA     EQ               2021-03-15
#> 6        GAIL     EQ               2021-03-22
#> 7        NMDC     EQ               2021-03-22
#> 8  BHARATFORG     EQ               2020-03-03
#> 9    GMRINFRA     EQ               2020-09-10
#> 10  BRITANNIA     EQ               2021-05-25
#>                                                 lastCorpAnnouncement openPrice
#> 1                              Interim Dividend - Rs  1.00 Per Share    125.00
#> 2                  Annual General Meeting/Dividend - Rs 25 Per Share   1131.10
#> 3                                             Annual General Meeting    408.50
#> 4                   Annual General Meeting/Dividend - Re 1 Per Share    390.00
#> 5                                  Interim Dividend - Rs 5 Per Share    152.40
#> 6                                Interim Dividend - Rs 2.5 Per Share    151.95
#> 7                              Interim Dividend - Rs 7.76  Per Share    180.00
#> 8                                  Interim Dividend - Rs 2 Per Share    681.00
#> 9                                             Annual General Meeting     26.50
#> 10 Scheme Of Arangement- Bonus - 1 Debenture For 1 Equity Share Held   3500.00
#>    highPrice lowPrice     ltp previousPrice netPrice tradedQuantity
#> 1     125.40   121.00  121.30        128.60    -5.68       59942600
#> 2    1135.50  1098.05 1104.90       1164.25    -5.10       28446253
#> 3     418.25   390.00  401.40        421.85    -4.85       35292844
#> 4     393.20   379.45  386.20        403.30    -4.24       30689219
#> 5     152.40   146.05  146.65        151.85    -3.42       22983457
#> 6     152.30   146.75  147.70        152.80    -3.34       14905120
#> 7     183.70   176.70  178.45        184.35    -3.20       19475577
#> 8     684.35   654.00  658.00        679.35    -3.14        1943758
#> 9      26.85    25.90   26.00         26.80    -2.99       12506758
#> 10   3517.70  3415.00 3428.90       3534.40    -2.98        1046540
#>    turnoverInLakhs
#> 1         73969.17
#> 2        316697.82
#> 3        142777.20
#> 4        118810.24
#> 5         34192.49
#> 6         22122.18
#> 7         35036.56
#> 8         12917.24
#> 9          3304.29
#> 10        36059.48

# }