Most actively traded stocks

Fetch most actively traded stocks in a month on NSE.

nse_stock_most_traded(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

security

Name of the firm.

share_turnover

Share in total turnover (percentage).

traded_quantity

Total traded quantity (in lakhs)

no_of_trades

Number of trades.

avg_daily_turnover

Average daily turnover (in crores).

turnover

Turnover (in crores.)

Examples

# \donttest{
nse_stock_most_traded()

#>                                         security share_turnover traded_quantity
#> 1                             Tata Steel Limited           3.48            5045
#> 2  Adani Ports and Special Economic Zone Limited           3.22            5547
#> 3                            Tata Motors Limited           2.61           11479
#> 4                              HDFC Bank Limited           2.53            2359
#> 5                              JSW Steel Limited           2.50            5392
#> 6                            State Bank of India           2.42            9291
#> 7                          Bajaj Finance Limited           2.42             651
#> 8                             ICICI Bank Limited           2.36            5423
#> 9                    Reliance Industries Limited           2.23            1510
#> 10                             Axis Bank Limited           1.91            3726
#> 11                   TOTAL of Top Ten securities          25.68           50423
#> 12                                         TOTAL         100.00          500847
#>    no_of_trades avg_daily_turnonver   turnover
#> 1       6297239             2439.37   46347.95
#> 2       6740117             2254.63   42838.03
#> 3       6261597             1829.61   34762.50
#> 4       5137713             1770.64   33642.10
#> 5       5326300             1751.55   33279.37
#> 6       5946848             1696.45   32232.60
#> 7       4065851             1691.77   32143.68
#> 8       5594907             1652.91   31405.21
#> 9       4398258             1560.00   29640.06
#> 10      4429981             1340.99   25478.72
#> 11     54198811            17987.91  341770.24
#> 12    385038115            70036.15 1330686.78

# retain original column names as returned by NSE
nse_stock_most_traded(clean_names = FALSE)

#>                                          skipped sharetotal trdQty nooftrades
#> 1                             Tata Steel Limited       3.48   5045    6297239
#> 2  Adani Ports and Special Economic Zone Limited       3.22   5547    6740117
#> 3                            Tata Motors Limited       2.61  11479    6261597
#> 4                              HDFC Bank Limited       2.53   2359    5137713
#> 5                              JSW Steel Limited       2.50   5392    5326300
#> 6                            State Bank of India       2.42   9291    5946848
#> 7                          Bajaj Finance Limited       2.42    651    4065851
#> 8                             ICICI Bank Limited       2.36   5423    5594907
#> 9                    Reliance Industries Limited       2.23   1510    4398258
#> 10                             Axis Bank Limited       1.91   3726    4429981
#> 11                   TOTAL of Top Ten securities      25.68  50423   54198811
#> 12                                         TOTAL     100.00 500847  385038115
#>    avgdailyturn   turnover
#> 1       2439.37   46347.95
#> 2       2254.63   42838.03
#> 3       1829.61   34762.50
#> 4       1770.64   33642.10
#> 5       1751.55   33279.37
#> 6       1696.45   32232.60
#> 7       1691.77   32143.68
#> 8       1652.91   31405.21
#> 9       1560.00   29640.06
#> 10      1340.99   25478.72
#> 11     17987.91  341770.24
#> 12     70036.15 1330686.78

# }