Most actively traded stocks

Fetch most actively traded stocks in a month on NSE.

nse_stock_most_traded(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

security

Name of the firm.

share_turnover

Share in total turnover (percentage).

traded_quantity

Total traded quantity (in lakhs)

no_of_trades

Number of trades.

avg_daily_turnover

Average daily turnover (in crores).

turnover

Turnover (in crores.)

Examples

# \donttest{
nse_stock_most_traded()

#> # A tibble: 12 x 6
#>    security share_turnover traded_quantity no_of_trades avg_daily_turno~
#>    <chr>             <dbl>           <dbl>        <dbl>            <dbl>
#>  1 Yes Ban~           4.38           57140     11058530            1423.
#>  2 Relianc~           3.93            1715      3803882            1276.
#>  3 Indiabu~           3.91            9077      8060572            1271.
#>  4 State B~           3.43            7114      4624946            1114.
#>  5 Tata Co~           2.69             866      2891204             875.
#>  6 ICICI B~           2.51            3204      2969145             814.
#>  7 HDFC Li~           2.33             670      2510572             756.
#>  8 Tata St~           2.26            3541      3008166             733.
#>  9 HDFC Ba~           2.18            1170      2969084             707.
#> 10 Tata Mo~           2.07            8128      3536571             671.
#> 11 TOTAL o~          29.7            92625     45432672            9640.
#> 12 TOTAL            100             391115    223723721           32475.
#> # ... with 1 more variable: turnover <dbl>

# retain original column names as returned by NSE
nse_stock_most_traded(clean_names = FALSE)

#> # A tibble: 12 x 6
#>    security                   sharetotal trdQty nooftrades avgdailyturn turnover
#>    <chr>                           <dbl>  <dbl>      <dbl>        <dbl>    <dbl>
#>  1 Yes Bank Limited                 4.38  57140   11058530        1423.   29880.
#>  2 Reliance Industries Limit~       3.93   1715    3803882        1276.   26799.
#>  3 Indiabulls Hsg Fin Limited       3.91   9077    8060572        1271.   26683.
#>  4 State Bank of India              3.43   7114    4624946        1114.   23391.
#>  5 Tata Consultancy Services~       2.69    866    2891204         875.   18367.
#>  6 ICICI Bank Limited               2.51   3204    2969145         814.   17098.
#>  7 HDFC Limited                     2.33    670    2510572         756.   15868.
#>  8 Tata Steel Limited               2.26   3541    3008166         733.   15396.
#>  9 HDFC Bank Limited                2.18   1170    2969084         707.   14848.
#> 10 Tata Motors Limited              2.07   8128    3536571         671.   14100.
#> 11 TOTAL of Top Ten securiti~      29.7   92625   45432672        9640.  202431.
#> 12 TOTAL                          100    391115  223723721       32475.  681983.

# }