Most actively traded stocks

Fetch most actively traded stocks in a month on NSE.

nse_stock_most_traded(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

security

Name of the firm.

share_turnover

Share in total turnover (percentage).

traded_quantity

Total traded quantity (in lakhs)

no_of_trades

Number of trades.

avg_daily_turnover

Average daily turnover (in crores).

turnover

Turnover (in crores.)

Examples

# \donttest{
nse_stock_most_traded()

#>                       security share_turnover traded_quantity no_of_trades
#> 1          Tata Motors Limited           3.95           20118     10750273
#> 2          State Bank of India           3.58           15323     10397758
#> 3  Reliance Industries Limited           2.97            2411      6543286
#> 4           ICICI Bank Limited           2.28            5954      6508391
#> 5       Bharati Airtel Limited           2.18            6107      4987691
#> 6           Tata Steel Limited           2.00            4713      5165939
#> 7            HDFC Bank Limited           1.92            1994      4925517
#> 8            Axis Bank Limited           1.90            4162      5820333
#> 9        Bajaj Finance Limited           1.89             565      3138564
#> 10       Indusind Bank Limited           1.76            2782      4933750
#> 11 TOTAL of Top Ten securities          24.43           64129     63171502
#> 12                       TOTAL         100.00          716614    455786495
#>    avg_daily_turnonver   turnover
#> 1              3211.22   64224.31
#> 2              2910.30   58205.98
#> 3              2419.53   48390.54
#> 4              1854.63   37092.52
#> 5              1776.81   35536.11
#> 6              1630.03   32600.54
#> 7              1562.98   31259.67
#> 8              1545.32   30906.37
#> 9              1536.38   30727.60
#> 10             1430.42   28608.42
#> 11            19877.60  397552.06
#> 12            81373.18 1627463.62

# retain original column names as returned by NSE
nse_stock_most_traded(clean_names = FALSE)

#>                        skipped sharetotal trdQty nooftrades avgdailyturn
#> 1          Tata Motors Limited       3.95  20118   10750273      3211.22
#> 2          State Bank of India       3.58  15323   10397758      2910.30
#> 3  Reliance Industries Limited       2.97   2411    6543286      2419.53
#> 4           ICICI Bank Limited       2.28   5954    6508391      1854.63
#> 5       Bharati Airtel Limited       2.18   6107    4987691      1776.81
#> 6           Tata Steel Limited       2.00   4713    5165939      1630.03
#> 7            HDFC Bank Limited       1.92   1994    4925517      1562.98
#> 8            Axis Bank Limited       1.90   4162    5820333      1545.32
#> 9        Bajaj Finance Limited       1.89    565    3138564      1536.38
#> 10       Indusind Bank Limited       1.76   2782    4933750      1430.42
#> 11 TOTAL of Top Ten securities      24.43  64129   63171502     19877.60
#> 12                       TOTAL     100.00 716614  455786495     81373.18
#>      turnover
#> 1    64224.31
#> 2    58205.98
#> 3    48390.54
#> 4    37092.52
#> 5    35536.11
#> 6    32600.54
#> 7    31259.67
#> 8    30906.37
#> 9    30727.60
#> 10   28608.42
#> 11  397552.06
#> 12 1627463.62

# }