NSE top gainers & losers

Fetch top gainers and losers for the last trading session.

nse_stock_top_gainers(clean_names = TRUE)

nse_stock_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_stock_top_gainers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 VEDL   EQ     2019-07-04       Annual General ~       153.       157.
#>  2 ZEEL   EQ     2019-07-15       Annual General ~       265        271.
#>  3 ULTRA~ EQ     2019-07-10       Annual General ~      4200.      4254 
#>  4 UPL    EQ     2019-07-02       Bonus 1:2              585.       604 
#>  5 RELIA~ EQ     2019-08-02       Annual General ~      1519       1534.
#>  6 HDFCB~ EQ     2019-09-19       Face Value Spli~      1259.      1271.
#>  7 SUNPH~ EQ     2019-08-20       Annual General ~       441.       448 
#>  8 NTPC   EQ     2019-08-13       Dividend - Rs 2~       119.       122.
#>  9 HDFC   EQ     2019-07-19       Annual General ~      2401.      2429.
#> 10 ASIAN~ EQ     2019-10-30       Interim Dividen~      1711       1740.
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_stock_top_gainers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 VEDL   EQ     2019-07-04       Annual General ~      153.      157.     153.
#>  2 ZEEL   EQ     2019-07-15       Annual General ~      265       271.     263.
#>  3 ULTRA~ EQ     2019-07-10       Annual General ~     4200.     4254     4164.
#>  4 UPL    EQ     2019-07-02       Bonus 1:2             585.      604      583.
#>  5 RELIA~ EQ     2019-08-02       Annual General ~     1519      1534.    1514.
#>  6 HDFCB~ EQ     2019-09-19       Face Value Spli~     1259.     1271.    1252.
#>  7 SUNPH~ EQ     2019-08-20       Annual General ~      441.      448      438.
#>  8 NTPC   EQ     2019-08-13       Dividend - Rs 2~      119.      122.     119.
#>  9 HDFC   EQ     2019-07-19       Annual General ~     2401.     2429.    2380.
#> 10 ASIAN~ EQ     2019-10-30       Interim Dividen~     1711      1740.    1711 
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# top losers
nse_stock_top_losers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 INFRA~ EQ     2019-12-19       Interim Dividen~       242.       247.
#>  2 BPCL   EQ     2019-08-21       Dividend - Rs 8~       468.       475.
#>  3 INFY   EQ     2019-10-23       Interim Dividen~       738        743.
#>  4 BHART~ EQ     2019-04-23       Rights 19:67 @ ~       452        455.
#>  5 POWER~ EQ     2019-08-19       Annual General ~       193.       194.
#>  6 INDUS~ EQ     2019-08-08       Annual General ~      1488.      1508 
#>  7 TATAM~ EQ     2016-07-18       Dividend - Re 0~       187        189.
#>  8 NESTL~ EQ     2019-12-10       Interim Dividen~     14424.     14547.
#>  9 HEROM~ EQ     2019-07-16       Annual General ~      2378       2400.
#> 10 ONGC   EQ     2019-08-22       Annual General ~       126.       128.
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_stock_top_losers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 INFRA~ EQ     2019-12-19       Interim Dividen~      242.      247.     237.
#>  2 BPCL   EQ     2019-08-21       Dividend - Rs 8~      468.      475.     460.
#>  3 INFY   EQ     2019-10-23       Interim Dividen~      738       743.     725.
#>  4 BHART~ EQ     2019-04-23       Rights 19:67 @ ~      452       455.     443.
#>  5 POWER~ EQ     2019-08-19       Annual General ~      193.      194.     191.
#>  6 INDUS~ EQ     2019-08-08       Annual General ~     1488.     1508     1442.
#>  7 TATAM~ EQ     2016-07-18       Dividend - Re 0~      187       189.     182.
#>  8 NESTL~ EQ     2019-12-10       Interim Dividen~    14424.    14547.   14261.
#>  9 HEROM~ EQ     2019-07-16       Annual General ~     2378      2400.    2337 
#> 10 ONGC   EQ     2019-08-22       Annual General ~      126.      128.     125.
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# }