NSE top gainers & losers

Fetch top gainers and losers for the last trading session.

nse_stock_top_gainers(clean_names = TRUE)

nse_stock_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_stock_top_gainers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 ADANI~ EQ     2019-07-26       Annual General ~       368.      378. 
#>  2 HDFC   EQ     2019-07-19       Annual General ~      2263.     2310  
#>  3 AXISB~ EQ     2019-07-04       Annual General ~       713.      732. 
#>  4 BPCL   EQ     2019-08-21       Dividend - Rs 8~       487.      500. 
#>  5 MARUTI EQ     2019-08-14       Dividend - Rs 8~      7007      7108  
#>  6 YESBA~ EQ     2019-06-03       Annual General ~        56        58.8
#>  7 JSWST~ EQ     2019-07-08       Annual General ~       253       259. 
#>  8 UPL    EQ     2019-07-02       Bonus 1:2              565       574. 
#>  9 POWER~ EQ     2019-08-19       Annual General ~       187       189. 
#> 10 TATAS~ EQ     2019-07-04       Annual General ~       403       412. 
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_stock_top_gainers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 ADANI~ EQ     2019-07-26       Annual General ~      368.     378.     365. 
#>  2 HDFC   EQ     2019-07-19       Annual General ~     2263.    2310     2260. 
#>  3 AXISB~ EQ     2019-07-04       Annual General ~      713.     732.     710  
#>  4 BPCL   EQ     2019-08-21       Dividend - Rs 8~      487.     500.     486. 
#>  5 MARUTI EQ     2019-08-14       Dividend - Rs 8~     7007     7108     6952. 
#>  6 YESBA~ EQ     2019-06-03       Annual General ~       56       58.8     53.8
#>  7 JSWST~ EQ     2019-07-08       Annual General ~      253      259.     252. 
#>  8 UPL    EQ     2019-07-02       Bonus 1:2             565      574.     560. 
#>  9 POWER~ EQ     2019-08-19       Annual General ~      187      189.     186. 
#> 10 TATAS~ EQ     2019-07-04       Annual General ~      403      412.     402. 
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# top losers
nse_stock_top_losers()

#> # A tibble: 10 x 12
#>    symbol series last_corp_annou~ last_corp_annou~ open_price high_price
#>    <chr>  <chr>  <date>           <chr>                 <dbl>      <dbl>
#>  1 TCS    EQ     2019-10-17       Interim Dividen~      2128.      2128.
#>  2 CIPLA  EQ     2019-07-31       Annual General ~       457.       458.
#>  3 HCLTE~ EQ     2019-12-05       Bonus 1:1              563.       563.
#>  4 HINDU~ EQ     2019-10-23       Interim Dividen~      2030       2032.
#>  5 NESTL~ EQ     2019-12-10       Interim Dividen~     14230      14230 
#>  6 KOTAK~ EQ     2019-07-12       Annual General ~      1678.      1679 
#>  7 INDUS~ EQ     2019-08-08       Annual General ~      1457       1485.
#>  8 SBIN   EQ     2018-06-15       Annual General ~       319.       328.
#>  9 ITC    EQ     2019-05-22       Dividend - Rs 5~       243        243.
#> 10 TECHM  EQ     2019-07-25       Annual General ~       759.       759.
#> # ... with 6 more variables: low_price <dbl>, last_traded_price <dbl>,
#> #   prev_close_price <dbl>, percent_change <dbl>, traded_quantity <dbl>,
#> #   turnover_in_lakhs <dbl>

# retain original column names as returned by NSE
nse_stock_top_losers(clean_names = FALSE)

#> # A tibble: 10 x 12
#>    symbol series lastCorpAnnounc~ lastCorpAnnounc~ openPrice highPrice lowPrice
#>    <chr>  <chr>  <date>           <chr>                <dbl>     <dbl>    <dbl>
#>  1 TCS    EQ     2019-10-17       Interim Dividen~     2128.     2128.    2070.
#>  2 CIPLA  EQ     2019-07-31       Annual General ~      457.      458.     445.
#>  3 HCLTE~ EQ     2019-12-05       Bonus 1:1             563.      563.     547 
#>  4 HINDU~ EQ     2019-10-23       Interim Dividen~     2030      2032.    1991.
#>  5 NESTL~ EQ     2019-12-10       Interim Dividen~    14230     14230    14005 
#>  6 KOTAK~ EQ     2019-07-12       Annual General ~     1678.     1679     1656.
#>  7 INDUS~ EQ     2019-08-08       Annual General ~     1457      1485.    1445.
#>  8 SBIN   EQ     2018-06-15       Annual General ~      319.      328.     315.
#>  9 ITC    EQ     2019-05-22       Dividend - Rs 5~      243       243.     241.
#> 10 TECHM  EQ     2019-07-25       Annual General ~      759.      759.     746.
#> # ... with 5 more variables: ltp <dbl>, previousPrice <dbl>, netPrice <dbl>,
#> #   tradedQuantity <dbl>, turnoverInLakhs <dbl>

# }