NSE top gainers & losers

Fetch top gainers and losers for the last trading session.

nse_stock_top_gainers(clean_names = TRUE)

nse_stock_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_stock_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1  BAJAJFINSV     EQ                  2020-03-03
#> 2   COALINDIA     EQ                  2020-03-19
#> 3  BAJFINANCE     EQ                  2020-03-03
#> 4        ZEEL     EQ                  2019-07-15
#> 5        VEDL     EQ                  2020-03-05
#> 6   KOTAKBANK     EQ                  2019-07-12
#> 7        SBIN     EQ                  2018-06-15
#> 8    HDFCBANK     EQ                  2019-09-19
#> 9   POWERGRID     EQ                  2020-03-16
#> 10   AXISBANK     EQ                  2019-07-04
#>                                                     last_corp_announcement
#> 1                                        Interim Dividend - Rs 5 Per Share
#> 2                                       Interim Dividend - Rs 12 Per Share
#> 3                                       Interim Dividend - Rs 10 Per Share
#> 4                     Annual General Meeting/ Dividend - Rs 3.50 Per Share
#> 5                                      Interim Dividend - Rs 3.9 Per Share
#> 6                        Annual General Meeting/Dividend Rs 0.80 Per Share
#> 7           Annual General Meeting/ Change In Registrar And Transfer Agent
#> 8  Face Value Split (Sub-Division) - From Rs 2 Per Share To Rs 1 Per Share
#> 9                                     Interim Dividend - Rs 5.96 Per Share
#> 10                       Annual General Meeting/ Dividend - Re 1 Per Share
#>    open_price high_price low_price last_traded_price prev_close_price
#> 1     5036.95    5499.00   5021.55           5450.75          5036.95
#> 2      129.00     139.70    128.15            137.65           129.45
#> 3     2375.00    2541.70   2373.00           2527.20          2396.00
#> 4      166.50     175.45    165.50            173.95           165.00
#> 5      102.80     107.80    102.00            106.90           102.10
#> 6     1240.00    1309.90   1228.35           1299.95          1243.80
#> 7      172.40     180.65    171.10            180.00           172.95
#> 8      980.00    1024.90    972.00           1017.60           979.25
#> 9      160.70     166.30    159.50            165.70           159.50
#> 10     386.00     408.05    384.00            404.75           389.60
#>    percent_change traded_quantity turnover_in_lakhs
#> 1            8.22         1806208          95403.73
#> 2            6.33        34587382          47031.92
#> 3            5.48        12738830         315647.83
#> 4            5.42        37917230          65103.88
#> 5            4.70        27586795          29197.86
#> 6            4.51         6672822          84453.24
#> 7            4.08        58262367         102757.34
#> 8            3.92        20435469         204078.81
#> 9            3.89        11093113          18153.88
#> 10           3.89        43774453         173556.95

# retain original column names as returned by NSE
nse_stock_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1  BAJAJFINSV     EQ               2020-03-03
#> 2   COALINDIA     EQ               2020-03-19
#> 3  BAJFINANCE     EQ               2020-03-03
#> 4        ZEEL     EQ               2019-07-15
#> 5        VEDL     EQ               2020-03-05
#> 6   KOTAKBANK     EQ               2019-07-12
#> 7        SBIN     EQ               2018-06-15
#> 8    HDFCBANK     EQ               2019-09-19
#> 9   POWERGRID     EQ               2020-03-16
#> 10   AXISBANK     EQ               2019-07-04
#>                                                       lastCorpAnnouncement
#> 1                                        Interim Dividend - Rs 5 Per Share
#> 2                                       Interim Dividend - Rs 12 Per Share
#> 3                                       Interim Dividend - Rs 10 Per Share
#> 4                     Annual General Meeting/ Dividend - Rs 3.50 Per Share
#> 5                                      Interim Dividend - Rs 3.9 Per Share
#> 6                        Annual General Meeting/Dividend Rs 0.80 Per Share
#> 7           Annual General Meeting/ Change In Registrar And Transfer Agent
#> 8  Face Value Split (Sub-Division) - From Rs 2 Per Share To Rs 1 Per Share
#> 9                                     Interim Dividend - Rs 5.96 Per Share
#> 10                       Annual General Meeting/ Dividend - Re 1 Per Share
#>    openPrice highPrice lowPrice     ltp previousPrice netPrice tradedQuantity
#> 1    5036.95   5499.00  5021.55 5450.75       5036.95     8.22        1806208
#> 2     129.00    139.70   128.15  137.65        129.45     6.33       34587382
#> 3    2375.00   2541.70  2373.00 2527.20       2396.00     5.48       12738830
#> 4     166.50    175.45   165.50  173.95        165.00     5.42       37917230
#> 5     102.80    107.80   102.00  106.90        102.10     4.70       27586795
#> 6    1240.00   1309.90  1228.35 1299.95       1243.80     4.51        6672822
#> 7     172.40    180.65   171.10  180.00        172.95     4.08       58262367
#> 8     980.00   1024.90   972.00 1017.60        979.25     3.92       20435469
#> 9     160.70    166.30   159.50  165.70        159.50     3.89       11093113
#> 10    386.00    408.05   384.00  404.75        389.60     3.89       43774453
#>    turnoverInLakhs
#> 1         95403.73
#> 2         47031.92
#> 3        315647.83
#> 4         65103.88
#> 5         29197.86
#> 6         84453.24
#> 7        102757.34
#> 8        204078.81
#> 9         18153.88
#> 10       173556.95

# top losers
nse_stock_top_losers()

#>        symbol series last_corp_announcement_date
#> 1         TCS     EQ                  2020-06-03
#> 2  HINDUNILVR     EQ                  2020-06-19
#> 3  BHARTIARTL     EQ                  2019-04-23
#> 4  ASIANPAINT     EQ                  2020-03-03
#> 5  BAJAJ-AUTO     EQ                  2020-03-03
#> 6        ONGC     EQ                  2020-03-23
#> 7   SUNPHARMA     EQ                  2020-02-17
#> 8  HEROMOTOCO     EQ                  2020-02-17
#> 9      MARUTI     EQ                  2019-08-14
#> 10  BRITANNIA     EQ                  2020-06-29
#>                               last_corp_announcement open_price high_price
#> 1                          Dividend - Rs 6 Per Share    2045.00    2059.20
#> 2  Annual General Meeting/Dividend - Rs 14 Per Share    2070.25    2084.85
#> 3            Rights 19:67 @ Premium Rs 215 Per Share     564.25     572.90
#> 4               Interim Dividend - Rs 7.15 Per Share    1595.40    1612.20
#> 5                Interim Dividend - Rs 120 Per Share    2689.45    2697.15
#> 6                  Interim Dividend - Rs 5 Per Share      83.00      83.95
#> 7                  Interim Dividend - Rs 3 Per Share     485.65     488.50
#> 8                    Interim Dividend - Rs 65 Per Sh    2360.25    2380.00
#> 9                         Dividend - Rs 80 Per Share    5695.00    5758.00
#> 10                            Annual General Meeting    3415.00    3457.00
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1    2025.20           2036.20          2047.80          -0.57         1944471
#> 2    2054.05           2067.00          2078.25          -0.54         2326992
#> 3     558.20            560.95           563.45          -0.44        19329827
#> 4    1587.30           1589.50          1595.40          -0.37         1417293
#> 5    2646.15           2680.00          2689.45          -0.35          579581
#> 6      81.80             83.95            84.20          -0.30        18460702
#> 7     483.00            484.45           485.65          -0.25         4077225
#> 8    2342.20           2359.85          2363.30          -0.15         1087905
#> 9    5641.00           5724.00          5726.20          -0.04         1580217
#> 10   3381.55           3401.00          3401.45          -0.01          898020
#>    turnover_in_lakhs
#> 1           39660.01
#> 2           48115.45
#> 3          109234.79
#> 4           22650.75
#> 5           15503.21
#> 6           15313.15
#> 7           19782.29
#> 8           25698.60
#> 9           90144.11
#> 10          30633.53

# retain original column names as returned by NSE
nse_stock_top_losers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1         TCS     EQ               2020-06-03
#> 2  HINDUNILVR     EQ               2020-06-19
#> 3  BHARTIARTL     EQ               2019-04-23
#> 4  ASIANPAINT     EQ               2020-03-03
#> 5  BAJAJ-AUTO     EQ               2020-03-03
#> 6        ONGC     EQ               2020-03-23
#> 7   SUNPHARMA     EQ               2020-02-17
#> 8  HEROMOTOCO     EQ               2020-02-17
#> 9      MARUTI     EQ               2019-08-14
#> 10  BRITANNIA     EQ               2020-06-29
#>                                 lastCorpAnnouncement openPrice highPrice
#> 1                          Dividend - Rs 6 Per Share   2045.00   2059.20
#> 2  Annual General Meeting/Dividend - Rs 14 Per Share   2070.25   2084.85
#> 3            Rights 19:67 @ Premium Rs 215 Per Share    564.25    572.90
#> 4               Interim Dividend - Rs 7.15 Per Share   1595.40   1612.20
#> 5                Interim Dividend - Rs 120 Per Share   2689.45   2697.15
#> 6                  Interim Dividend - Rs 5 Per Share     83.00     83.95
#> 7                  Interim Dividend - Rs 3 Per Share    485.65    488.50
#> 8                    Interim Dividend - Rs 65 Per Sh   2360.25   2380.00
#> 9                         Dividend - Rs 80 Per Share   5695.00   5758.00
#> 10                            Annual General Meeting   3415.00   3457.00
#>    lowPrice     ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1   2025.20 2036.20       2047.80    -0.57        1944471        39660.01
#> 2   2054.05 2067.00       2078.25    -0.54        2326992        48115.45
#> 3    558.20  560.95        563.45    -0.44       19329827       109234.79
#> 4   1587.30 1589.50       1595.40    -0.37        1417293        22650.75
#> 5   2646.15 2680.00       2689.45    -0.35         579581        15503.21
#> 6     81.80   83.95         84.20    -0.30       18460702        15313.15
#> 7    483.00  484.45        485.65    -0.25        4077225        19782.29
#> 8   2342.20 2359.85       2363.30    -0.15        1087905        25698.60
#> 9   5641.00 5724.00       5726.20    -0.04        1580217        90144.11
#> 10  3381.55 3401.00       3401.45    -0.01         898020        30633.53

# }