NSE top gainers & losers

Fetch top gainers and losers for the last trading session.

nse_stock_top_gainers(clean_names = TRUE)

nse_stock_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_stock_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1  ULTRACEMCO     EQ                  2020-07-29
#> 2  ADANIPORTS     EQ                  2020-06-17
#> 3    SHREECEM     EQ                  2020-02-24
#> 4      GRASIM     EQ                  2020-08-26
#> 5     DRREDDY     EQ                  2020-07-13
#> 6     SBILIFE     EQ                  2019-04-03
#> 7       WIPRO     EQ                  2021-01-22
#> 8  ASIANPAINT     EQ                  2020-10-28
#> 9   BRITANNIA     EQ                  2020-08-26
#> 10 HEROMOTOCO     EQ                  2021-02-15
#>                                                  last_corp_announcement
#> 1                     Annual General Meeting/Dividend - Rs 13 Per Share
#> 2                                                Annual General Meeting
#> 3                                   Interim Dividend - Rs 110 Per Share
#> 4                                             Dividend - Rs 4 Per Share
#> 5                     Annual General Meeting/Dividend - Rs 25 Per Share
#> 6                                     Interim Dividend - Rs 2 Per Share
#> 7                                     Interim Dividend - Re 1 Per Share
#> 8                                  Interim Dividend - Rs 3.35 Per Share
#> 9                                    Interim Dividend - Rs 83 Per Share
#> 10 Interim Dividend - Rs 65 Per Share Special Dividend - Rs 5 Per Share
#>    open_price high_price low_price last_traded_price prev_close_price
#> 1     6429.00    6830.00   6394.45           6755.00          6500.50
#> 2      733.25     765.70    715.85            751.55           729.85
#> 3    27450.00   28850.00  27335.05          28500.00         27695.25
#> 4     1306.00    1360.00   1300.50           1353.10          1318.10
#> 5     4482.00    4583.35   4444.00           4567.00          4502.55
#> 6      899.70     924.80    892.40            915.00           906.45
#> 7      428.90     444.30    427.50            439.00           435.50
#> 8     2380.00    2435.00   2350.00           2416.00          2402.10
#> 9     3450.00    3514.80   3450.00           3500.05          3481.15
#> 10    3398.00    3460.00   3368.25           3429.90          3424.35
#>    percent_change traded_quantity turnover_in_lakhs
#> 1            3.92         1303144          86771.93
#> 2            2.97        30607959         229403.59
#> 3            2.91           83170          23524.20
#> 4            2.66         1950846          26068.37
#> 5            1.43         1472991          66779.67
#> 6            0.94         3140379          28790.37
#> 7            0.80        12963797          56913.66
#> 8            0.58         2149984          51837.83
#> 9            0.54          537283          18777.83
#> 10           0.16          934953          32039.62

# retain original column names as returned by NSE
nse_stock_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1  ULTRACEMCO     EQ               2020-07-29
#> 2  ADANIPORTS     EQ               2020-06-17
#> 3    SHREECEM     EQ               2020-02-24
#> 4      GRASIM     EQ               2020-08-26
#> 5     DRREDDY     EQ               2020-07-13
#> 6     SBILIFE     EQ               2019-04-03
#> 7       WIPRO     EQ               2021-01-22
#> 8  ASIANPAINT     EQ               2020-10-28
#> 9   BRITANNIA     EQ               2020-08-26
#> 10 HEROMOTOCO     EQ               2021-02-15
#>                                                    lastCorpAnnouncement
#> 1                     Annual General Meeting/Dividend - Rs 13 Per Share
#> 2                                                Annual General Meeting
#> 3                                   Interim Dividend - Rs 110 Per Share
#> 4                                             Dividend - Rs 4 Per Share
#> 5                     Annual General Meeting/Dividend - Rs 25 Per Share
#> 6                                     Interim Dividend - Rs 2 Per Share
#> 7                                     Interim Dividend - Re 1 Per Share
#> 8                                  Interim Dividend - Rs 3.35 Per Share
#> 9                                    Interim Dividend - Rs 83 Per Share
#> 10 Interim Dividend - Rs 65 Per Share Special Dividend - Rs 5 Per Share
#>    openPrice highPrice lowPrice      ltp previousPrice netPrice tradedQuantity
#> 1    6429.00   6830.00  6394.45  6755.00       6500.50     3.92        1303144
#> 2     733.25    765.70   715.85   751.55        729.85     2.97       30607959
#> 3   27450.00  28850.00 27335.05 28500.00      27695.25     2.91          83170
#> 4    1306.00   1360.00  1300.50  1353.10       1318.10     2.66        1950846
#> 5    4482.00   4583.35  4444.00  4567.00       4502.55     1.43        1472991
#> 6     899.70    924.80   892.40   915.00        906.45     0.94        3140379
#> 7     428.90    444.30   427.50   439.00        435.50     0.80       12963797
#> 8    2380.00   2435.00  2350.00  2416.00       2402.10     0.58        2149984
#> 9    3450.00   3514.80  3450.00  3500.05       3481.15     0.54         537283
#> 10   3398.00   3460.00  3368.25  3429.90       3424.35     0.16         934953
#>    turnoverInLakhs
#> 1         86771.93
#> 2        229403.59
#> 3         23524.20
#> 4         26068.37
#> 5         66779.67
#> 6         28790.37
#> 7         56913.66
#> 8         51837.83
#> 9         18777.83
#> 10        32039.62

# top losers
nse_stock_top_losers()

#>        symbol series last_corp_announcement_date
#> 1    JSWSTEEL     EQ                  2020-07-06
#> 2        HDFC     EQ                  2020-07-09
#> 3    HINDALCO     EQ                  2020-09-03
#> 4   TATASTEEL     EQ                  2020-08-06
#> 5  TATAMOTORS     EQ                  2016-07-18
#> 6  BAJAJFINSV     EQ                  2020-03-03
#> 7    AXISBANK     EQ                  2019-07-04
#> 8          LT     EQ                  2020-11-04
#> 9    HDFCBANK     EQ                  2019-09-19
#> 10       SBIN     EQ                  2018-06-15
#>                                                     last_corp_announcement
#> 1                                                Dividend - Rs 2 Per Share
#> 2                        Annual General Meeting/Dividend - Rs 21 Per Share
#> 3                         Annual General Meeting/Dividend - Re 1 Per Share
#> 4                                               Dividend - Rs 10 Per Share
#> 5                                           Dividend - Re 0.20/- Per Share
#> 6                                        Interim Dividend - Rs 5 Per Share
#> 7                        Annual General Meeting/ Dividend - Re 1 Per Share
#> 8                                       Special Dividend - Rs 18 Per Share
#> 9  Face Value Split (Sub-Division) - From Rs 2 Per Share To Rs 1 Per Share
#> 10          Annual General Meeting/ Change In Registrar And Transfer Agent
#>    open_price high_price low_price last_traded_price prev_close_price
#> 1      415.10     421.40    411.05             415.5           428.15
#> 2     2610.00    2610.00   2568.00            2584.9          2654.90
#> 3      352.00     357.70    347.00             349.9           359.35
#> 4      751.00     766.90    743.65             757.6           777.15
#> 5      339.00     345.40    337.30             340.0           348.50
#> 6    10150.00   10280.00  10077.75           10148.0         10395.60
#> 7      731.10     751.45    731.10             737.0           753.95
#> 8     1473.00    1494.35   1458.00            1464.0          1496.90
#> 9     1548.55    1571.00   1539.10            1553.0          1586.85
#> 10     398.70     403.50    394.45             396.8           405.05
#>    percent_change traded_quantity turnover_in_lakhs
#> 1           -2.95         8655068          35973.06
#> 2           -2.64         5315682         137372.11
#> 3           -2.63        13623173          47971.28
#> 4           -2.52        22873822         173209.73
#> 5           -2.44        59863169         204432.72
#> 6           -2.38          412117          41921.98
#> 7           -2.25        16896947         125052.62
#> 8           -2.20         3530366          51800.35
#> 9           -2.13         8145719         126449.25
#> 10          -2.04        43603653         174187.87

# retain original column names as returned by NSE
nse_stock_top_losers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1    JSWSTEEL     EQ               2020-07-06
#> 2        HDFC     EQ               2020-07-09
#> 3    HINDALCO     EQ               2020-09-03
#> 4   TATASTEEL     EQ               2020-08-06
#> 5  TATAMOTORS     EQ               2016-07-18
#> 6  BAJAJFINSV     EQ               2020-03-03
#> 7    AXISBANK     EQ               2019-07-04
#> 8          LT     EQ               2020-11-04
#> 9    HDFCBANK     EQ               2019-09-19
#> 10       SBIN     EQ               2018-06-15
#>                                                       lastCorpAnnouncement
#> 1                                                Dividend - Rs 2 Per Share
#> 2                        Annual General Meeting/Dividend - Rs 21 Per Share
#> 3                         Annual General Meeting/Dividend - Re 1 Per Share
#> 4                                               Dividend - Rs 10 Per Share
#> 5                                           Dividend - Re 0.20/- Per Share
#> 6                                        Interim Dividend - Rs 5 Per Share
#> 7                        Annual General Meeting/ Dividend - Re 1 Per Share
#> 8                                       Special Dividend - Rs 18 Per Share
#> 9  Face Value Split (Sub-Division) - From Rs 2 Per Share To Rs 1 Per Share
#> 10          Annual General Meeting/ Change In Registrar And Transfer Agent
#>    openPrice highPrice lowPrice     ltp previousPrice netPrice tradedQuantity
#> 1     415.10    421.40   411.05   415.5        428.15    -2.95        8655068
#> 2    2610.00   2610.00  2568.00  2584.9       2654.90    -2.64        5315682
#> 3     352.00    357.70   347.00   349.9        359.35    -2.63       13623173
#> 4     751.00    766.90   743.65   757.6        777.15    -2.52       22873822
#> 5     339.00    345.40   337.30   340.0        348.50    -2.44       59863169
#> 6   10150.00  10280.00 10077.75 10148.0      10395.60    -2.38         412117
#> 7     731.10    751.45   731.10   737.0        753.95    -2.25       16896947
#> 8    1473.00   1494.35  1458.00  1464.0       1496.90    -2.20        3530366
#> 9    1548.55   1571.00  1539.10  1553.0       1586.85    -2.13        8145719
#> 10    398.70    403.50   394.45   396.8        405.05    -2.04       43603653
#>    turnoverInLakhs
#> 1         35973.06
#> 2        137372.11
#> 3         47971.28
#> 4        173209.73
#> 5        204432.72
#> 6         41921.98
#> 7        125052.62
#> 8         51800.35
#> 9        126449.25
#> 10       174187.87

# }