NSE top gainers & losers

Fetch top gainers and losers for the last trading session.

nse_stock_top_gainers(clean_names = TRUE)

nse_stock_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_stock_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1   TATASTEEL     EQ                  2020-08-06
#> 2  BAJAJFINSV     EQ                  2020-03-03
#> 3    RELIANCE     EQ                  2020-07-02
#> 4  BAJFINANCE     EQ                  2020-03-03
#> 5         UPL     EQ                  2020-08-19
#> 6    HINDALCO     EQ                  2020-09-03
#> 7        HDFC     EQ                  2020-07-09
#> 8   ICICIBANK     EQ                  2019-07-22
#> 9    JSWSTEEL     EQ                  2020-07-06
#> 10    SBILIFE     EQ                  2019-04-03
#>                               last_corp_announcement open_price high_price
#> 1                         Dividend - Rs 10 Per Share     741.00     782.50
#> 2                  Interim Dividend - Rs 5 Per Share    9926.40   10444.95
#> 3                      Dividend - Rs 6.50  Per Share    2121.05    2219.90
#> 4                 Interim Dividend - Rs 10 Per Share    5364.15    5573.35
#> 5                             Dividend - Rs 6 Per Sh     595.80     616.85
#> 6   Annual General Meeting/Dividend - Re 1 Per Share     350.70     361.30
#> 7  Annual General Meeting/Dividend - Rs 21 Per Share    2608.90    2664.15
#> 8   Annual General Meeting/Dividend - Re 1 Per Share     618.00     633.75
#> 9                          Dividend - Rs 2 Per Share     415.30     435.00
#> 10                 Interim Dividend - Rs 2 Per Share     888.00     910.00
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1     741.00            774.80           735.50           5.34        28842835
#> 2    9890.40          10369.30          9877.00           4.98          504122
#> 3    2107.20           2207.10          2106.00           4.80        14733134
#> 4    5340.10           5544.00          5316.30           4.28         2252413
#> 5     589.20            613.95           589.45           4.16         8740632
#> 6     350.50            358.75           345.25           3.91        20104887
#> 7    2592.15           2654.85          2564.00           3.54         5640063
#> 8     613.85            632.00           610.50           3.52        21036174
#> 9     415.30            426.80           413.30           3.27        11903775
#> 10    883.50            906.95           879.15           3.16         2127137
#>    turnover_in_lakhs
#> 1          222929.16
#> 2           51502.72
#> 3          318461.11
#> 4          122799.75
#> 5           52585.39
#> 6           71782.49
#> 7          148149.23
#> 8          131461.36
#> 9           51157.66
#> 10          19123.17

# retain original column names as returned by NSE
nse_stock_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1   TATASTEEL     EQ               2020-08-06
#> 2  BAJAJFINSV     EQ               2020-03-03
#> 3    RELIANCE     EQ               2020-07-02
#> 4  BAJFINANCE     EQ               2020-03-03
#> 5         UPL     EQ               2020-08-19
#> 6    HINDALCO     EQ               2020-09-03
#> 7        HDFC     EQ               2020-07-09
#> 8   ICICIBANK     EQ               2019-07-22
#> 9    JSWSTEEL     EQ               2020-07-06
#> 10    SBILIFE     EQ               2019-04-03
#>                                 lastCorpAnnouncement openPrice highPrice
#> 1                         Dividend - Rs 10 Per Share    741.00    782.50
#> 2                  Interim Dividend - Rs 5 Per Share   9926.40  10444.95
#> 3                      Dividend - Rs 6.50  Per Share   2121.05   2219.90
#> 4                 Interim Dividend - Rs 10 Per Share   5364.15   5573.35
#> 5                             Dividend - Rs 6 Per Sh    595.80    616.85
#> 6   Annual General Meeting/Dividend - Re 1 Per Share    350.70    361.30
#> 7  Annual General Meeting/Dividend - Rs 21 Per Share   2608.90   2664.15
#> 8   Annual General Meeting/Dividend - Re 1 Per Share    618.00    633.75
#> 9                          Dividend - Rs 2 Per Share    415.30    435.00
#> 10                 Interim Dividend - Rs 2 Per Share    888.00    910.00
#>    lowPrice      ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1    741.00   774.80        735.50     5.34       28842835       222929.16
#> 2   9890.40 10369.30       9877.00     4.98         504122        51502.72
#> 3   2107.20  2207.10       2106.00     4.80       14733134       318461.11
#> 4   5340.10  5544.00       5316.30     4.28        2252413       122799.75
#> 5    589.20   613.95        589.45     4.16        8740632        52585.39
#> 6    350.50   358.75        345.25     3.91       20104887        71782.49
#> 7   2592.15  2654.85       2564.00     3.54        5640063       148149.23
#> 8    613.85   632.00        610.50     3.52       21036174       131461.36
#> 9    415.30   426.80        413.30     3.27       11903775        51157.66
#> 10   883.50   906.95        879.15     3.16        2127137        19123.17

# top losers
nse_stock_top_losers()

#>       symbol series last_corp_announcement_date
#> 1 HEROMOTOCO     EQ                  2021-02-15
#> 2     MARUTI     EQ                  2020-08-13
#> 3 BAJAJ-AUTO     EQ                  2020-03-03
#> 4        M&M     EQ                  2020-07-16
#> 5       BPCL     EQ                  2021-02-17
#> 6      TITAN     EQ                  2020-08-03
#> 7  EICHERMOT     EQ                  2020-08-24
#>                                                         last_corp_announcement
#> 1         Interim Dividend - Rs 65 Per Share Special Dividend - Rs 5 Per Share
#> 2                                                   Dividend - Rs 60 Per Share
#> 3                                          Interim Dividend - Rs 120 Per Share
#> 4                          Annual General Meeting/Dividend - Rs 2.35 Per Share
#> 5                                           Interim Dividend - Rs 16 Per Share
#> 6                             Annual General Meeting/Dividend - Rs 4 Per Share
#> 7 Face Value Split (Sub-Division) - From Rs 10/- Per Share To Re 1/- Per Share
#>   open_price high_price low_price last_traded_price prev_close_price
#> 1    3510.00    3510.90   3394.00           3422.90          3476.40
#> 2    7245.60    7262.30   7101.20           7129.25          7214.10
#> 3    3985.00    3985.00   3863.60           3897.95          3944.05
#> 4     866.95     868.00    847.05            851.45           860.05
#> 5     480.00     480.00    462.30            466.35           469.00
#> 6    1472.00    1481.65   1466.15           1470.05          1471.15
#> 7    2605.80    2626.85   2571.10           2601.00          2602.30
#>   percent_change traded_quantity turnover_in_lakhs
#> 1          -1.54          965062          33036.58
#> 2          -1.18          796258          56794.54
#> 3          -1.17          615104          23956.82
#> 4          -1.00         5028771          42969.34
#> 5          -0.57         8837642          41355.75
#> 6          -0.07         1227772          18109.88
#> 7          -0.05         1289447          33412.15

# retain original column names as returned by NSE
nse_stock_top_losers(clean_names = FALSE)

#>       symbol series lastCorpAnnouncementDate
#> 1 HEROMOTOCO     EQ               2021-02-15
#> 2     MARUTI     EQ               2020-08-13
#> 3 BAJAJ-AUTO     EQ               2020-03-03
#> 4        M&M     EQ               2020-07-16
#> 5       BPCL     EQ               2021-02-17
#> 6      TITAN     EQ               2020-08-03
#> 7  EICHERMOT     EQ               2020-08-24
#>                                                           lastCorpAnnouncement
#> 1         Interim Dividend - Rs 65 Per Share Special Dividend - Rs 5 Per Share
#> 2                                                   Dividend - Rs 60 Per Share
#> 3                                          Interim Dividend - Rs 120 Per Share
#> 4                          Annual General Meeting/Dividend - Rs 2.35 Per Share
#> 5                                           Interim Dividend - Rs 16 Per Share
#> 6                             Annual General Meeting/Dividend - Rs 4 Per Share
#> 7 Face Value Split (Sub-Division) - From Rs 10/- Per Share To Re 1/- Per Share
#>   openPrice highPrice lowPrice     ltp previousPrice netPrice tradedQuantity
#> 1   3510.00   3510.90  3394.00 3422.90       3476.40    -1.54         965062
#> 2   7245.60   7262.30  7101.20 7129.25       7214.10    -1.18         796258
#> 3   3985.00   3985.00  3863.60 3897.95       3944.05    -1.17         615104
#> 4    866.95    868.00   847.05  851.45        860.05    -1.00        5028771
#> 5    480.00    480.00   462.30  466.35        469.00    -0.57        8837642
#> 6   1472.00   1481.65  1466.15 1470.05       1471.15    -0.07        1227772
#> 7   2605.80   2626.85  2571.10 2601.00       2602.30    -0.05        1289447
#>   turnoverInLakhs
#> 1        33036.58
#> 2        56794.54
#> 3        23956.82
#> 4        42969.34
#> 5        41355.75
#> 6        18109.88
#> 7        33412.15

# }