NSE top gainers & losers

Fetch top gainers and losers for the last trading session.

nse_stock_top_gainers(clean_names = TRUE)

nse_stock_top_losers(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ).

last_corp_announcement_date

Last corporate announcement date.

last_corp_announcement

Last corporate announcement.

open_price

Open price.

high_price

High price.

low_price

Low price.

last_traded_price

Last traded price.

prev_close_price

Previous close price.

percent_change

Percentage change in price.

traded_quantity

Total traded quantity.

turnover

Turnover in lakhs.

Examples

# \donttest{
# top gainers
nse_stock_top_gainers()

#>        symbol series last_corp_announcement_date
#> 1  ULTRACEMCO     EQ                  2020-07-29
#> 2    SHREECEM     EQ                  2020-02-24
#> 3  ADANIPORTS     EQ                  2020-06-17
#> 4      GRASIM     EQ                  2020-08-26
#> 5     DRREDDY     EQ                  2020-07-13
#> 6     SBILIFE     EQ                  2019-04-03
#> 7         UPL     EQ                  2020-08-19
#> 8       WIPRO     EQ                  2021-01-22
#> 9   BRITANNIA     EQ                  2020-08-26
#> 10 ASIANPAINT     EQ                  2020-10-28
#>                               last_corp_announcement open_price high_price
#> 1  Annual General Meeting/Dividend - Rs 13 Per Share    6429.00    6830.00
#> 2                Interim Dividend - Rs 110 Per Share   27450.00   28850.00
#> 3                             Annual General Meeting     733.25     765.70
#> 4                          Dividend - Rs 4 Per Share    1306.00    1360.00
#> 5  Annual General Meeting/Dividend - Rs 25 Per Share    4482.00    4582.30
#> 6                  Interim Dividend - Rs 2 Per Share     899.70     924.80
#> 7                             Dividend - Rs 6 Per Sh     602.00     631.45
#> 8                  Interim Dividend - Re 1 Per Share     428.90     444.30
#> 9                 Interim Dividend - Rs 83 Per Share    3450.00    3514.80
#> 10              Interim Dividend - Rs 3.35 Per Share    2380.00    2435.00
#>    low_price last_traded_price prev_close_price percent_change traded_quantity
#> 1    6394.45           6793.40          6500.50           4.51         1117090
#> 2   27335.05          28662.55         27695.25           3.49           65962
#> 3     715.85            754.15           729.85           3.33        28006261
#> 4    1300.50           1356.00          1318.10           2.88         1630901
#> 5    4444.00           4580.60          4502.55           1.73         1267131
#> 6     892.40            921.65           906.45           1.68         2860767
#> 7     599.10            620.75           613.60           1.17        12818231
#> 8     427.50            439.70           435.50           0.96        11100023
#> 9    3450.00           3509.70          3481.15           0.82          458290
#> 10   2350.00           2420.10          2402.10           0.75         1799718
#>    turnover_in_lakhs
#> 1           74167.96
#> 2           18588.86
#> 3          209828.51
#> 4           21748.88
#> 5           57362.01
#> 6           26230.66
#> 7           79371.77
#> 8           48736.87
#> 9           16011.83
#> 10          43371.94

# retain original column names as returned by NSE
nse_stock_top_gainers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1  ULTRACEMCO     EQ               2020-07-29
#> 2    SHREECEM     EQ               2020-02-24
#> 3  ADANIPORTS     EQ               2020-06-17
#> 4      GRASIM     EQ               2020-08-26
#> 5     DRREDDY     EQ               2020-07-13
#> 6     SBILIFE     EQ               2019-04-03
#> 7         UPL     EQ               2020-08-19
#> 8       WIPRO     EQ               2021-01-22
#> 9   BRITANNIA     EQ               2020-08-26
#> 10 ASIANPAINT     EQ               2020-10-28
#>                                 lastCorpAnnouncement openPrice highPrice
#> 1  Annual General Meeting/Dividend - Rs 13 Per Share   6429.00   6830.00
#> 2                Interim Dividend - Rs 110 Per Share  27450.00  28850.00
#> 3                             Annual General Meeting    733.25    765.70
#> 4                          Dividend - Rs 4 Per Share   1306.00   1360.00
#> 5  Annual General Meeting/Dividend - Rs 25 Per Share   4482.00   4582.30
#> 6                  Interim Dividend - Rs 2 Per Share    899.70    924.80
#> 7                             Dividend - Rs 6 Per Sh    602.00    631.45
#> 8                  Interim Dividend - Re 1 Per Share    428.90    444.30
#> 9                 Interim Dividend - Rs 83 Per Share   3450.00   3514.80
#> 10              Interim Dividend - Rs 3.35 Per Share   2380.00   2435.00
#>    lowPrice      ltp previousPrice netPrice tradedQuantity turnoverInLakhs
#> 1   6394.45  6793.40       6500.50     4.51        1117090        74167.96
#> 2  27335.05 28662.55      27695.25     3.49          65962        18588.86
#> 3    715.85   754.15        729.85     3.33       28006261       209828.51
#> 4   1300.50  1356.00       1318.10     2.88        1630901        21748.88
#> 5   4444.00  4580.60       4502.55     1.73        1267131        57362.01
#> 6    892.40   921.65        906.45     1.68        2860767        26230.66
#> 7    599.10   620.75        613.60     1.17       12818231        79371.77
#> 8    427.50   439.70        435.50     0.96       11100023        48736.87
#> 9   3450.00  3509.70       3481.15     0.82         458290        16011.83
#> 10  2350.00  2420.10       2402.10     0.75        1799718        43371.94

# top losers
nse_stock_top_losers()

#>        symbol series last_corp_announcement_date
#> 1    JSWSTEEL     EQ                  2020-07-06
#> 2  BAJAJFINSV     EQ                  2020-03-03
#> 3    HDFCBANK     EQ                  2019-09-19
#> 4  TATAMOTORS     EQ                  2016-07-18
#> 5        HDFC     EQ                  2020-07-09
#> 6          LT     EQ                  2020-11-04
#> 7        SBIN     EQ                  2018-06-15
#> 8    AXISBANK     EQ                  2019-07-04
#> 9   TATASTEEL     EQ                  2020-08-06
#> 10   HINDALCO     EQ                  2020-09-03
#>                                                     last_corp_announcement
#> 1                                                Dividend - Rs 2 Per Share
#> 2                                        Interim Dividend - Rs 5 Per Share
#> 3  Face Value Split (Sub-Division) - From Rs 2 Per Share To Rs 1 Per Share
#> 4                                           Dividend - Re 0.20/- Per Share
#> 5                        Annual General Meeting/Dividend - Rs 21 Per Share
#> 6                                       Special Dividend - Rs 18 Per Share
#> 7           Annual General Meeting/ Change In Registrar And Transfer Agent
#> 8                        Annual General Meeting/ Dividend - Re 1 Per Share
#> 9                                               Dividend - Rs 10 Per Share
#> 10                        Annual General Meeting/Dividend - Re 1 Per Share
#>    open_price high_price low_price last_traded_price prev_close_price
#> 1      415.10     421.40    411.05            415.05           428.15
#> 2    10150.00   10280.00  10077.75          10165.00         10395.60
#> 3     1548.55    1571.00   1539.10           1552.25          1586.85
#> 4      339.00     345.40    337.30            340.95           348.50
#> 5     2610.00    2610.00   2568.00           2598.45          2654.90
#> 6     1473.00    1494.35   1458.00           1466.00          1496.90
#> 7      398.70     403.50    396.60            396.80           405.05
#> 8      731.10     751.45    731.10            739.35           753.95
#> 9      751.00     766.90    743.65            763.30           777.15
#> 10     352.00     357.70    347.00            353.25           359.35
#>    percent_change traded_quantity turnover_in_lakhs
#> 1           -3.06         7723273          32114.91
#> 2           -2.22          359675          36610.99
#> 3           -2.18         7047272         109401.15
#> 4           -2.17        51921657         177499.38
#> 5           -2.13         4253858         109959.68
#> 6           -2.06         2931033          43036.65
#> 7           -2.04        35878943         143627.00
#> 8           -1.94        14538342         107679.68
#> 9           -1.78        20172404         152737.37
#> 10          -1.70        11173567          39415.87

# retain original column names as returned by NSE
nse_stock_top_losers(clean_names = FALSE)

#>        symbol series lastCorpAnnouncementDate
#> 1    JSWSTEEL     EQ               2020-07-06
#> 2  BAJAJFINSV     EQ               2020-03-03
#> 3    HDFCBANK     EQ               2019-09-19
#> 4  TATAMOTORS     EQ               2016-07-18
#> 5        HDFC     EQ               2020-07-09
#> 6          LT     EQ               2020-11-04
#> 7        SBIN     EQ               2018-06-15
#> 8    AXISBANK     EQ               2019-07-04
#> 9   TATASTEEL     EQ               2020-08-06
#> 10   HINDALCO     EQ               2020-09-03
#>                                                       lastCorpAnnouncement
#> 1                                                Dividend - Rs 2 Per Share
#> 2                                        Interim Dividend - Rs 5 Per Share
#> 3  Face Value Split (Sub-Division) - From Rs 2 Per Share To Rs 1 Per Share
#> 4                                           Dividend - Re 0.20/- Per Share
#> 5                        Annual General Meeting/Dividend - Rs 21 Per Share
#> 6                                       Special Dividend - Rs 18 Per Share
#> 7           Annual General Meeting/ Change In Registrar And Transfer Agent
#> 8                        Annual General Meeting/ Dividend - Re 1 Per Share
#> 9                                               Dividend - Rs 10 Per Share
#> 10                        Annual General Meeting/Dividend - Re 1 Per Share
#>    openPrice highPrice lowPrice      ltp previousPrice netPrice tradedQuantity
#> 1     415.10    421.40   411.05   415.05        428.15    -3.06        7723273
#> 2   10150.00  10280.00 10077.75 10165.00      10395.60    -2.22         359675
#> 3    1548.55   1571.00  1539.10  1552.25       1586.85    -2.18        7047272
#> 4     339.00    345.40   337.30   340.95        348.50    -2.17       51921657
#> 5    2610.00   2610.00  2568.00  2598.45       2654.90    -2.13        4253858
#> 6    1473.00   1494.35  1458.00  1466.00       1496.90    -2.06        2931033
#> 7     398.70    403.50   396.60   396.80        405.05    -2.04       35878943
#> 8     731.10    751.45   731.10   739.35        753.95    -1.94       14538342
#> 9     751.00    766.90   743.65   763.30        777.15    -1.78       20172404
#> 10    352.00    357.70   347.00   353.25        359.35    -1.70       11173567
#>    turnoverInLakhs
#> 1         32114.91
#> 2         36610.99
#> 3        109401.15
#> 4        177499.38
#> 5        109959.68
#> 6         43036.65
#> 7        143627.00
#> 8        107679.68
#> 9        152737.37
#> 10        39415.87

# }