Pre Open market data

Fetch data of pre open session of Nifty & Nifty Bank.

nse_preopen_nifty(clean_names = TRUE)

nse_preopen_nifty_bank(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ)

corp_action_date

Corporate action date.

corp_action

Corporate Action

price

Price

change

Change in price

percent_change

Percentage change in price.

prev_close

Previous close.

quantity

Quantity

value

Value (in lakhs),

mkt_cap

Free float market capitalization (in crores).

year_high

Normal market 52 week high.

year_low

Normal market 52 week low.

Examples

# \donttest{
# nifty
nse_preopen_nifty()

#> # A tibble: 50 x 17
#>    symbol series corp_action_date corp_action price change percent_change
#>    <chr>  <chr>  <date>           <chr>       <dbl>  <dbl>          <dbl>
#>  1 MARUTI EQ     NA               NA          7007  120.            1.75 
#>  2 INFRA~ EQ     NA               NA           258   -1.7          -0.65 
#>  3 INDUS~ EQ     NA               NA          1457   -8.85         -0.6  
#>  4 AXISB~ EQ     NA               NA           713.  -4.15         -0.580
#>  5 JSWST~ EQ     NA               NA           253   -1.1          -0.43 
#>  6 SBIN   EQ     NA               NA           319.  -1.05         -0.33 
#>  7 BPCL   EQ     NA               NA           487.  -1.05         -0.21 
#>  8 ITC    EQ     NA               NA           243   -0.35         -0.14 
#>  9 ICICI~ EQ     NA               NA           524.  -0.75         -0.14 
#> 10 WIPRO  EQ     NA               NA           241.  -0.2          -0.08 
#> # ... with 40 more rows, and 10 more variables: prev_close <dbl>,
#> #   quantity <dbl>, value <dbl>, mkt_cap <dbl>, year_high <dbl>,
#> #   year_low <dbl>, sum_val <dbl>, sum_quantity <dbl>, fin_quantity <dbl>,
#> #   sum_fin_quantity <dbl>

# retain original column names as returned by NSE
nse_preopen_nifty(clean_names = FALSE)

#> # A tibble: 50 x 17
#>    symbol series xDt        caAct   iep    chn perChn  pCls trdQnty  iVal mktCap
#>    <chr>  <chr>  <date>     <chr> <dbl>  <dbl>  <dbl> <dbl>   <dbl> <dbl>  <dbl>
#>  1 MARUTI EQ     NA         NA    7007  120.    1.75  6887.    5594 392.  9.15e4
#>  2 INFRA~ EQ     NA         NA     258   -1.7  -0.65   260.   24591  63.4 2.21e4
#>  3 INDUS~ EQ     NA         NA    1457   -8.85 -0.6   1466.    3829  55.8 8.84e4
#>  4 AXISB~ EQ     NA         NA     713.  -4.15 -0.580  718.   22777 162.  1.60e5
#>  5 JSWST~ EQ     NA         NA     253   -1.1  -0.43   254.    5663  14.3 2.58e4
#>  6 SBIN   EQ     NA         NA     319.  -1.05 -0.33   320   112019 357.  1.23e5
#>  7 BPCL   EQ     NA         NA     487.  -1.05 -0.21   488.   13495  65.8 3.92e4
#>  8 ITC    EQ     NA         NA     243   -0.35 -0.14   243.   20211  49.1 2.09e5
#>  9 ICICI~ EQ     NA         NA     524.  -0.75 -0.14   525.   32809 172.  3.39e5
#> 10 WIPRO  EQ     NA         NA     241.  -0.2  -0.08   241     7231  17.4 3.58e4
#> # ... with 40 more rows, and 6 more variables: yHigh <dbl>, yLow <dbl>,
#> #   sumVal <dbl>, sumQnty <dbl>, finQnty <dbl>, sumfinQnty <dbl>

# nifty bank
nse_preopen_nifty_bank()

#> # A tibble: 12 x 17
#>    symbol series corp_action_date corp_action  price change percent_change
#>    <chr>  <chr>  <date>           <chr>        <dbl>  <dbl>          <dbl>
#>  1 INDUS~ EQ     NA               NA          1457    -8.85         -0.6  
#>  2 AXISB~ EQ     NA               NA           713.   -4.15         -0.580
#>  3 SBIN   EQ     NA               NA           319.   -1.05         -0.33 
#>  4 ICICI~ EQ     NA               NA           524.   -0.75         -0.14 
#>  5 IDFCF~ EQ     NA               NA            43.2  -0.05         -0.12 
#>  6 FEDER~ EQ     NA               NA            84.0  -0.1          -0.12 
#>  7 HDFCB~ EQ     2019-09-19       FACE VALUE~ 1250.    3.45          0.28 
#>  8 KOTAK~ EQ     NA               NA          1678.    3.65          0.22 
#>  9 RBLBA~ EQ     NA               NA           339.    0.6           0.18 
#> 10 PNB    EQ     NA               NA            60.3   0.1           0.17 
#> 11 BANKB~ EQ     NA               NA            97.0   0             0    
#> 12 YESBA~ EQ     NA               NA            56     0             0    
#> # ... with 10 more variables: prev_close <dbl>, quantity <dbl>, value <dbl>,
#> #   mkt_cap <dbl>, year_high <dbl>, year_low <dbl>, sum_val <dbl>,
#> #   sum_quantity <dbl>, fin_quantity <dbl>, sum_fin_quantity <dbl>

# retain original column names as returned by NSE
nse_preopen_nifty_bank(clean_names = FALSE)

#> # A tibble: 12 x 17
#>    symbol series xDt        caAct    iep   chn perChn   pCls trdQnty   iVal
#>    <chr>  <chr>  <date>     <chr>  <dbl> <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1 INDUS~ EQ     NA         NA    1457   -8.85 -0.6   1466.     3829  55.8 
#>  2 AXISB~ EQ     NA         NA     713.  -4.15 -0.580  718.    22777 162.  
#>  3 SBIN   EQ     NA         NA     319.  -1.05 -0.33   320    112019 357.  
#>  4 ICICI~ EQ     NA         NA     524.  -0.75 -0.14   525.    32809 172.  
#>  5 IDFCF~ EQ     NA         NA      43.2 -0.05 -0.12    43.3   17400   7.53
#>  6 FEDER~ EQ     NA         NA      84.0 -0.1  -0.12    84.2   18359  15.4 
#>  7 HDFCB~ EQ     2019-09-19 FACE~ 1250.   3.45  0.28  1246.     8763 109.  
#>  8 KOTAK~ EQ     NA         NA    1678.   3.65  0.22  1674.     3550  59.6 
#>  9 RBLBA~ EQ     NA         NA     339.   0.6   0.18   339.    13298  45.1 
#> 10 PNB    EQ     NA         NA      60.3  0.1   0.17    60.2   16552   9.98
#> 11 BANKB~ EQ     NA         NA      97.0  0     0       97.0   43390  42.1 
#> 12 YESBA~ EQ     NA         NA      56    0     0       56    449621 252.  
#> # ... with 7 more variables: mktCap <dbl>, yHigh <dbl>, yLow <dbl>,
#> #   sumVal <dbl>, sumQnty <dbl>, finQnty <dbl>, sumfinQnty <dbl>

# }