Pre Open market data

Fetch data of pre open session of Nifty & Nifty Bank.

nse_preopen_nifty(clean_names = TRUE)

nse_preopen_nifty_bank(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ)

corp_action_date

Corporate action date.

corp_action

Corporate Action

price

Price

change

Change in price

percent_change

Percentage change in price.

prev_close

Previous close.

quantity

Quantity

value

Value (in lakhs),

mkt_cap

Free float market capitalization (in crores).

year_high

Normal market 52 week high.

year_low

Normal market 52 week low.

Examples

# \donttest{
# nifty
nse_preopen_nifty()

#>        symbol series corp_action_date
#> 1    JSWSTEEL     EQ             <NA>
#> 2        ONGC     EQ       2020-03-23
#> 3    INFRATEL     EQ       2020-05-05
#> 4          LT     EQ       2020-03-24
#> 5    AXISBANK     EQ             <NA>
#> 6  BAJFINANCE     EQ             <NA>
#> 7     HCLTECH     EQ             <NA>
#> 8   ICICIBANK     EQ             <NA>
#> 9      MARUTI     EQ             <NA>
#> 10   RELIANCE     EQ       2020-05-13
#> 11      TITAN     EQ             <NA>
#> 12      WIPRO     EQ             <NA>
#> 13 TATAMOTORS     EQ             <NA>
#> 14 HINDUNILVR     EQ             <NA>
#> 15  COALINDIA     EQ             <NA>
#> 16 INDUSINDBK     EQ             <NA>
#> 17       SBIN     EQ             <NA>
#> 18  KOTAKBANK     EQ             <NA>
#> 19 ADANIPORTS     EQ       2020-06-17
#> 20       HDFC     EQ             <NA>
#> 21 ULTRACEMCO     EQ             <NA>
#> 22     GRASIM     EQ             <NA>
#> 23        TCS     EQ       2020-06-03
#> 24        M&M     EQ             <NA>
#> 25       BPCL     EQ       2020-03-23
#> 26 HEROMOTOCO     EQ             <NA>
#> 27      TECHM     EQ             <NA>
#> 28        IOC     EQ       2020-03-23
#> 29  EICHERMOT     EQ             <NA>
#> 30       ZEEL     EQ             <NA>
#> 31  POWERGRID     EQ             <NA>
#> 32       VEDL     EQ             <NA>
#> 33       INFY     EQ       2020-05-29
#> 34    DRREDDY     EQ             <NA>
#> 35        UPL     EQ             <NA>
#> 36  BRITANNIA     EQ             <NA>
#> 37   SHREECEM     EQ             <NA>
#> 38       NTPC     EQ       2020-03-26
#> 39   HINDALCO     EQ             <NA>
#> 40       GAIL     EQ             <NA>
#> 41      CIPLA     EQ             <NA>
#> 42 BHARTIARTL     EQ             <NA>
#> 43  NESTLEIND     EQ       2020-06-11
#> 44   HDFCBANK     EQ             <NA>
#> 45  TATASTEEL     EQ             <NA>
#> 46 BAJAJFINSV     EQ             <NA>
#> 47        ITC     EQ             <NA>
#> 48  SUNPHARMA     EQ             <NA>
#> 49 ASIANPAINT     EQ             <NA>
#> 50 BAJAJ-AUTO     EQ             <NA>
#>                                          corp_action    price change
#> 1                                               <NA>   189.00  -2.85
#> 2                  INTERIM DIVIDEND - RS 5 PER SHARE    83.00  -1.20
#> 3               INTERIM DIVIDEND - RS 4.10 PER SHARE   208.95   2.70
#> 4                 INTERIM DIVIDEND - RS 10 PER SHARE   876.00  -8.80
#> 5                                               <NA>   386.00  -3.60
#> 6                                               <NA>  2375.00 -21.00
#> 7                                               <NA>   571.20  -4.65
#> 8                                               <NA>   340.00  -1.95
#> 9                                               <NA>  5695.00 -31.20
#> 10                     RIGHTS 1:15 @ PREMIUM RS 1247  1607.00  -8.35
#> 11                                              <NA>   958.10  -4.90
#> 12                                              <NA>   216.80  -1.00
#> 13                                              <NA>    95.00  -0.40
#> 14                                              <NA>  2070.25  -8.00
#> 15                                              <NA>   129.00  -0.45
#> 16                                              <NA>   489.50  -1.65
#> 17                                              <NA>   172.40  -0.55
#> 18                                              <NA>  1240.00  -3.80
#> 19                            ANNUAL GENERAL MEETING   340.35  -0.95
#> 20                                              <NA>  1790.00  -4.65
#> 21                                              <NA>  3799.70  -9.50
#> 22                                              <NA>   583.00  -0.85
#> 23                         DIVIDEND - RS 6 PER SHARE  2045.00  -2.80
#> 24                                              <NA>   496.90  -0.70
#> 25             INTERIM DIVIDEND - RS 16.50 PER SHARE   364.10  -0.50
#> 26                                              <NA>  2360.25  -3.05
#> 27                                              <NA>   535.05  -0.50
#> 28              INTERIM DIVIDEND - RS 4.25 PER SHARE    84.30  -0.05
#> 29                                              <NA> 16788.50  -8.35
#> 30                                              <NA>   166.50   1.50
#> 31                                              <NA>   160.70   1.20
#> 32                                              <NA>   102.80   0.70
#> 33                      DIVIDEND - RS 9.50 PER SHARE   708.20   4.55
#> 34                                              <NA>  4010.00  24.80
#> 35                                              <NA>   422.85   2.05
#> 36                                              <NA>  3415.00  13.55
#> 37                                              <NA> 21718.00  82.60
#> 38              INTERIM DIVIDEND - RS 0.50 PER HSARE    91.80   0.30
#> 39                                              <NA>   148.45   0.45
#> 40                                              <NA>    96.50   0.25
#> 41                                              <NA>   641.65   1.30
#> 42                                              <NA>   564.25   0.80
#> 43 ANNUAL GENERAL MEETING/DIVIDEND - RS 61 PER SHARE 16218.50  20.35
#> 44                                              <NA>   980.00   0.75
#> 45                                              <NA>   308.20   0.00
#> 46                                              <NA>  5036.95   0.00
#> 47                                              <NA>   181.20   0.00
#> 48                                              <NA>   485.65   0.00
#> 49                                              <NA>  1595.40   0.00
#> 50                                              <NA>  2689.45   0.00
#>    percent_change prev_close quantity  value   mkt_cap year_high year_low
#> 1           -1.49     191.85    10577  19.99  19477.24    296.75   132.50
#> 2           -1.43      84.20    32140  26.68  30718.53    173.45    50.00
#> 3            1.31     206.25    28951  60.49  17548.16    296.50   120.05
#> 4           -0.99     884.80    16578 145.22 109298.80   1591.65   661.00
#> 5           -0.92     389.60    38462 148.46  87921.92    819.00   286.00
#> 6           -0.88    2396.00     6468 153.62  63432.47   4923.40  1783.00
#> 7           -0.81     575.85     2417  13.81  62506.56    623.50   375.25
#> 8           -0.57     341.95    36447 123.92 221323.07    552.20   268.30
#> 9           -0.54    5726.20     1902 108.32  76109.92   7758.70  4001.10
#> 10          -0.52    1615.35    28200 453.17 522246.83   1648.55   866.98
#> 11          -0.51     963.00     1245  11.93  40182.09   1389.95   720.90
#> 12          -0.46     217.80     5193  11.26  32353.00    300.50   159.40
#> 13          -0.42      95.40   111472 105.90  17091.91    201.70    63.50
#> 14          -0.38    2078.25     4694  97.18 148468.30   2614.30  1659.05
#> 15          -0.35     129.45    39277  50.67  27124.02    265.35   119.20
#> 16          -0.34     491.15    24913 121.95  29634.63   1596.55   235.55
#> 17          -0.32     172.95    77741 134.03  66371.00    373.80   149.45
#> 18          -0.31    1243.80     6537  81.06 166510.81   1740.00  1001.00
#> 19          -0.28     341.30     2285   7.78  26350.60    428.85   203.00
#> 20          -0.26    1794.65     8593 153.81 310288.49   2499.90  1473.45
#> 21          -0.25    3809.20      215   8.17  43976.98   4754.10  2910.00
#> 22          -0.15     583.85     1561   9.10  23043.35    949.50   385.05
#> 23          -0.14    2047.80     1647  33.68 215155.74   2296.20  1506.05
#> 24          -0.14     497.60     4533  22.52  47633.17    675.30   245.40
#> 25          -0.14     364.60     3226  11.75  29263.65    549.00   252.00
#> 26          -0.13    2363.30     1617  38.17  30682.86   3022.90  1475.00
#> 27          -0.09     535.55     2960  15.84  33101.32    845.90   471.40
#> 28          -0.06      84.35    18850  15.89  21440.28    158.00    71.15
#> 29          -0.05   16796.85      360  60.44  23390.17  23450.00 12450.10
#> 30           0.91     165.00    28857  48.05  15055.57    405.20   114.00
#> 31           0.75     159.50    13314  21.40  40887.49    216.25   122.15
#> 32           0.69     102.10    14546  14.95  18596.76    180.00    60.20
#> 33           0.65     703.65    22552 159.71 257721.77    847.00   509.25
#> 34           0.62    3985.20     1530  61.35  48340.12   4190.00  2351.20
#> 35           0.49     420.80    14166  59.90  23148.74    709.05   240.15
#> 36           0.40    3401.45     1461  49.89  40079.10   3708.00  2100.00
#> 37           0.38   21635.40      113  24.54  28882.99  25355.00 15410.00
#> 38           0.33      91.50    14327  13.15  44362.25    145.85    73.20
#> 39           0.30     148.00    21655  32.15  21607.12    221.20    84.90
#> 40           0.26      96.25    12682  12.24  17798.15    157.38    65.00
#> 41           0.20     640.35     1873  12.02  32525.19    669.75   355.30
#> 42           0.14     563.45    29784 168.06 126031.29    612.00   321.15
#> 43           0.13   16198.15      145  23.52  57784.98  18369.95 11265.00
#> 44           0.08     979.25    44801 439.05 424081.52   1305.50   738.75
#> 45           0.00     308.20    17080  52.64  23261.33    516.45   250.85
#> 46           0.00    5036.95     1113  56.06  30459.53   9950.00  3985.30
#> 47           0.00     181.20    41650  75.47 158142.01    282.90   134.60
#> 48           0.00     485.65     4026  19.55  52435.67    512.70   312.00
#> 49           0.00    1595.40     1744  27.82  71924.30   1916.70  1291.25
#> 50           0.00    2689.45      594  15.98  35020.72   3315.15  1788.65
#>    sum_val sum_quantity fin_quantity sum_fin_quantity
#> 1  3632.29       413159        10577           807074
#> 2  3632.29       413159        32140           807074
#> 3  3632.29       413159        28951           807074
#> 4  3632.29       413159        16578           807074
#> 5  3632.29       413159        38462           807074
#> 6  3632.29       413159         6468           807074
#> 7  3632.29       413159         2417           807074
#> 8  3632.29       413159        36447           807074
#> 9  3632.29       413159         1902           807074
#> 10 3632.29       413159        28200           807074
#> 11 3632.29       413159         1245           807074
#> 12 3632.29       413159         5193           807074
#> 13 3632.29       413159       111472           807074
#> 14 3632.29       413159         4694           807074
#> 15 3632.29       413159        39277           807074
#> 16 3632.29       413159        24913           807074
#> 17 3632.29       413159        77741           807074
#> 18 3632.29       413159         6537           807074
#> 19 3632.29       413159         2285           807074
#> 20 3632.29       413159         8593           807074
#> 21 3632.29       413159          215           807074
#> 22 3632.29       413159         1561           807074
#> 23 3632.29       413159         1647           807074
#> 24 3632.29       413159         4533           807074
#> 25 3632.29       413159         3226           807074
#> 26 3632.29       413159         1617           807074
#> 27 3632.29       413159         2960           807074
#> 28 3632.29       413159        18850           807074
#> 29 3632.29       413159          360           807074
#> 30 3632.29       413159        28857           807074
#> 31 3632.29       413159        13314           807074
#> 32 3632.29       413159        14546           807074
#> 33 3632.29       413159        22552           807074
#> 34 3632.29       413159         1530           807074
#> 35 3632.29       413159        14166           807074
#> 36 3632.29       413159         1461           807074
#> 37 3632.29       413159          113           807074
#> 38 3632.29       413159        14327           807074
#> 39 3632.29       413159        21655           807074
#> 40 3632.29       413159        12682           807074
#> 41 3632.29       413159         1873           807074
#> 42 3632.29       413159        29784           807074
#> 43 3632.29       413159          145           807074
#> 44 3632.29       413159        44801           807074
#> 45 3632.29       413159        17080           807074
#> 46 3632.29       413159         1113           807074
#> 47 3632.29       413159        41650           807074
#> 48 3632.29       413159         4026           807074
#> 49 3632.29       413159         1744           807074
#> 50 3632.29       413159          594           807074

# retain original column names as returned by NSE
nse_preopen_nifty(clean_names = FALSE)

#>        symbol series        xDt
#> 1    JSWSTEEL     EQ       <NA>
#> 2        ONGC     EQ 2020-03-23
#> 3    INFRATEL     EQ 2020-05-05
#> 4          LT     EQ 2020-03-24
#> 5    AXISBANK     EQ       <NA>
#> 6  BAJFINANCE     EQ       <NA>
#> 7     HCLTECH     EQ       <NA>
#> 8   ICICIBANK     EQ       <NA>
#> 9      MARUTI     EQ       <NA>
#> 10   RELIANCE     EQ 2020-05-13
#> 11      TITAN     EQ       <NA>
#> 12      WIPRO     EQ       <NA>
#> 13 TATAMOTORS     EQ       <NA>
#> 14 HINDUNILVR     EQ       <NA>
#> 15  COALINDIA     EQ       <NA>
#> 16 INDUSINDBK     EQ       <NA>
#> 17       SBIN     EQ       <NA>
#> 18  KOTAKBANK     EQ       <NA>
#> 19 ADANIPORTS     EQ 2020-06-17
#> 20       HDFC     EQ       <NA>
#> 21 ULTRACEMCO     EQ       <NA>
#> 22     GRASIM     EQ       <NA>
#> 23        TCS     EQ 2020-06-03
#> 24        M&M     EQ       <NA>
#> 25       BPCL     EQ 2020-03-23
#> 26 HEROMOTOCO     EQ       <NA>
#> 27      TECHM     EQ       <NA>
#> 28        IOC     EQ 2020-03-23
#> 29  EICHERMOT     EQ       <NA>
#> 30       ZEEL     EQ       <NA>
#> 31  POWERGRID     EQ       <NA>
#> 32       VEDL     EQ       <NA>
#> 33       INFY     EQ 2020-05-29
#> 34    DRREDDY     EQ       <NA>
#> 35        UPL     EQ       <NA>
#> 36  BRITANNIA     EQ       <NA>
#> 37   SHREECEM     EQ       <NA>
#> 38       NTPC     EQ 2020-03-26
#> 39   HINDALCO     EQ       <NA>
#> 40       GAIL     EQ       <NA>
#> 41      CIPLA     EQ       <NA>
#> 42 BHARTIARTL     EQ       <NA>
#> 43  NESTLEIND     EQ 2020-06-11
#> 44   HDFCBANK     EQ       <NA>
#> 45  TATASTEEL     EQ       <NA>
#> 46 BAJAJFINSV     EQ       <NA>
#> 47        ITC     EQ       <NA>
#> 48  SUNPHARMA     EQ       <NA>
#> 49 ASIANPAINT     EQ       <NA>
#> 50 BAJAJ-AUTO     EQ       <NA>
#>                                                caAct      iep    chn perChn
#> 1                                               <NA>   189.00  -2.85  -1.49
#> 2                  INTERIM DIVIDEND - RS 5 PER SHARE    83.00  -1.20  -1.43
#> 3               INTERIM DIVIDEND - RS 4.10 PER SHARE   208.95   2.70   1.31
#> 4                 INTERIM DIVIDEND - RS 10 PER SHARE   876.00  -8.80  -0.99
#> 5                                               <NA>   386.00  -3.60  -0.92
#> 6                                               <NA>  2375.00 -21.00  -0.88
#> 7                                               <NA>   571.20  -4.65  -0.81
#> 8                                               <NA>   340.00  -1.95  -0.57
#> 9                                               <NA>  5695.00 -31.20  -0.54
#> 10                     RIGHTS 1:15 @ PREMIUM RS 1247  1607.00  -8.35  -0.52
#> 11                                              <NA>   958.10  -4.90  -0.51
#> 12                                              <NA>   216.80  -1.00  -0.46
#> 13                                              <NA>    95.00  -0.40  -0.42
#> 14                                              <NA>  2070.25  -8.00  -0.38
#> 15                                              <NA>   129.00  -0.45  -0.35
#> 16                                              <NA>   489.50  -1.65  -0.34
#> 17                                              <NA>   172.40  -0.55  -0.32
#> 18                                              <NA>  1240.00  -3.80  -0.31
#> 19                            ANNUAL GENERAL MEETING   340.35  -0.95  -0.28
#> 20                                              <NA>  1790.00  -4.65  -0.26
#> 21                                              <NA>  3799.70  -9.50  -0.25
#> 22                                              <NA>   583.00  -0.85  -0.15
#> 23                         DIVIDEND - RS 6 PER SHARE  2045.00  -2.80  -0.14
#> 24                                              <NA>   496.90  -0.70  -0.14
#> 25             INTERIM DIVIDEND - RS 16.50 PER SHARE   364.10  -0.50  -0.14
#> 26                                              <NA>  2360.25  -3.05  -0.13
#> 27                                              <NA>   535.05  -0.50  -0.09
#> 28              INTERIM DIVIDEND - RS 4.25 PER SHARE    84.30  -0.05  -0.06
#> 29                                              <NA> 16788.50  -8.35  -0.05
#> 30                                              <NA>   166.50   1.50   0.91
#> 31                                              <NA>   160.70   1.20   0.75
#> 32                                              <NA>   102.80   0.70   0.69
#> 33                      DIVIDEND - RS 9.50 PER SHARE   708.20   4.55   0.65
#> 34                                              <NA>  4010.00  24.80   0.62
#> 35                                              <NA>   422.85   2.05   0.49
#> 36                                              <NA>  3415.00  13.55   0.40
#> 37                                              <NA> 21718.00  82.60   0.38
#> 38              INTERIM DIVIDEND - RS 0.50 PER HSARE    91.80   0.30   0.33
#> 39                                              <NA>   148.45   0.45   0.30
#> 40                                              <NA>    96.50   0.25   0.26
#> 41                                              <NA>   641.65   1.30   0.20
#> 42                                              <NA>   564.25   0.80   0.14
#> 43 ANNUAL GENERAL MEETING/DIVIDEND - RS 61 PER SHARE 16218.50  20.35   0.13
#> 44                                              <NA>   980.00   0.75   0.08
#> 45                                              <NA>   308.20   0.00   0.00
#> 46                                              <NA>  5036.95   0.00   0.00
#> 47                                              <NA>   181.20   0.00   0.00
#> 48                                              <NA>   485.65   0.00   0.00
#> 49                                              <NA>  1595.40   0.00   0.00
#> 50                                              <NA>  2689.45   0.00   0.00
#>        pCls trdQnty   iVal    mktCap    yHigh     yLow  sumVal sumQnty finQnty
#> 1    191.85   10577  19.99  19477.24   296.75   132.50 3632.29  413159   10577
#> 2     84.20   32140  26.68  30718.53   173.45    50.00 3632.29  413159   32140
#> 3    206.25   28951  60.49  17548.16   296.50   120.05 3632.29  413159   28951
#> 4    884.80   16578 145.22 109298.80  1591.65   661.00 3632.29  413159   16578
#> 5    389.60   38462 148.46  87921.92   819.00   286.00 3632.29  413159   38462
#> 6   2396.00    6468 153.62  63432.47  4923.40  1783.00 3632.29  413159    6468
#> 7    575.85    2417  13.81  62506.56   623.50   375.25 3632.29  413159    2417
#> 8    341.95   36447 123.92 221323.07   552.20   268.30 3632.29  413159   36447
#> 9   5726.20    1902 108.32  76109.92  7758.70  4001.10 3632.29  413159    1902
#> 10  1615.35   28200 453.17 522246.83  1648.55   866.98 3632.29  413159   28200
#> 11   963.00    1245  11.93  40182.09  1389.95   720.90 3632.29  413159    1245
#> 12   217.80    5193  11.26  32353.00   300.50   159.40 3632.29  413159    5193
#> 13    95.40  111472 105.90  17091.91   201.70    63.50 3632.29  413159  111472
#> 14  2078.25    4694  97.18 148468.30  2614.30  1659.05 3632.29  413159    4694
#> 15   129.45   39277  50.67  27124.02   265.35   119.20 3632.29  413159   39277
#> 16   491.15   24913 121.95  29634.63  1596.55   235.55 3632.29  413159   24913
#> 17   172.95   77741 134.03  66371.00   373.80   149.45 3632.29  413159   77741
#> 18  1243.80    6537  81.06 166510.81  1740.00  1001.00 3632.29  413159    6537
#> 19   341.30    2285   7.78  26350.60   428.85   203.00 3632.29  413159    2285
#> 20  1794.65    8593 153.81 310288.49  2499.90  1473.45 3632.29  413159    8593
#> 21  3809.20     215   8.17  43976.98  4754.10  2910.00 3632.29  413159     215
#> 22   583.85    1561   9.10  23043.35   949.50   385.05 3632.29  413159    1561
#> 23  2047.80    1647  33.68 215155.74  2296.20  1506.05 3632.29  413159    1647
#> 24   497.60    4533  22.52  47633.17   675.30   245.40 3632.29  413159    4533
#> 25   364.60    3226  11.75  29263.65   549.00   252.00 3632.29  413159    3226
#> 26  2363.30    1617  38.17  30682.86  3022.90  1475.00 3632.29  413159    1617
#> 27   535.55    2960  15.84  33101.32   845.90   471.40 3632.29  413159    2960
#> 28    84.35   18850  15.89  21440.28   158.00    71.15 3632.29  413159   18850
#> 29 16796.85     360  60.44  23390.17 23450.00 12450.10 3632.29  413159     360
#> 30   165.00   28857  48.05  15055.57   405.20   114.00 3632.29  413159   28857
#> 31   159.50   13314  21.40  40887.49   216.25   122.15 3632.29  413159   13314
#> 32   102.10   14546  14.95  18596.76   180.00    60.20 3632.29  413159   14546
#> 33   703.65   22552 159.71 257721.77   847.00   509.25 3632.29  413159   22552
#> 34  3985.20    1530  61.35  48340.12  4190.00  2351.20 3632.29  413159    1530
#> 35   420.80   14166  59.90  23148.74   709.05   240.15 3632.29  413159   14166
#> 36  3401.45    1461  49.89  40079.10  3708.00  2100.00 3632.29  413159    1461
#> 37 21635.40     113  24.54  28882.99 25355.00 15410.00 3632.29  413159     113
#> 38    91.50   14327  13.15  44362.25   145.85    73.20 3632.29  413159   14327
#> 39   148.00   21655  32.15  21607.12   221.20    84.90 3632.29  413159   21655
#> 40    96.25   12682  12.24  17798.15   157.38    65.00 3632.29  413159   12682
#> 41   640.35    1873  12.02  32525.19   669.75   355.30 3632.29  413159    1873
#> 42   563.45   29784 168.06 126031.29   612.00   321.15 3632.29  413159   29784
#> 43 16198.15     145  23.52  57784.98 18369.95 11265.00 3632.29  413159     145
#> 44   979.25   44801 439.05 424081.52  1305.50   738.75 3632.29  413159   44801
#> 45   308.20   17080  52.64  23261.33   516.45   250.85 3632.29  413159   17080
#> 46  5036.95    1113  56.06  30459.53  9950.00  3985.30 3632.29  413159    1113
#> 47   181.20   41650  75.47 158142.01   282.90   134.60 3632.29  413159   41650
#> 48   485.65    4026  19.55  52435.67   512.70   312.00 3632.29  413159    4026
#> 49  1595.40    1744  27.82  71924.30  1916.70  1291.25 3632.29  413159    1744
#> 50  2689.45     594  15.98  35020.72  3315.15  1788.65 3632.29  413159     594
#>    sumfinQnty
#> 1      807074
#> 2      807074
#> 3      807074
#> 4      807074
#> 5      807074
#> 6      807074
#> 7      807074
#> 8      807074
#> 9      807074
#> 10     807074
#> 11     807074
#> 12     807074
#> 13     807074
#> 14     807074
#> 15     807074
#> 16     807074
#> 17     807074
#> 18     807074
#> 19     807074
#> 20     807074
#> 21     807074
#> 22     807074
#> 23     807074
#> 24     807074
#> 25     807074
#> 26     807074
#> 27     807074
#> 28     807074
#> 29     807074
#> 30     807074
#> 31     807074
#> 32     807074
#> 33     807074
#> 34     807074
#> 35     807074
#> 36     807074
#> 37     807074
#> 38     807074
#> 39     807074
#> 40     807074
#> 41     807074
#> 42     807074
#> 43     807074
#> 44     807074
#> 45     807074
#> 46     807074
#> 47     807074
#> 48     807074
#> 49     807074
#> 50     807074

# nifty bank
nse_preopen_nifty_bank()

#>        symbol series corp_action_date                          corp_action
#> 1  FEDERALBNK     EQ             <NA>                                 <NA>
#> 2    AXISBANK     EQ             <NA>                                 <NA>
#> 3     RBLBANK     EQ       2020-03-23 INTERIM DIVIDEND - RS 1.50 PER SHARE
#> 4   ICICIBANK     EQ             <NA>                                 <NA>
#> 5  INDUSINDBK     EQ             <NA>                                 <NA>
#> 6        SBIN     EQ             <NA>                                 <NA>
#> 7   KOTAKBANK     EQ             <NA>                                 <NA>
#> 8  BANKBARODA     EQ             <NA>                                 <NA>
#> 9         PNB     EQ             <NA>                                 <NA>
#> 10 BANDHANBNK     EQ             <NA>                                 <NA>
#> 11 IDFCFIRSTB     EQ             <NA>                                 <NA>
#> 12   HDFCBANK     EQ             <NA>                                 <NA>
#>      price change percent_change prev_close quantity  value   mkt_cap year_high
#> 1    46.25  -0.45          -0.96      46.70    34005  15.73   9310.19    110.40
#> 2   386.00  -3.60          -0.92     389.60    38462 148.46  87921.92    819.00
#> 3   161.00  -1.35          -0.83     162.35    25355  40.82   7677.08    663.00
#> 4   340.00  -1.95          -0.57     341.95    36447 123.92 221323.07    552.20
#> 5   489.50  -1.65          -0.34     491.15    24913 121.95  29634.63   1596.55
#> 6   172.40  -0.55          -0.32     172.95    77741 134.03  66371.00    373.80
#> 7  1240.00  -3.80          -0.31    1243.80     6537  81.06 166510.81   1740.00
#> 8    44.80  -0.10          -0.22      44.90    17855   8.00   5394.05    132.85
#> 9    33.10  -0.05          -0.15      33.15    28628   9.48   3796.96     83.40
#> 10  263.00  -0.20          -0.08     263.20     9118  23.98  11019.05    650.00
#> 11   25.55   0.10           0.39      25.45    73524  18.79   5263.72     47.95
#> 12  980.00   0.75           0.08     979.25    44801 439.05 424081.52   1305.50
#>    year_low sum_val sum_quantity fin_quantity sum_fin_quantity
#> 1     35.70 1165.26       205313        34005           417386
#> 2    286.00 1165.26       205313        38462           417386
#> 3    101.55 1165.26       205313        25355           417386
#> 4    268.30 1165.26       205313        36447           417386
#> 5    235.55 1165.26       205313        24913           417386
#> 6    149.45 1165.26       205313        77741           417386
#> 7   1001.00 1165.26       205313         6537           417386
#> 8     36.00 1165.26       205313        17855           417386
#> 9     26.30 1165.26       205313        28628           417386
#> 10   152.20 1165.26       205313         9118           417386
#> 11    17.65 1165.26       205313        73524           417386
#> 12   738.75 1165.26       205313        44801           417386

# retain original column names as returned by NSE
nse_preopen_nifty_bank(clean_names = FALSE)

#>        symbol series        xDt                                caAct     iep
#> 1  FEDERALBNK     EQ       <NA>                                 <NA>   46.25
#> 2    AXISBANK     EQ       <NA>                                 <NA>  386.00
#> 3     RBLBANK     EQ 2020-03-23 INTERIM DIVIDEND - RS 1.50 PER SHARE  161.00
#> 4   ICICIBANK     EQ       <NA>                                 <NA>  340.00
#> 5  INDUSINDBK     EQ       <NA>                                 <NA>  489.50
#> 6        SBIN     EQ       <NA>                                 <NA>  172.40
#> 7   KOTAKBANK     EQ       <NA>                                 <NA> 1240.00
#> 8  BANKBARODA     EQ       <NA>                                 <NA>   44.80
#> 9         PNB     EQ       <NA>                                 <NA>   33.10
#> 10 BANDHANBNK     EQ       <NA>                                 <NA>  263.00
#> 11 IDFCFIRSTB     EQ       <NA>                                 <NA>   25.55
#> 12   HDFCBANK     EQ       <NA>                                 <NA>  980.00
#>      chn perChn    pCls trdQnty   iVal    mktCap   yHigh    yLow  sumVal
#> 1  -0.45  -0.96   46.70   34005  15.73   9310.19  110.40   35.70 1165.26
#> 2  -3.60  -0.92  389.60   38462 148.46  87921.92  819.00  286.00 1165.26
#> 3  -1.35  -0.83  162.35   25355  40.82   7677.08  663.00  101.55 1165.26
#> 4  -1.95  -0.57  341.95   36447 123.92 221323.07  552.20  268.30 1165.26
#> 5  -1.65  -0.34  491.15   24913 121.95  29634.63 1596.55  235.55 1165.26
#> 6  -0.55  -0.32  172.95   77741 134.03  66371.00  373.80  149.45 1165.26
#> 7  -3.80  -0.31 1243.80    6537  81.06 166510.81 1740.00 1001.00 1165.26
#> 8  -0.10  -0.22   44.90   17855   8.00   5394.05  132.85   36.00 1165.26
#> 9  -0.05  -0.15   33.15   28628   9.48   3796.96   83.40   26.30 1165.26
#> 10 -0.20  -0.08  263.20    9118  23.98  11019.05  650.00  152.20 1165.26
#> 11  0.10   0.39   25.45   73524  18.79   5263.72   47.95   17.65 1165.26
#> 12  0.75   0.08  979.25   44801 439.05 424081.52 1305.50  738.75 1165.26
#>    sumQnty finQnty sumfinQnty
#> 1   205313   34005     417386
#> 2   205313   38462     417386
#> 3   205313   25355     417386
#> 4   205313   36447     417386
#> 5   205313   24913     417386
#> 6   205313   77741     417386
#> 7   205313    6537     417386
#> 8   205313   17855     417386
#> 9   205313   28628     417386
#> 10  205313    9118     417386
#> 11  205313   73524     417386
#> 12  205313   44801     417386

# }