Pre Open market data

Fetch data of pre open session of Nifty & Nifty Bank.

nse_preopen_nifty(clean_names = TRUE)

nse_preopen_nifty_bank(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ)

corp_action_date

Corporate action date.

corp_action

Corporate Action

price

Price

change

Change in price

percent_change

Percentage change in price.

prev_close

Previous close.

quantity

Quantity

value

Value (in lakhs),

mkt_cap

Free float market capitalization (in crores).

year_high

Normal market 52 week high.

year_low

Normal market 52 week low.

Examples

# \donttest{
# nifty
nse_preopen_nifty()

#> # A tibble: 50 x 17
#>    symbol series corp_action_date corp_action  price change percent_change
#>    <chr>  <chr>  <date>           <chr>        <dbl>  <dbl>          <dbl>
#>  1 SBIN   EQ     NA               NA           324.    5.45           1.71
#>  2 VEDL   EQ     NA               NA           153.    2.25           1.49
#>  3 BAJFI~ EQ     NA               NA          4055    58.3            1.46
#>  4 HDFCB~ EQ     NA               NA          1259.   18.0            1.45
#>  5 ZEEL   EQ     NA               NA           265     3.5            1.34
#>  6 INDUS~ EQ     NA               NA          1488.   19.0            1.3 
#>  7 YESBA~ EQ     NA               NA            45.7   0.55           1.22
#>  8 RELIA~ EQ     NA               NA          1519    17.5            1.17
#>  9 TITAN  EQ     NA               NA          1171.   12.4            1.07
#> 10 ICICI~ EQ     NA               NA           531.    5.6            1.07
#> # ... with 40 more rows, and 10 more variables: prev_close <dbl>,
#> #   quantity <dbl>, value <dbl>, mkt_cap <dbl>, year_high <dbl>,
#> #   year_low <dbl>, sum_val <dbl>, sum_quantity <dbl>, fin_quantity <dbl>,
#> #   sum_fin_quantity <dbl>

# retain original column names as returned by NSE
nse_preopen_nifty(clean_names = FALSE)

#> # A tibble: 50 x 17
#>    symbol series xDt        caAct    iep   chn perChn   pCls trdQnty   iVal
#>    <chr>  <chr>  <date>     <chr>  <dbl> <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1 SBIN   EQ     NA         NA     324.   5.45   1.71  319    119226  387. 
#>  2 VEDL   EQ     NA         NA     153.   2.25   1.49  151.    29896   45.7
#>  3 BAJFI~ EQ     NA         NA    4055   58.3    1.46 3997.     5803  235. 
#>  4 HDFCB~ EQ     NA         NA    1259.  18.0    1.45 1241.    95521 1203. 
#>  5 ZEEL   EQ     NA         NA     265    3.5    1.34  262.    37585   99.6
#>  6 INDUS~ EQ     NA         NA    1488.  19.0    1.3  1469.     8382  125. 
#>  7 YESBA~ EQ     NA         NA      45.7  0.55   1.22   45.2  186780   85.4
#>  8 RELIA~ EQ     NA         NA    1519   17.5    1.17 1502.    36027  547. 
#>  9 TITAN  EQ     NA         NA    1171.  12.4    1.07 1159.     7974   93.4
#> 10 ICICI~ EQ     NA         NA     531.   5.6    1.07  526.    32942  175. 
#> # ... with 40 more rows, and 7 more variables: mktCap <dbl>, yHigh <dbl>,
#> #   yLow <dbl>, sumVal <dbl>, sumQnty <dbl>, finQnty <dbl>, sumfinQnty <dbl>

# nifty bank
nse_preopen_nifty_bank()

#> # A tibble: 12 x 17
#>    symbol series corp_action_date corp_action  price change percent_change
#>    <chr>  <chr>  <date>           <chr>        <dbl>  <dbl>          <dbl>
#>  1 BANKB~ EQ     NA               NA            97.6   1.8            1.88
#>  2 RBLBA~ EQ     NA               NA           343     6.15           1.83
#>  3 SBIN   EQ     NA               NA           324.    5.45           1.71
#>  4 HDFCB~ EQ     NA               NA          1259.   18.0            1.45
#>  5 INDUS~ EQ     NA               NA          1488.   19.0            1.3 
#>  6 YESBA~ EQ     NA               NA            45.7   0.55           1.22
#>  7 PNB    EQ     NA               NA            62.1   0.7            1.14
#>  8 ICICI~ EQ     NA               NA           531.    5.6            1.07
#>  9 IDFCF~ EQ     NA               NA            44     0.35           0.8 
#> 10 FEDER~ EQ     NA               NA            87.8   0.7            0.8 
#> 11 AXISB~ EQ     NA               NA           728     4.75           0.66
#> 12 KOTAK~ EQ     NA               NA          1661.    8.5            0.51
#> # ... with 10 more variables: prev_close <dbl>, quantity <dbl>, value <dbl>,
#> #   mkt_cap <dbl>, year_high <dbl>, year_low <dbl>, sum_val <dbl>,
#> #   sum_quantity <dbl>, fin_quantity <dbl>, sum_fin_quantity <dbl>

# retain original column names as returned by NSE
nse_preopen_nifty_bank(clean_names = FALSE)

#> # A tibble: 12 x 17
#>    symbol series xDt        caAct    iep   chn perChn   pCls trdQnty   iVal
#>    <chr>  <chr>  <date>     <chr>  <dbl> <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1 BANKB~ EQ     NA         NA      97.6  1.8    1.88   95.8  102872  100. 
#>  2 RBLBA~ EQ     NA         NA     343    6.15   1.83  337.    16805   57.6
#>  3 SBIN   EQ     NA         NA     324.   5.45   1.71  319    119226  387. 
#>  4 HDFCB~ EQ     NA         NA    1259.  18.0    1.45 1241.    95521 1203. 
#>  5 INDUS~ EQ     NA         NA    1488.  19.0    1.3  1469.     8382  125. 
#>  6 YESBA~ EQ     NA         NA      45.7  0.55   1.22   45.2  186780   85.4
#>  7 PNB    EQ     NA         NA      62.1  0.7    1.14   61.4   59732   37.1
#>  8 ICICI~ EQ     NA         NA     531.   5.6    1.07  526.    32942  175. 
#>  9 IDFCF~ EQ     NA         NA      44    0.35   0.8    43.6   44166   19.4
#> 10 FEDER~ EQ     NA         NA      87.8  0.7    0.8    87.1   27146   23.8
#> 11 AXISB~ EQ     NA         NA     728    4.75   0.66  723.    14952  109. 
#> 12 KOTAK~ EQ     NA         NA    1661.   8.5    0.51 1653.     5594   92.9
#> # ... with 7 more variables: mktCap <dbl>, yHigh <dbl>, yLow <dbl>,
#> #   sumVal <dbl>, sumQnty <dbl>, finQnty <dbl>, sumfinQnty <dbl>

# }