Pre Open market data

Fetch data of pre open session of Nifty & Nifty Bank.

nse_preopen_nifty(clean_names = TRUE)

nse_preopen_nifty_bank(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ)

corp_action_date

Corporate action date.

corp_action

Corporate Action

price

Price

change

Change in price

percent_change

Percentage change in price.

prev_close

Previous close.

quantity

Quantity

value

Value (in lakhs),

mkt_cap

Free float market capitalization (in crores).

year_high

Normal market 52 week high.

year_low

Normal market 52 week low.

Examples

# \donttest{
# nifty
nse_preopen_nifty()

#>        symbol series corp_action_date
#> 1        BPCL     EQ       2021-02-17
#> 2   POWERGRID     EQ             <NA>
#> 3        HDFC     EQ             <NA>
#> 4        NTPC     EQ       2021-02-11
#> 5  ADANIPORTS     EQ             <NA>
#> 6    HINDALCO     EQ             <NA>
#> 7  BHARTIARTL     EQ             <NA>
#> 8     DRREDDY     EQ             <NA>
#> 9    AXISBANK     EQ             <NA>
#> 10 HINDUNILVR     EQ             <NA>
#> 11  ICICIBANK     EQ             <NA>
#> 12   DIVISLAB     EQ             <NA>
#> 13    HCLTECH     EQ       2021-01-21
#> 14       INFY     EQ             <NA>
#> 15        UPL     EQ             <NA>
#> 16 BAJAJ-AUTO     EQ             <NA>
#> 17  COALINDIA     EQ             <NA>
#> 18   HDFCBANK     EQ             <NA>
#> 19    SBILIFE     EQ             <NA>
#> 20        ITC     EQ       2021-02-22
#> 21 HEROMOTOCO     EQ       2021-02-15
#> 22        TCS     EQ       2021-01-14
#> 23 BAJFINANCE     EQ             <NA>
#> 24 TATAMOTORS     EQ             <NA>
#> 25        M&M     EQ             <NA>
#> 26       GAIL     EQ       2021-01-27
#> 27  NESTLEIND     EQ             <NA>
#> 28        IOC     EQ       2021-02-09
#> 29      CIPLA     EQ             <NA>
#> 30  TATASTEEL     EQ             <NA>
#> 31   RELIANCE     EQ             <NA>
#> 32 ASIANPAINT     EQ             <NA>
#> 33 INDUSINDBK     EQ             <NA>
#> 34      WIPRO     EQ       2021-01-22
#> 35       SBIN     EQ             <NA>
#> 36  KOTAKBANK     EQ             <NA>
#> 37      TECHM     EQ             <NA>
#> 38         LT     EQ             <NA>
#> 39 BAJAJFINSV     EQ             <NA>
#> 40       ONGC     EQ       2021-02-17
#> 41   JSWSTEEL     EQ             <NA>
#> 42   HDFCLIFE     EQ             <NA>
#> 43     MARUTI     EQ             <NA>
#> 44 ULTRACEMCO     EQ             <NA>
#> 45  SUNPHARMA     EQ       2021-02-09
#> 46   SHREECEM     EQ             <NA>
#> 47  BRITANNIA     EQ             <NA>
#> 48     GRASIM     EQ             <NA>
#> 49  EICHERMOT     EQ             <NA>
#> 50      TITAN     EQ             <NA>
#>                                                             corp_action
#> 1                                    INTERIM DIVIDEND - RS 16 PER SHARE
#> 2                                                                  <NA>
#> 3                                                                  <NA>
#> 4                                     INTERIM DIVIDEND - RS 3 PER SHARE
#> 5                                                                  <NA>
#> 6                                                                  <NA>
#> 7                                                                  <NA>
#> 8                                                                  <NA>
#> 9                                                                  <NA>
#> 10                                                                 <NA>
#> 11                                                                 <NA>
#> 12                                                                 <NA>
#> 13                                    INTERIM DIVIDEND - RS 4 PER SHARE
#> 14                                                                 <NA>
#> 15                                                                 <NA>
#> 16                                                                 <NA>
#> 17                                                                 <NA>
#> 18                                                                 <NA>
#> 19                                                                 <NA>
#> 20                                    INTERIM DIVIDEND - RS 5 PER SHARE
#> 21 INTERIM DIVIDEND - RS 65 PER SHARE SPECIAL DIVIDEND - RS 5 PER SHARE
#> 22                                    INTERIM DIVIDEND - RS 6 PER SHARE
#> 23                                                                 <NA>
#> 24                                                                 <NA>
#> 25                                                                 <NA>
#> 26                                 INTERIM DIVIDEND - RS 2.50 PER SHARE
#> 27                                                                 <NA>
#> 28                                INTERIM DIVIDEND - RS 7.50  PER SHARE
#> 29                                                                 <NA>
#> 30                                                                 <NA>
#> 31                                                                 <NA>
#> 32                                                                 <NA>
#> 33                                                                 <NA>
#> 34                                    INTERIM DIVIDEND - RE 1 PER SHARE
#> 35                                                                 <NA>
#> 36                                                                 <NA>
#> 37                                                                 <NA>
#> 38                                                                 <NA>
#> 39                                                                 <NA>
#> 40                                 INTERIM DIVIDEND - RS 1.75 PER SHARE
#> 41                                                                 <NA>
#> 42                                                                 <NA>
#> 43                                                                 <NA>
#> 44                                                                 <NA>
#> 45                                 INTERIM DIVIDEND - RS 5.50 PER SHARE
#> 46                                                                 <NA>
#> 47                                                                 <NA>
#> 48                                                                 <NA>
#> 49                                                                 <NA>
#> 50                                                                 <NA>
#>       price change percent_change prev_close quantity   value   mkt_cap
#> 1    480.00  11.00           2.35     469.00    53538  256.98  37643.04
#> 2    230.00   4.95           2.20     225.05    57869  133.10  57691.09
#> 3   2608.90  44.90           1.75    2564.00    64336 1678.46 460781.28
#> 4    114.45   1.90           1.69     112.55   237834  272.20  54567.99
#> 5    732.00  11.50           1.60     720.50    58291  426.69  52699.58
#> 6    350.70   5.45           1.58     345.25    57270  200.85  50416.13
#> 7    548.75   7.55           1.40     541.20   127095  697.43 129912.10
#> 8   4465.00  60.30           1.37    4404.70     4737  211.51  53472.01
#> 9    740.50   9.25           1.26     731.25    43619  323.00 181316.88
#> 10  2205.00  27.05           1.24    2177.95    26354  581.11 194454.21
#> 11   618.00   7.50           1.23     610.50    77500  478.95 421448.96
#> 12  3560.00  43.10           1.23    3516.90     4682  166.68  44814.07
#> 13   965.00  11.25           1.18     953.75    49159  474.38 103526.32
#> 14  1319.20  14.70           1.13    1304.50    86168 1136.73 483420.52
#> 15   595.80   6.35           1.08     589.45     8830   52.61  32426.39
#> 16  3985.00  40.95           1.04    3944.05     3001  119.59  51357.51
#> 17   155.00   1.60           1.04     153.40    34066   52.80  32142.33
#> 18  1584.00  15.80           1.01    1568.20    32200  510.05 682331.50
#> 19   888.00   8.85           1.01     879.15     3040   27.00  34288.53
#> 20   212.00   2.10           1.00     209.90    93839  198.94 183380.73
#> 21  3510.00  33.60           0.97    3476.40     7881  276.62  45140.69
#> 22  3035.00  28.65           0.95    3006.35    21758  660.36 315867.49
#> 23  5364.15  47.85           0.90    5316.30     3388  181.74 140955.54
#> 24   348.65   2.90           0.84     345.75   223924  780.71  61944.74
#> 25   866.95   6.90           0.80     860.05    11930  103.43  82329.00
#> 26   145.95   1.15           0.79     144.80    34910   50.95  26775.81
#> 27 16705.00 128.85           0.78   16576.15      647  108.08  59133.45
#> 28   103.00   0.80           0.78     102.20   176423  181.72  25977.43
#> 29   817.95   6.15           0.76     811.80    11728   95.93  41244.40
#> 30   741.00   5.50           0.75     735.50    25496  188.93  55511.71
#> 31  2121.05  15.05           0.71    2106.00    52276 1108.80 680893.97
#> 32  2406.00  15.60           0.65    2390.40     4600  110.68 107764.72
#> 33  1080.00   6.65           0.62    1073.35     7236   78.15  60939.75
#> 34   433.00   2.60           0.60     430.40    40004  173.22  63956.92
#> 35   396.50   2.30           0.58     394.20    72853  288.86 151277.52
#> 36  1855.00  10.35           0.56    1844.65    16583  307.61 270248.78
#> 37   972.20   5.10           0.53     967.10     7931   77.11  59869.87
#> 38  1493.50   7.85           0.53    1485.65     3370   50.33 179416.12
#> 39  9926.40  49.40           0.50    9877.00      256   25.41  59728.42
#> 40   114.05   0.55           0.48     113.50   127686  145.63  41407.99
#> 41   415.30   2.00           0.48     413.30     7395   30.71  39961.49
#> 42   724.00   3.20           0.44     720.80     4139   29.97  56782.67
#> 43  7245.60  31.50           0.44    7214.10     1513  109.63  95886.37
#> 44  6450.00  26.90           0.42    6423.10      845   54.50  74157.75
#> 45   616.50   2.30           0.37     614.20     9749   60.10  66315.22
#> 46 27755.50  96.75           0.35   27658.75      141   39.14  36924.09
#> 47  3460.00  12.05           0.35    3447.95      866   29.96  40694.59
#> 48  1299.00   4.25           0.33    1294.75     3212   41.72  50260.14
#> 49  2605.80   3.50           0.13    2602.30     2632   68.58  36264.71
#> 50  1472.00   0.85           0.06    1471.15     5260   77.43  61385.13
#>    year_high year_low  sum_val sum_quantity fin_quantity sum_fin_quantity
#> 1     482.40   252.00 13535.04      1342962        53538          2010060
#> 2     238.30   122.15 13535.04      1342962        57869          2010060
#> 3    2896.00  1473.45 13535.04      1342962        64336          2010060
#> 4     113.70    73.20 13535.04      1342962       237834          2010060
#> 5     724.70   203.00 13535.04      1342962        58291          2010060
#> 6     354.60    84.90 13535.04      1342962        57270          2010060
#> 7     623.00   361.75 13535.04      1342962       127095          2010060
#> 8    5512.65  2495.05 13535.04      1342962         4737          2010060
#> 9     799.00   286.00 13535.04      1342962        43619          2010060
#> 10   2614.30  1757.30 13535.04      1342962        26354          2010060
#> 11    679.40   268.30 13535.04      1342962        77500          2010060
#> 12   3914.95  1626.00 13535.04      1342962         4682          2010060
#> 13   1067.00   375.25 13535.04      1342962        49159          2010060
#> 14   1392.80   509.25 13535.04      1342962        86168          2010060
#> 15    608.80   240.15 13535.04      1342962         8830          2010060
#> 16   4361.40  1788.65 13535.04      1342962         3001          2010060
#> 17    180.20   109.55 13535.04      1342962        34066          2010060
#> 18   1641.00   738.75 13535.04      1342962        32200          2010060
#> 19    954.50   519.40 13535.04      1342962         3040          2010060
#> 20    239.20   134.60 13535.04      1342962        93839          2010060
#> 21   3629.05  1475.00 13535.04      1342962         7881          2010060
#> 22   3339.80  1506.05 13535.04      1342962        21758          2010060
#> 23   5822.20  1783.00 13535.04      1342962         3388          2010060
#> 24    347.20    63.50 13535.04      1342962       223924          2010060
#> 25    952.05   245.40 13535.04      1342962        11930          2010060
#> 26    152.80    65.00 13535.04      1342962        34910          2010060
#> 27  18844.00 12200.00 13535.04      1342962          647          2010060
#> 28    112.90    71.15 13535.04      1342962       176423          2010060
#> 29    878.90   355.30 13535.04      1342962        11728          2010060
#> 30    753.00   250.85 13535.04      1342962        25496          2010060
#> 31   2369.35   866.98 13535.04      1342962        52276          2010060
#> 32   2873.45  1431.20 13535.04      1342962         4600          2010060
#> 33   1136.95   235.55 13535.04      1342962         7236          2010060
#> 34    467.45   159.40 13535.04      1342962        40004          2010060
#> 35    427.70   149.45 13535.04      1342962        72853          2010060
#> 36   2049.00  1001.00 13535.04      1342962        16583          2010060
#> 37   1081.55   471.40 13535.04      1342962         7931          2010060
#> 38   1593.00   661.00 13535.04      1342962         3370          2010060
#> 39  10579.00  3985.30 13535.04      1342962          256          2010060
#> 40    120.50    50.00 13535.04      1342962       127686          2010060
#> 41    426.60   132.50 13535.04      1342962         7395          2010060
#> 42    744.00   340.00 13535.04      1342962         4139          2010060
#> 43   8329.00  4001.10 13535.04      1342962         1513          2010060
#> 44   6589.00  2910.00 13535.04      1342962          845          2010060
#> 45    654.40   312.00 13535.04      1342962         9749          2010060
#> 46  29090.00 15410.00 13535.04      1342962          141          2010060
#> 47   4010.00  2100.00 13535.04      1342962          866          2010060
#> 48   1298.50   385.05 13535.04      1342962         3212          2010060
#> 49   3037.00  1245.01 13535.04      1342962         2632          2010060
#> 50   1621.35   720.90 13535.04      1342962         5260          2010060

# retain original column names as returned by NSE
nse_preopen_nifty(clean_names = FALSE)

#>        symbol series        xDt
#> 1        BPCL     EQ 2021-02-17
#> 2   POWERGRID     EQ       <NA>
#> 3        HDFC     EQ       <NA>
#> 4        NTPC     EQ 2021-02-11
#> 5  ADANIPORTS     EQ       <NA>
#> 6    HINDALCO     EQ       <NA>
#> 7  BHARTIARTL     EQ       <NA>
#> 8     DRREDDY     EQ       <NA>
#> 9    AXISBANK     EQ       <NA>
#> 10 HINDUNILVR     EQ       <NA>
#> 11  ICICIBANK     EQ       <NA>
#> 12   DIVISLAB     EQ       <NA>
#> 13    HCLTECH     EQ 2021-01-21
#> 14       INFY     EQ       <NA>
#> 15        UPL     EQ       <NA>
#> 16 BAJAJ-AUTO     EQ       <NA>
#> 17  COALINDIA     EQ       <NA>
#> 18   HDFCBANK     EQ       <NA>
#> 19    SBILIFE     EQ       <NA>
#> 20        ITC     EQ 2021-02-22
#> 21 HEROMOTOCO     EQ 2021-02-15
#> 22        TCS     EQ 2021-01-14
#> 23 BAJFINANCE     EQ       <NA>
#> 24 TATAMOTORS     EQ       <NA>
#> 25        M&M     EQ       <NA>
#> 26       GAIL     EQ 2021-01-27
#> 27  NESTLEIND     EQ       <NA>
#> 28        IOC     EQ 2021-02-09
#> 29      CIPLA     EQ       <NA>
#> 30  TATASTEEL     EQ       <NA>
#> 31   RELIANCE     EQ       <NA>
#> 32 ASIANPAINT     EQ       <NA>
#> 33 INDUSINDBK     EQ       <NA>
#> 34      WIPRO     EQ 2021-01-22
#> 35       SBIN     EQ       <NA>
#> 36  KOTAKBANK     EQ       <NA>
#> 37      TECHM     EQ       <NA>
#> 38         LT     EQ       <NA>
#> 39 BAJAJFINSV     EQ       <NA>
#> 40       ONGC     EQ 2021-02-17
#> 41   JSWSTEEL     EQ       <NA>
#> 42   HDFCLIFE     EQ       <NA>
#> 43     MARUTI     EQ       <NA>
#> 44 ULTRACEMCO     EQ       <NA>
#> 45  SUNPHARMA     EQ 2021-02-09
#> 46   SHREECEM     EQ       <NA>
#> 47  BRITANNIA     EQ       <NA>
#> 48     GRASIM     EQ       <NA>
#> 49  EICHERMOT     EQ       <NA>
#> 50      TITAN     EQ       <NA>
#>                                                                   caAct
#> 1                                    INTERIM DIVIDEND - RS 16 PER SHARE
#> 2                                                                  <NA>
#> 3                                                                  <NA>
#> 4                                     INTERIM DIVIDEND - RS 3 PER SHARE
#> 5                                                                  <NA>
#> 6                                                                  <NA>
#> 7                                                                  <NA>
#> 8                                                                  <NA>
#> 9                                                                  <NA>
#> 10                                                                 <NA>
#> 11                                                                 <NA>
#> 12                                                                 <NA>
#> 13                                    INTERIM DIVIDEND - RS 4 PER SHARE
#> 14                                                                 <NA>
#> 15                                                                 <NA>
#> 16                                                                 <NA>
#> 17                                                                 <NA>
#> 18                                                                 <NA>
#> 19                                                                 <NA>
#> 20                                    INTERIM DIVIDEND - RS 5 PER SHARE
#> 21 INTERIM DIVIDEND - RS 65 PER SHARE SPECIAL DIVIDEND - RS 5 PER SHARE
#> 22                                    INTERIM DIVIDEND - RS 6 PER SHARE
#> 23                                                                 <NA>
#> 24                                                                 <NA>
#> 25                                                                 <NA>
#> 26                                 INTERIM DIVIDEND - RS 2.50 PER SHARE
#> 27                                                                 <NA>
#> 28                                INTERIM DIVIDEND - RS 7.50  PER SHARE
#> 29                                                                 <NA>
#> 30                                                                 <NA>
#> 31                                                                 <NA>
#> 32                                                                 <NA>
#> 33                                                                 <NA>
#> 34                                    INTERIM DIVIDEND - RE 1 PER SHARE
#> 35                                                                 <NA>
#> 36                                                                 <NA>
#> 37                                                                 <NA>
#> 38                                                                 <NA>
#> 39                                                                 <NA>
#> 40                                 INTERIM DIVIDEND - RS 1.75 PER SHARE
#> 41                                                                 <NA>
#> 42                                                                 <NA>
#> 43                                                                 <NA>
#> 44                                                                 <NA>
#> 45                                 INTERIM DIVIDEND - RS 5.50 PER SHARE
#> 46                                                                 <NA>
#> 47                                                                 <NA>
#> 48                                                                 <NA>
#> 49                                                                 <NA>
#> 50                                                                 <NA>
#>         iep    chn perChn     pCls trdQnty    iVal    mktCap    yHigh     yLow
#> 1    480.00  11.00   2.35   469.00   53538  256.98  37643.04   482.40   252.00
#> 2    230.00   4.95   2.20   225.05   57869  133.10  57691.09   238.30   122.15
#> 3   2608.90  44.90   1.75  2564.00   64336 1678.46 460781.28  2896.00  1473.45
#> 4    114.45   1.90   1.69   112.55  237834  272.20  54567.99   113.70    73.20
#> 5    732.00  11.50   1.60   720.50   58291  426.69  52699.58   724.70   203.00
#> 6    350.70   5.45   1.58   345.25   57270  200.85  50416.13   354.60    84.90
#> 7    548.75   7.55   1.40   541.20  127095  697.43 129912.10   623.00   361.75
#> 8   4465.00  60.30   1.37  4404.70    4737  211.51  53472.01  5512.65  2495.05
#> 9    740.50   9.25   1.26   731.25   43619  323.00 181316.88   799.00   286.00
#> 10  2205.00  27.05   1.24  2177.95   26354  581.11 194454.21  2614.30  1757.30
#> 11   618.00   7.50   1.23   610.50   77500  478.95 421448.96   679.40   268.30
#> 12  3560.00  43.10   1.23  3516.90    4682  166.68  44814.07  3914.95  1626.00
#> 13   965.00  11.25   1.18   953.75   49159  474.38 103526.32  1067.00   375.25
#> 14  1319.20  14.70   1.13  1304.50   86168 1136.73 483420.52  1392.80   509.25
#> 15   595.80   6.35   1.08   589.45    8830   52.61  32426.39   608.80   240.15
#> 16  3985.00  40.95   1.04  3944.05    3001  119.59  51357.51  4361.40  1788.65
#> 17   155.00   1.60   1.04   153.40   34066   52.80  32142.33   180.20   109.55
#> 18  1584.00  15.80   1.01  1568.20   32200  510.05 682331.50  1641.00   738.75
#> 19   888.00   8.85   1.01   879.15    3040   27.00  34288.53   954.50   519.40
#> 20   212.00   2.10   1.00   209.90   93839  198.94 183380.73   239.20   134.60
#> 21  3510.00  33.60   0.97  3476.40    7881  276.62  45140.69  3629.05  1475.00
#> 22  3035.00  28.65   0.95  3006.35   21758  660.36 315867.49  3339.80  1506.05
#> 23  5364.15  47.85   0.90  5316.30    3388  181.74 140955.54  5822.20  1783.00
#> 24   348.65   2.90   0.84   345.75  223924  780.71  61944.74   347.20    63.50
#> 25   866.95   6.90   0.80   860.05   11930  103.43  82329.00   952.05   245.40
#> 26   145.95   1.15   0.79   144.80   34910   50.95  26775.81   152.80    65.00
#> 27 16705.00 128.85   0.78 16576.15     647  108.08  59133.45 18844.00 12200.00
#> 28   103.00   0.80   0.78   102.20  176423  181.72  25977.43   112.90    71.15
#> 29   817.95   6.15   0.76   811.80   11728   95.93  41244.40   878.90   355.30
#> 30   741.00   5.50   0.75   735.50   25496  188.93  55511.71   753.00   250.85
#> 31  2121.05  15.05   0.71  2106.00   52276 1108.80 680893.97  2369.35   866.98
#> 32  2406.00  15.60   0.65  2390.40    4600  110.68 107764.72  2873.45  1431.20
#> 33  1080.00   6.65   0.62  1073.35    7236   78.15  60939.75  1136.95   235.55
#> 34   433.00   2.60   0.60   430.40   40004  173.22  63956.92   467.45   159.40
#> 35   396.50   2.30   0.58   394.20   72853  288.86 151277.52   427.70   149.45
#> 36  1855.00  10.35   0.56  1844.65   16583  307.61 270248.78  2049.00  1001.00
#> 37   972.20   5.10   0.53   967.10    7931   77.11  59869.87  1081.55   471.40
#> 38  1493.50   7.85   0.53  1485.65    3370   50.33 179416.12  1593.00   661.00
#> 39  9926.40  49.40   0.50  9877.00     256   25.41  59728.42 10579.00  3985.30
#> 40   114.05   0.55   0.48   113.50  127686  145.63  41407.99   120.50    50.00
#> 41   415.30   2.00   0.48   413.30    7395   30.71  39961.49   426.60   132.50
#> 42   724.00   3.20   0.44   720.80    4139   29.97  56782.67   744.00   340.00
#> 43  7245.60  31.50   0.44  7214.10    1513  109.63  95886.37  8329.00  4001.10
#> 44  6450.00  26.90   0.42  6423.10     845   54.50  74157.75  6589.00  2910.00
#> 45   616.50   2.30   0.37   614.20    9749   60.10  66315.22   654.40   312.00
#> 46 27755.50  96.75   0.35 27658.75     141   39.14  36924.09 29090.00 15410.00
#> 47  3460.00  12.05   0.35  3447.95     866   29.96  40694.59  4010.00  2100.00
#> 48  1299.00   4.25   0.33  1294.75    3212   41.72  50260.14  1298.50   385.05
#> 49  2605.80   3.50   0.13  2602.30    2632   68.58  36264.71  3037.00  1245.01
#> 50  1472.00   0.85   0.06  1471.15    5260   77.43  61385.13  1621.35   720.90
#>      sumVal sumQnty finQnty sumfinQnty
#> 1  13535.04 1342962   53538    2010060
#> 2  13535.04 1342962   57869    2010060
#> 3  13535.04 1342962   64336    2010060
#> 4  13535.04 1342962  237834    2010060
#> 5  13535.04 1342962   58291    2010060
#> 6  13535.04 1342962   57270    2010060
#> 7  13535.04 1342962  127095    2010060
#> 8  13535.04 1342962    4737    2010060
#> 9  13535.04 1342962   43619    2010060
#> 10 13535.04 1342962   26354    2010060
#> 11 13535.04 1342962   77500    2010060
#> 12 13535.04 1342962    4682    2010060
#> 13 13535.04 1342962   49159    2010060
#> 14 13535.04 1342962   86168    2010060
#> 15 13535.04 1342962    8830    2010060
#> 16 13535.04 1342962    3001    2010060
#> 17 13535.04 1342962   34066    2010060
#> 18 13535.04 1342962   32200    2010060
#> 19 13535.04 1342962    3040    2010060
#> 20 13535.04 1342962   93839    2010060
#> 21 13535.04 1342962    7881    2010060
#> 22 13535.04 1342962   21758    2010060
#> 23 13535.04 1342962    3388    2010060
#> 24 13535.04 1342962  223924    2010060
#> 25 13535.04 1342962   11930    2010060
#> 26 13535.04 1342962   34910    2010060
#> 27 13535.04 1342962     647    2010060
#> 28 13535.04 1342962  176423    2010060
#> 29 13535.04 1342962   11728    2010060
#> 30 13535.04 1342962   25496    2010060
#> 31 13535.04 1342962   52276    2010060
#> 32 13535.04 1342962    4600    2010060
#> 33 13535.04 1342962    7236    2010060
#> 34 13535.04 1342962   40004    2010060
#> 35 13535.04 1342962   72853    2010060
#> 36 13535.04 1342962   16583    2010060
#> 37 13535.04 1342962    7931    2010060
#> 38 13535.04 1342962    3370    2010060
#> 39 13535.04 1342962     256    2010060
#> 40 13535.04 1342962  127686    2010060
#> 41 13535.04 1342962    7395    2010060
#> 42 13535.04 1342962    4139    2010060
#> 43 13535.04 1342962    1513    2010060
#> 44 13535.04 1342962     845    2010060
#> 45 13535.04 1342962    9749    2010060
#> 46 13535.04 1342962     141    2010060
#> 47 13535.04 1342962     866    2010060
#> 48 13535.04 1342962    3212    2010060
#> 49 13535.04 1342962    2632    2010060
#> 50 13535.04 1342962    5260    2010060

# nifty bank
nse_preopen_nifty_bank()

#>        symbol series corp_action_date corp_action  price change percent_change
#> 1    AXISBANK     EQ             <NA>        <NA>  740.5   9.25           1.26
#> 2   ICICIBANK     EQ             <NA>        <NA>  618.0   7.50           1.23
#> 3  BANDHANBNK     EQ             <NA>        <NA>  349.0   3.80           1.10
#> 4    HDFCBANK     EQ             <NA>        <NA> 1584.0  15.80           1.01
#> 5         PNB     EQ             <NA>        <NA>   41.4   0.40           0.98
#> 6  BANKBARODA     EQ             <NA>        <NA>   84.1   0.80           0.96
#> 7     RBLBANK     EQ             <NA>        <NA>  238.9   2.15           0.91
#> 8  INDUSINDBK     EQ             <NA>        <NA> 1080.0   6.65           0.62
#> 9        SBIN     EQ             <NA>        <NA>  396.5   2.30           0.58
#> 10  KOTAKBANK     EQ             <NA>        <NA> 1855.0  10.35           0.56
#> 11 FEDERALBNK     EQ             <NA>        <NA>   83.7   0.40           0.48
#> 12 IDFCFIRSTB     EQ             <NA>        <NA>   64.8   0.30           0.47
#>    prev_close quantity  value   mkt_cap year_high year_low sum_val sum_quantity
#> 1      731.25    43619 323.00 181316.88    799.00   286.00 2326.69       609386
#> 2      610.50    77500 478.95 421448.96    679.40   268.30 2326.69       609386
#> 3      345.20    16205  56.56  23347.82    430.70   152.20 2326.69       609386
#> 4     1568.20    32200 510.05 682331.50   1641.00   738.75 2326.69       609386
#> 5       41.00   208449  86.30   9880.45     46.60    26.30 2326.69       609386
#> 6       83.30   158345 133.17  10007.22     99.85    36.00 2326.69       609386
#> 7      236.75     6920  16.53  10613.87    308.10   101.55 2326.69       609386
#> 8     1073.35     7236  78.15  60939.75   1136.95   235.55 2326.69       609386
#> 9      394.20    72853 288.86 151277.52    427.70   149.45 2326.69       609386
#> 10    1844.65    16583 307.61 270248.78   2049.00  1001.00 2326.69       609386
#> 11      83.30    33114  27.72  16632.33     88.20    35.70 2326.69       609386
#> 12      64.50    30545  19.79  13537.21     66.80    17.65 2326.69       609386
#>    fin_quantity sum_fin_quantity
#> 1         43619           703569
#> 2         77500           703569
#> 3         16205           703569
#> 4         32200           703569
#> 5        208449           703569
#> 6        158345           703569
#> 7          6920           703569
#> 8          7236           703569
#> 9         72853           703569
#> 10        16583           703569
#> 11        33114           703569
#> 12        30545           703569

# retain original column names as returned by NSE
nse_preopen_nifty_bank(clean_names = FALSE)

#>        symbol series  xDt caAct    iep   chn perChn    pCls trdQnty   iVal
#> 1    AXISBANK     EQ <NA>  <NA>  740.5  9.25   1.26  731.25   43619 323.00
#> 2   ICICIBANK     EQ <NA>  <NA>  618.0  7.50   1.23  610.50   77500 478.95
#> 3  BANDHANBNK     EQ <NA>  <NA>  349.0  3.80   1.10  345.20   16205  56.56
#> 4    HDFCBANK     EQ <NA>  <NA> 1584.0 15.80   1.01 1568.20   32200 510.05
#> 5         PNB     EQ <NA>  <NA>   41.4  0.40   0.98   41.00  208449  86.30
#> 6  BANKBARODA     EQ <NA>  <NA>   84.1  0.80   0.96   83.30  158345 133.17
#> 7     RBLBANK     EQ <NA>  <NA>  238.9  2.15   0.91  236.75    6920  16.53
#> 8  INDUSINDBK     EQ <NA>  <NA> 1080.0  6.65   0.62 1073.35    7236  78.15
#> 9        SBIN     EQ <NA>  <NA>  396.5  2.30   0.58  394.20   72853 288.86
#> 10  KOTAKBANK     EQ <NA>  <NA> 1855.0 10.35   0.56 1844.65   16583 307.61
#> 11 FEDERALBNK     EQ <NA>  <NA>   83.7  0.40   0.48   83.30   33114  27.72
#> 12 IDFCFIRSTB     EQ <NA>  <NA>   64.8  0.30   0.47   64.50   30545  19.79
#>       mktCap   yHigh    yLow  sumVal sumQnty finQnty sumfinQnty
#> 1  181316.88  799.00  286.00 2326.69  609386   43619     703569
#> 2  421448.96  679.40  268.30 2326.69  609386   77500     703569
#> 3   23347.82  430.70  152.20 2326.69  609386   16205     703569
#> 4  682331.50 1641.00  738.75 2326.69  609386   32200     703569
#> 5    9880.45   46.60   26.30 2326.69  609386  208449     703569
#> 6   10007.22   99.85   36.00 2326.69  609386  158345     703569
#> 7   10613.87  308.10  101.55 2326.69  609386    6920     703569
#> 8   60939.75 1136.95  235.55 2326.69  609386    7236     703569
#> 9  151277.52  427.70  149.45 2326.69  609386   72853     703569
#> 10 270248.78 2049.00 1001.00 2326.69  609386   16583     703569
#> 11  16632.33   88.20   35.70 2326.69  609386   33114     703569
#> 12  13537.21   66.80   17.65 2326.69  609386   30545     703569

# }