Pre Open market data

Fetch data of pre open session of Nifty & Nifty Bank.

nse_preopen_nifty(clean_names = TRUE)

nse_preopen_nifty_bank(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ)

corp_action_date

Corporate action date.

corp_action

Corporate Action

price

Price

change

Change in price

percent_change

Percentage change in price.

prev_close

Previous close.

quantity

Quantity

value

Value (in lakhs),

mkt_cap

Free float market capitalization (in crores).

year_high

Normal market 52 week high.

year_low

Normal market 52 week low.

Examples

# \donttest{
# nifty
nse_preopen_nifty()

#>        symbol series corp_action_date
#> 1   TATASTEEL     EQ             <NA>
#> 2    JSWSTEEL     EQ             <NA>
#> 3    AXISBANK     EQ             <NA>
#> 4  TATAMOTORS     EQ             <NA>
#> 5    HDFCBANK     EQ             <NA>
#> 6  BAJAJFINSV     EQ             <NA>
#> 7   ICICIBANK     EQ             <NA>
#> 8    HINDALCO     EQ             <NA>
#> 9  BAJFINANCE     EQ             <NA>
#> 10 INDUSINDBK     EQ             <NA>
#> 11  POWERGRID     EQ             <NA>
#> 12        UPL     EQ             <NA>
#> 13  KOTAKBANK     EQ             <NA>
#> 14       HDFC     EQ             <NA>
#> 15      TECHM     EQ             <NA>
#> 16         LT     EQ             <NA>
#> 17       SBIN     EQ             <NA>
#> 18      TITAN     EQ             <NA>
#> 19      WIPRO     EQ       2021-01-22
#> 20    HCLTECH     EQ       2021-01-21
#> 21        M&M     EQ             <NA>
#> 22        TCS     EQ       2021-01-14
#> 23 BHARTIARTL     EQ             <NA>
#> 24  EICHERMOT     EQ             <NA>
#> 25 ULTRACEMCO     EQ             <NA>
#> 26        IOC     EQ       2021-02-09
#> 27       NTPC     EQ       2021-02-11
#> 28   RELIANCE     EQ             <NA>
#> 29       GAIL     EQ       2021-01-27
#> 30       BPCL     EQ       2021-02-17
#> 31 ASIANPAINT     EQ             <NA>
#> 32     GRASIM     EQ             <NA>
#> 33   HDFCLIFE     EQ             <NA>
#> 34   SHREECEM     EQ             <NA>
#> 35  BRITANNIA     EQ             <NA>
#> 36 HEROMOTOCO     EQ       2021-02-15
#> 37    SBILIFE     EQ             <NA>
#> 38 BAJAJ-AUTO     EQ             <NA>
#> 39       INFY     EQ             <NA>
#> 40     MARUTI     EQ             <NA>
#> 41      CIPLA     EQ             <NA>
#> 42  COALINDIA     EQ             <NA>
#> 43        ITC     EQ       2021-02-22
#> 44    DRREDDY     EQ             <NA>
#> 45   DIVISLAB     EQ             <NA>
#> 46 HINDUNILVR     EQ             <NA>
#> 47  SUNPHARMA     EQ       2021-02-09
#> 48  NESTLEIND     EQ             <NA>
#> 49       ONGC     EQ       2021-02-17
#> 50 ADANIPORTS     EQ             <NA>
#>                                                             corp_action
#> 1                                                                  <NA>
#> 2                                                                  <NA>
#> 3                                                                  <NA>
#> 4                                                                  <NA>
#> 5                                                                  <NA>
#> 6                                                                  <NA>
#> 7                                                                  <NA>
#> 8                                                                  <NA>
#> 9                                                                  <NA>
#> 10                                                                 <NA>
#> 11                                                                 <NA>
#> 12                                                                 <NA>
#> 13                                                                 <NA>
#> 14                                                                 <NA>
#> 15                                                                 <NA>
#> 16                                                                 <NA>
#> 17                                                                 <NA>
#> 18                                                                 <NA>
#> 19                                    INTERIM DIVIDEND - RE 1 PER SHARE
#> 20                                    INTERIM DIVIDEND - RS 4 PER SHARE
#> 21                                                                 <NA>
#> 22                                    INTERIM DIVIDEND - RS 6 PER SHARE
#> 23                                                                 <NA>
#> 24                                                                 <NA>
#> 25                                                                 <NA>
#> 26                                INTERIM DIVIDEND - RS 7.50  PER SHARE
#> 27                                    INTERIM DIVIDEND - RS 3 PER SHARE
#> 28                                                                 <NA>
#> 29                                 INTERIM DIVIDEND - RS 2.50 PER SHARE
#> 30                                   INTERIM DIVIDEND - RS 16 PER SHARE
#> 31                                                                 <NA>
#> 32                                                                 <NA>
#> 33                                                                 <NA>
#> 34                                                                 <NA>
#> 35                                                                 <NA>
#> 36 INTERIM DIVIDEND - RS 65 PER SHARE SPECIAL DIVIDEND - RS 5 PER SHARE
#> 37                                                                 <NA>
#> 38                                                                 <NA>
#> 39                                                                 <NA>
#> 40                                                                 <NA>
#> 41                                                                 <NA>
#> 42                                                                 <NA>
#> 43                                    INTERIM DIVIDEND - RS 5 PER SHARE
#> 44                                                                 <NA>
#> 45                                                                 <NA>
#> 46                                                                 <NA>
#> 47                                 INTERIM DIVIDEND - RS 5.50 PER SHARE
#> 48                                                                 <NA>
#> 49                                 INTERIM DIVIDEND - RS 1.75 PER SHARE
#> 50                                                                 <NA>
#>       price  change percent_change prev_close quantity   value   mkt_cap
#> 1    751.00  -26.15          -3.36     777.15    84566  635.09  58655.24
#> 2    415.10  -13.05          -3.05     428.15    45527  188.98  41397.32
#> 3    731.10  -22.85          -3.03     753.95    57795  422.54 186945.45
#> 4    339.00   -9.50          -2.73     348.50   222386  753.89  62437.43
#> 5   1548.55  -38.30          -2.41    1586.85    60993  944.51 690446.21
#> 6  10150.00 -245.60          -2.36   10395.60      921   93.48  62864.51
#> 7    618.75  -13.35          -2.11     632.10   110691  684.90 436360.17
#> 8    352.00   -7.35          -2.05     359.35   123649  435.24  52475.12
#> 9   5430.00 -112.90          -2.04    5542.90     7567  410.89 146963.58
#> 10  1078.00  -21.75          -1.98    1099.75    16466  177.50  62438.62
#> 11   223.05   -4.45          -1.96     227.50    22232   49.59  58319.15
#> 12   602.00  -11.60          -1.89     613.60    15660   94.27  33754.92
#> 13  1863.90  -35.60          -1.87    1899.50     5176   96.48 278284.53
#> 14  2610.00  -44.90          -1.69    2654.90    16205  422.95 477117.09
#> 15   960.00  -16.00          -1.64     976.00     4469   42.90  60420.83
#> 16  1473.00  -23.90          -1.60    1496.90     6007   88.48 180774.74
#> 17   398.70   -6.35          -1.57     405.05   165134  658.39 155441.30
#> 18  1452.00  -22.60          -1.53    1474.60     9089  131.97  61529.09
#> 19   428.90   -6.60          -1.52     435.50    26909  115.41  64714.77
#> 20   952.00  -13.25          -1.37     965.25    45035  428.73 104774.61
#> 21   841.85  -10.20          -1.20     852.05     9358   78.78  81563.19
#> 22  3023.00  -36.10          -1.18    3059.10    16407  495.98 321409.76
#> 23   540.00   -6.30          -1.15     546.30    80561  435.03 131136.32
#> 24  2570.00  -29.05          -1.12    2599.05     2275   58.47  36219.42
#> 25  6429.00  -71.50          -1.10    6500.50      864   55.55  75051.37
#> 26   101.50   -1.10          -1.07     102.60    82112   83.34  26079.10
#> 27   111.50   -1.20          -1.06     112.70    87190   97.22  54640.71
#> 28  2180.00  -22.10          -1.00    2202.10   119360 2602.05 711964.20
#> 29   145.50   -1.45          -0.99     146.95    45042   65.54  27173.38
#> 30   462.00   -4.35          -0.93     466.35    16624   76.80  37430.35
#> 31  2380.00  -22.10          -0.92    2402.10     5634  134.09 108292.18
#> 32  1306.00  -12.10          -0.92    1318.10     2928   38.24  51166.55
#> 33   721.00   -6.50          -0.89     727.50     3606   26.00  57310.48
#> 34 27450.00 -245.25          -0.89   27695.25       39   10.71  36972.82
#> 35  3450.00  -31.15          -0.89    3481.15     1277   44.06  41086.43
#> 36  3398.00  -26.35          -0.77    3424.35     1749   59.43  44464.82
#> 37   899.70   -6.75          -0.74     906.45     5118   46.05  35353.29
#> 38  3875.00  -26.10          -0.67    3901.10      836   32.40  50798.24
#> 39  1335.00   -8.55          -0.64    1343.55   105459 1407.88 497891.64
#> 40  7080.00  -44.70          -0.63    7124.70     4552  322.28  94698.11
#> 41   807.40   -5.10          -0.63     812.50     4941   39.89  41279.96
#> 42   154.00   -0.90          -0.58     154.90    57942   89.23  32456.63
#> 43   208.80   -1.15          -0.55     209.95   125401  261.84 183424.41
#> 44  4482.00  -20.55          -0.46    4502.55     3441  154.23  54659.89
#> 45  3550.00  -15.15          -0.42    3565.15     1554   55.17  45428.89
#> 46  2185.00   -8.50          -0.39    2193.50    17442  381.11 195842.56
#> 47   622.00   -2.35          -0.38     624.35    24962  155.26  67411.12
#> 48 16656.55  -42.70          -0.26   16699.25      335   55.80  59572.60
#> 49   113.95   -0.05          -0.04     114.00   129498  147.56  41590.40
#> 50   733.25    3.40           0.47     729.85   161134 1181.52  53383.46
#>    year_high year_low  sum_val sum_quantity fin_quantity sum_fin_quantity
#> 1     782.50   250.85 15567.69      1552461        84566          2164118
#> 2     435.00   132.50 15567.69      1552461        45527          2164118
#> 3     799.00   286.00 15567.69      1552461        57795          2164118
#> 4     357.00    63.50 15567.69      1552461       222386          2164118
#> 5    1641.00   738.75 15567.69      1552461        60993          2164118
#> 6   10579.00  3985.30 15567.69      1552461          921          2164118
#> 7     679.40   268.30 15567.69      1552461       110691          2164118
#> 8     361.30    84.90 15567.69      1552461       123649          2164118
#> 9    5822.20  1783.00 15567.69      1552461         7567          2164118
#> 10   1119.50   235.55 15567.69      1552461        16466          2164118
#> 11    238.30   122.15 15567.69      1552461        22232          2164118
#> 12    616.85   240.15 15567.69      1552461        15660          2164118
#> 13   2049.00  1001.00 15567.69      1552461         5176          2164118
#> 14   2896.00  1473.45 15567.69      1552461        16205          2164118
#> 15   1081.55   471.40 15567.69      1552461         4469          2164118
#> 16   1593.00   661.00 15567.69      1552461         6007          2164118
#> 17    427.70   149.45 15567.69      1552461       165134          2164118
#> 18   1621.35   720.90 15567.69      1552461         9089          2164118
#> 19    467.45   159.40 15567.69      1552461        26909          2164118
#> 20   1067.00   375.25 15567.69      1552461        45035          2164118
#> 21    952.05   245.40 15567.69      1552461         9358          2164118
#> 22   3339.80  1506.05 15567.69      1552461        16407          2164118
#> 23    623.00   361.75 15567.69      1552461        80561          2164118
#> 24   3037.00  1245.01 15567.69      1552461         2275          2164118
#> 25   6589.00  2910.00 15567.69      1552461          864          2164118
#> 26    108.75    71.15 15567.69      1552461        82112          2164118
#> 27    114.70    73.20 15567.69      1552461        87190          2164118
#> 28   2369.35   866.98 15567.69      1552461       119360          2164118
#> 29    152.80    65.00 15567.69      1552461        45042          2164118
#> 30    482.40   252.00 15567.69      1552461        16624          2164118
#> 31   2873.45  1431.20 15567.69      1552461         5634          2164118
#> 32   1343.00   385.05 15567.69      1552461         2928          2164118
#> 33    744.00   340.00 15567.69      1552461         3606          2164118
#> 34  29090.00 15410.00 15567.69      1552461           39          2164118
#> 35   4010.00  2100.00 15567.69      1552461         1277          2164118
#> 36   3629.05  1475.00 15567.69      1552461         1749          2164118
#> 37    954.50   519.40 15567.69      1552461         5118          2164118
#> 38   4361.40  1788.65 15567.69      1552461          836          2164118
#> 39   1392.80   509.25 15567.69      1552461       105459          2164118
#> 40   8329.00  4001.10 15567.69      1552461         4552          2164118
#> 41    878.90   355.30 15567.69      1552461         4941          2164118
#> 42    180.20   109.55 15567.69      1552461        57942          2164118
#> 43    239.20   134.60 15567.69      1552461       125401          2164118
#> 44   5512.65  2495.05 15567.69      1552461         3441          2164118
#> 45   3914.95  1626.00 15567.69      1552461         1554          2164118
#> 46   2614.30  1757.30 15567.69      1552461        17442          2164118
#> 47    654.40   312.00 15567.69      1552461        24962          2164118
#> 48  18844.00 12200.00 15567.69      1552461          335          2164118
#> 49    120.50    50.00 15567.69      1552461       129498          2164118
#> 50    754.95   203.00 15567.69      1552461       161134          2164118

# retain original column names as returned by NSE
nse_preopen_nifty(clean_names = FALSE)

#>        symbol series        xDt
#> 1   TATASTEEL     EQ       <NA>
#> 2    JSWSTEEL     EQ       <NA>
#> 3    AXISBANK     EQ       <NA>
#> 4  TATAMOTORS     EQ       <NA>
#> 5    HDFCBANK     EQ       <NA>
#> 6  BAJAJFINSV     EQ       <NA>
#> 7   ICICIBANK     EQ       <NA>
#> 8    HINDALCO     EQ       <NA>
#> 9  BAJFINANCE     EQ       <NA>
#> 10 INDUSINDBK     EQ       <NA>
#> 11  POWERGRID     EQ       <NA>
#> 12        UPL     EQ       <NA>
#> 13  KOTAKBANK     EQ       <NA>
#> 14       HDFC     EQ       <NA>
#> 15      TECHM     EQ       <NA>
#> 16         LT     EQ       <NA>
#> 17       SBIN     EQ       <NA>
#> 18      TITAN     EQ       <NA>
#> 19      WIPRO     EQ 2021-01-22
#> 20    HCLTECH     EQ 2021-01-21
#> 21        M&M     EQ       <NA>
#> 22        TCS     EQ 2021-01-14
#> 23 BHARTIARTL     EQ       <NA>
#> 24  EICHERMOT     EQ       <NA>
#> 25 ULTRACEMCO     EQ       <NA>
#> 26        IOC     EQ 2021-02-09
#> 27       NTPC     EQ 2021-02-11
#> 28   RELIANCE     EQ       <NA>
#> 29       GAIL     EQ 2021-01-27
#> 30       BPCL     EQ 2021-02-17
#> 31 ASIANPAINT     EQ       <NA>
#> 32     GRASIM     EQ       <NA>
#> 33   HDFCLIFE     EQ       <NA>
#> 34   SHREECEM     EQ       <NA>
#> 35  BRITANNIA     EQ       <NA>
#> 36 HEROMOTOCO     EQ 2021-02-15
#> 37    SBILIFE     EQ       <NA>
#> 38 BAJAJ-AUTO     EQ       <NA>
#> 39       INFY     EQ       <NA>
#> 40     MARUTI     EQ       <NA>
#> 41      CIPLA     EQ       <NA>
#> 42  COALINDIA     EQ       <NA>
#> 43        ITC     EQ 2021-02-22
#> 44    DRREDDY     EQ       <NA>
#> 45   DIVISLAB     EQ       <NA>
#> 46 HINDUNILVR     EQ       <NA>
#> 47  SUNPHARMA     EQ 2021-02-09
#> 48  NESTLEIND     EQ       <NA>
#> 49       ONGC     EQ 2021-02-17
#> 50 ADANIPORTS     EQ       <NA>
#>                                                                   caAct
#> 1                                                                  <NA>
#> 2                                                                  <NA>
#> 3                                                                  <NA>
#> 4                                                                  <NA>
#> 5                                                                  <NA>
#> 6                                                                  <NA>
#> 7                                                                  <NA>
#> 8                                                                  <NA>
#> 9                                                                  <NA>
#> 10                                                                 <NA>
#> 11                                                                 <NA>
#> 12                                                                 <NA>
#> 13                                                                 <NA>
#> 14                                                                 <NA>
#> 15                                                                 <NA>
#> 16                                                                 <NA>
#> 17                                                                 <NA>
#> 18                                                                 <NA>
#> 19                                    INTERIM DIVIDEND - RE 1 PER SHARE
#> 20                                    INTERIM DIVIDEND - RS 4 PER SHARE
#> 21                                                                 <NA>
#> 22                                    INTERIM DIVIDEND - RS 6 PER SHARE
#> 23                                                                 <NA>
#> 24                                                                 <NA>
#> 25                                                                 <NA>
#> 26                                INTERIM DIVIDEND - RS 7.50  PER SHARE
#> 27                                    INTERIM DIVIDEND - RS 3 PER SHARE
#> 28                                                                 <NA>
#> 29                                 INTERIM DIVIDEND - RS 2.50 PER SHARE
#> 30                                   INTERIM DIVIDEND - RS 16 PER SHARE
#> 31                                                                 <NA>
#> 32                                                                 <NA>
#> 33                                                                 <NA>
#> 34                                                                 <NA>
#> 35                                                                 <NA>
#> 36 INTERIM DIVIDEND - RS 65 PER SHARE SPECIAL DIVIDEND - RS 5 PER SHARE
#> 37                                                                 <NA>
#> 38                                                                 <NA>
#> 39                                                                 <NA>
#> 40                                                                 <NA>
#> 41                                                                 <NA>
#> 42                                                                 <NA>
#> 43                                    INTERIM DIVIDEND - RS 5 PER SHARE
#> 44                                                                 <NA>
#> 45                                                                 <NA>
#> 46                                                                 <NA>
#> 47                                 INTERIM DIVIDEND - RS 5.50 PER SHARE
#> 48                                                                 <NA>
#> 49                                 INTERIM DIVIDEND - RS 1.75 PER SHARE
#> 50                                                                 <NA>
#>         iep     chn perChn     pCls trdQnty    iVal    mktCap    yHigh     yLow
#> 1    751.00  -26.15  -3.36   777.15   84566  635.09  58655.24   782.50   250.85
#> 2    415.10  -13.05  -3.05   428.15   45527  188.98  41397.32   435.00   132.50
#> 3    731.10  -22.85  -3.03   753.95   57795  422.54 186945.45   799.00   286.00
#> 4    339.00   -9.50  -2.73   348.50  222386  753.89  62437.43   357.00    63.50
#> 5   1548.55  -38.30  -2.41  1586.85   60993  944.51 690446.21  1641.00   738.75
#> 6  10150.00 -245.60  -2.36 10395.60     921   93.48  62864.51 10579.00  3985.30
#> 7    618.75  -13.35  -2.11   632.10  110691  684.90 436360.17   679.40   268.30
#> 8    352.00   -7.35  -2.05   359.35  123649  435.24  52475.12   361.30    84.90
#> 9   5430.00 -112.90  -2.04  5542.90    7567  410.89 146963.58  5822.20  1783.00
#> 10  1078.00  -21.75  -1.98  1099.75   16466  177.50  62438.62  1119.50   235.55
#> 11   223.05   -4.45  -1.96   227.50   22232   49.59  58319.15   238.30   122.15
#> 12   602.00  -11.60  -1.89   613.60   15660   94.27  33754.92   616.85   240.15
#> 13  1863.90  -35.60  -1.87  1899.50    5176   96.48 278284.53  2049.00  1001.00
#> 14  2610.00  -44.90  -1.69  2654.90   16205  422.95 477117.09  2896.00  1473.45
#> 15   960.00  -16.00  -1.64   976.00    4469   42.90  60420.83  1081.55   471.40
#> 16  1473.00  -23.90  -1.60  1496.90    6007   88.48 180774.74  1593.00   661.00
#> 17   398.70   -6.35  -1.57   405.05  165134  658.39 155441.30   427.70   149.45
#> 18  1452.00  -22.60  -1.53  1474.60    9089  131.97  61529.09  1621.35   720.90
#> 19   428.90   -6.60  -1.52   435.50   26909  115.41  64714.77   467.45   159.40
#> 20   952.00  -13.25  -1.37   965.25   45035  428.73 104774.61  1067.00   375.25
#> 21   841.85  -10.20  -1.20   852.05    9358   78.78  81563.19   952.05   245.40
#> 22  3023.00  -36.10  -1.18  3059.10   16407  495.98 321409.76  3339.80  1506.05
#> 23   540.00   -6.30  -1.15   546.30   80561  435.03 131136.32   623.00   361.75
#> 24  2570.00  -29.05  -1.12  2599.05    2275   58.47  36219.42  3037.00  1245.01
#> 25  6429.00  -71.50  -1.10  6500.50     864   55.55  75051.37  6589.00  2910.00
#> 26   101.50   -1.10  -1.07   102.60   82112   83.34  26079.10   108.75    71.15
#> 27   111.50   -1.20  -1.06   112.70   87190   97.22  54640.71   114.70    73.20
#> 28  2180.00  -22.10  -1.00  2202.10  119360 2602.05 711964.20  2369.35   866.98
#> 29   145.50   -1.45  -0.99   146.95   45042   65.54  27173.38   152.80    65.00
#> 30   462.00   -4.35  -0.93   466.35   16624   76.80  37430.35   482.40   252.00
#> 31  2380.00  -22.10  -0.92  2402.10    5634  134.09 108292.18  2873.45  1431.20
#> 32  1306.00  -12.10  -0.92  1318.10    2928   38.24  51166.55  1343.00   385.05
#> 33   721.00   -6.50  -0.89   727.50    3606   26.00  57310.48   744.00   340.00
#> 34 27450.00 -245.25  -0.89 27695.25      39   10.71  36972.82 29090.00 15410.00
#> 35  3450.00  -31.15  -0.89  3481.15    1277   44.06  41086.43  4010.00  2100.00
#> 36  3398.00  -26.35  -0.77  3424.35    1749   59.43  44464.82  3629.05  1475.00
#> 37   899.70   -6.75  -0.74   906.45    5118   46.05  35353.29   954.50   519.40
#> 38  3875.00  -26.10  -0.67  3901.10     836   32.40  50798.24  4361.40  1788.65
#> 39  1335.00   -8.55  -0.64  1343.55  105459 1407.88 497891.64  1392.80   509.25
#> 40  7080.00  -44.70  -0.63  7124.70    4552  322.28  94698.11  8329.00  4001.10
#> 41   807.40   -5.10  -0.63   812.50    4941   39.89  41279.96   878.90   355.30
#> 42   154.00   -0.90  -0.58   154.90   57942   89.23  32456.63   180.20   109.55
#> 43   208.80   -1.15  -0.55   209.95  125401  261.84 183424.41   239.20   134.60
#> 44  4482.00  -20.55  -0.46  4502.55    3441  154.23  54659.89  5512.65  2495.05
#> 45  3550.00  -15.15  -0.42  3565.15    1554   55.17  45428.89  3914.95  1626.00
#> 46  2185.00   -8.50  -0.39  2193.50   17442  381.11 195842.56  2614.30  1757.30
#> 47   622.00   -2.35  -0.38   624.35   24962  155.26  67411.12   654.40   312.00
#> 48 16656.55  -42.70  -0.26 16699.25     335   55.80  59572.60 18844.00 12200.00
#> 49   113.95   -0.05  -0.04   114.00  129498  147.56  41590.40   120.50    50.00
#> 50   733.25    3.40   0.47   729.85  161134 1181.52  53383.46   754.95   203.00
#>      sumVal sumQnty finQnty sumfinQnty
#> 1  15567.69 1552461   84566    2164118
#> 2  15567.69 1552461   45527    2164118
#> 3  15567.69 1552461   57795    2164118
#> 4  15567.69 1552461  222386    2164118
#> 5  15567.69 1552461   60993    2164118
#> 6  15567.69 1552461     921    2164118
#> 7  15567.69 1552461  110691    2164118
#> 8  15567.69 1552461  123649    2164118
#> 9  15567.69 1552461    7567    2164118
#> 10 15567.69 1552461   16466    2164118
#> 11 15567.69 1552461   22232    2164118
#> 12 15567.69 1552461   15660    2164118
#> 13 15567.69 1552461    5176    2164118
#> 14 15567.69 1552461   16205    2164118
#> 15 15567.69 1552461    4469    2164118
#> 16 15567.69 1552461    6007    2164118
#> 17 15567.69 1552461  165134    2164118
#> 18 15567.69 1552461    9089    2164118
#> 19 15567.69 1552461   26909    2164118
#> 20 15567.69 1552461   45035    2164118
#> 21 15567.69 1552461    9358    2164118
#> 22 15567.69 1552461   16407    2164118
#> 23 15567.69 1552461   80561    2164118
#> 24 15567.69 1552461    2275    2164118
#> 25 15567.69 1552461     864    2164118
#> 26 15567.69 1552461   82112    2164118
#> 27 15567.69 1552461   87190    2164118
#> 28 15567.69 1552461  119360    2164118
#> 29 15567.69 1552461   45042    2164118
#> 30 15567.69 1552461   16624    2164118
#> 31 15567.69 1552461    5634    2164118
#> 32 15567.69 1552461    2928    2164118
#> 33 15567.69 1552461    3606    2164118
#> 34 15567.69 1552461      39    2164118
#> 35 15567.69 1552461    1277    2164118
#> 36 15567.69 1552461    1749    2164118
#> 37 15567.69 1552461    5118    2164118
#> 38 15567.69 1552461     836    2164118
#> 39 15567.69 1552461  105459    2164118
#> 40 15567.69 1552461    4552    2164118
#> 41 15567.69 1552461    4941    2164118
#> 42 15567.69 1552461   57942    2164118
#> 43 15567.69 1552461  125401    2164118
#> 44 15567.69 1552461    3441    2164118
#> 45 15567.69 1552461    1554    2164118
#> 46 15567.69 1552461   17442    2164118
#> 47 15567.69 1552461   24962    2164118
#> 48 15567.69 1552461     335    2164118
#> 49 15567.69 1552461  129498    2164118
#> 50 15567.69 1552461  161134    2164118

# nifty bank
nse_preopen_nifty_bank()

#>        symbol series corp_action_date corp_action   price change percent_change
#> 1    AXISBANK     EQ             <NA>        <NA>  731.10 -22.85          -3.03
#> 2    HDFCBANK     EQ             <NA>        <NA> 1548.55 -38.30          -2.41
#> 3   ICICIBANK     EQ             <NA>        <NA>  618.75 -13.35          -2.11
#> 4     RBLBANK     EQ             <NA>        <NA>  243.80  -5.15          -2.07
#> 5  INDUSINDBK     EQ             <NA>        <NA> 1078.00 -21.75          -1.98
#> 6   KOTAKBANK     EQ             <NA>        <NA> 1863.90 -35.60          -1.87
#> 7  IDFCFIRSTB     EQ             <NA>        <NA>   65.95  -1.15          -1.71
#> 8  FEDERALBNK     EQ             <NA>        <NA>   86.70  -1.40          -1.59
#> 9        SBIN     EQ             <NA>        <NA>  398.70  -6.35          -1.57
#> 10 BANDHANBNK     EQ             <NA>        <NA>  350.00  -4.90          -1.38
#> 11        PNB     EQ             <NA>        <NA>   43.00  -0.55          -1.26
#> 12 BANKBARODA     EQ             <NA>        <NA>   85.00  -0.60          -0.70
#>    prev_close quantity  value   mkt_cap year_high year_low sum_val sum_quantity
#> 1      753.95    57795 422.54 186945.45    799.00   286.00 4120.66      1600400
#> 2     1586.85    60993 944.51 690446.21   1641.00   738.75 4120.66      1600400
#> 3      632.10   110691 684.90 436360.17    679.40   268.30 4120.66      1600400
#> 4      248.95    31732  77.36  11160.81    308.10   101.55 4120.66      1600400
#> 5     1099.75    16466 177.50  62438.62   1119.50   235.55 4120.66      1600400
#> 6     1899.50     5176  96.48 278284.53   2049.00  1001.00 4120.66      1600400
#> 7       67.10   160479 105.84  14082.89     68.15    17.65 4120.66      1600400
#> 8       88.10    80787  70.04  17590.73     88.50    35.70 4120.66      1600400
#> 9      405.05   165134 658.39 155441.30    427.70   149.45 4120.66      1600400
#> 10     354.90    86055 301.19  24003.88    430.70   152.20 4120.66      1600400
#> 11      43.55   694341 298.57  10494.97     46.35    26.30 4120.66      1600400
#> 12      85.60   333343 283.34  10283.53     99.85    36.00 4120.66      1600400
#>    fin_quantity sum_fin_quantity
#> 1         57795          1802992
#> 2         60993          1802992
#> 3        110691          1802992
#> 4         31732          1802992
#> 5         16466          1802992
#> 6          5176          1802992
#> 7        160479          1802992
#> 8         80787          1802992
#> 9        165134          1802992
#> 10        86055          1802992
#> 11       694341          1802992
#> 12       333343          1802992

# retain original column names as returned by NSE
nse_preopen_nifty_bank(clean_names = FALSE)

#>        symbol series  xDt caAct     iep    chn perChn    pCls trdQnty   iVal
#> 1    AXISBANK     EQ <NA>  <NA>  731.10 -22.85  -3.03  753.95   57795 422.54
#> 2    HDFCBANK     EQ <NA>  <NA> 1548.55 -38.30  -2.41 1586.85   60993 944.51
#> 3   ICICIBANK     EQ <NA>  <NA>  618.75 -13.35  -2.11  632.10  110691 684.90
#> 4     RBLBANK     EQ <NA>  <NA>  243.80  -5.15  -2.07  248.95   31732  77.36
#> 5  INDUSINDBK     EQ <NA>  <NA> 1078.00 -21.75  -1.98 1099.75   16466 177.50
#> 6   KOTAKBANK     EQ <NA>  <NA> 1863.90 -35.60  -1.87 1899.50    5176  96.48
#> 7  IDFCFIRSTB     EQ <NA>  <NA>   65.95  -1.15  -1.71   67.10  160479 105.84
#> 8  FEDERALBNK     EQ <NA>  <NA>   86.70  -1.40  -1.59   88.10   80787  70.04
#> 9        SBIN     EQ <NA>  <NA>  398.70  -6.35  -1.57  405.05  165134 658.39
#> 10 BANDHANBNK     EQ <NA>  <NA>  350.00  -4.90  -1.38  354.90   86055 301.19
#> 11        PNB     EQ <NA>  <NA>   43.00  -0.55  -1.26   43.55  694341 298.57
#> 12 BANKBARODA     EQ <NA>  <NA>   85.00  -0.60  -0.70   85.60  333343 283.34
#>       mktCap   yHigh    yLow  sumVal sumQnty finQnty sumfinQnty
#> 1  186945.45  799.00  286.00 4120.66 1600400   57795    1802992
#> 2  690446.21 1641.00  738.75 4120.66 1600400   60993    1802992
#> 3  436360.17  679.40  268.30 4120.66 1600400  110691    1802992
#> 4   11160.81  308.10  101.55 4120.66 1600400   31732    1802992
#> 5   62438.62 1119.50  235.55 4120.66 1600400   16466    1802992
#> 6  278284.53 2049.00 1001.00 4120.66 1600400    5176    1802992
#> 7   14082.89   68.15   17.65 4120.66 1600400  160479    1802992
#> 8   17590.73   88.50   35.70 4120.66 1600400   80787    1802992
#> 9  155441.30  427.70  149.45 4120.66 1600400  165134    1802992
#> 10  24003.88  430.70  152.20 4120.66 1600400   86055    1802992
#> 11  10494.97   46.35   26.30 4120.66 1600400  694341    1802992
#> 12  10283.53   99.85   36.00 4120.66 1600400  333343    1802992

# }