Pre Open market data

Fetch data of pre open session of Nifty & Nifty Bank.

nse_preopen_nifty(clean_names = TRUE)

nse_preopen_nifty_bank(clean_names = TRUE)

Arguments

clean_names
Logical; if TRUE, makes the column names descriptive and uses snake_case.

Value

A tibble with the following columns:

symbol

NSE ticker.

series

Equity (EQ)

corp_action_date

Corporate action date.

corp_action

Corporate Action

price

Price

change

Change in price

percent_change

Percentage change in price.

prev_close

Previous close.

quantity

Quantity

value

Value (in lakhs),

mkt_cap

Free float market capitalization (in crores).

year_high

Normal market 52 week high.

year_low

Normal market 52 week low.

Examples

# \donttest{
# nifty
nse_preopen_nifty()

#> # A tibble: 50 x 17
#>    symbol series corp_action_date corp_action price change percent_change
#>    <chr>  <chr>  <date>           <chr>       <dbl>  <dbl>          <dbl>
#>  1 TATAM~ EQ     NA               NA           180.  -4.5           -2.44
#>  2 BPCL   EQ     NA               NA           450  -11             -2.39
#>  3 SBIN   EQ     NA               NA           312.  -6.3           -1.98
#>  4 VEDL   EQ     NA               NA           153.  -2.75          -1.76
#>  5 ASIAN~ EQ     2019-10-30       INTERIM DI~ 1696  -28.4           -1.65
#>  6 UPL    EQ     NA               NA           586.  -9.3           -1.56
#>  7 ADANI~ EQ     NA               NA           379   -5.8           -1.51
#>  8 JSWST~ EQ     NA               NA           261.  -3.85          -1.45
#>  9 AXISB~ EQ     NA               NA           715. -10.5           -1.45
#> 10 BAJFI~ EQ     NA               NA          3950  -57.6           -1.44
#> # ... with 40 more rows, and 10 more variables: prev_close <dbl>,
#> #   quantity <dbl>, value <dbl>, mkt_cap <dbl>, year_high <dbl>,
#> #   year_low <dbl>, sum_val <dbl>, sum_quantity <dbl>, fin_quantity <dbl>,
#> #   sum_fin_quantity <dbl>

# retain original column names as returned by NSE
nse_preopen_nifty(clean_names = FALSE)

#> # A tibble: 50 x 17
#>    symbol series xDt        caAct   iep    chn perChn  pCls trdQnty  iVal mktCap
#>    <chr>  <chr>  <date>     <chr> <dbl>  <dbl>  <dbl> <dbl>   <dbl> <dbl>  <dbl>
#>  1 TATAM~ EQ     NA         NA     180.  -4.5   -2.44  185.   71141 128.  3.31e4
#>  2 BPCL   EQ     NA         NA     450  -11     -2.39  461    41761 188.  3.70e4
#>  3 SBIN   EQ     NA         NA     312.  -6.3   -1.98  318.  191444 598.  1.22e5
#>  4 VEDL   EQ     NA         NA     153.  -2.75  -1.76  156    67628 104.  2.84e4
#>  5 ASIAN~ EQ     2019-10-30 INTE~ 1696  -28.4   -1.65 1724.    9393 159.  7.77e4
#>  6 UPL    EQ     NA         NA     586.  -9.3   -1.56  595.    4609  27   3.27e4
#>  7 ADANI~ EQ     NA         NA     379   -5.8   -1.51  385.   10580  40.1 2.97e4
#>  8 JSWST~ EQ     NA         NA     261.  -3.85  -1.45  265.   15277  39.9 2.69e4
#>  9 AXISB~ EQ     NA         NA     715. -10.5   -1.45  726.   21060 151.  1.62e5
#> 10 BAJFI~ EQ     NA         NA    3950  -57.6   -1.44 4008.    5710 226.  1.04e5
#> # ... with 40 more rows, and 6 more variables: yHigh <dbl>, yLow <dbl>,
#> #   sumVal <dbl>, sumQnty <dbl>, finQnty <dbl>, sumfinQnty <dbl>

# nifty bank
nse_preopen_nifty_bank()

#> # A tibble: 12 x 17
#>    symbol series corp_action_date corp_action  price change percent_change
#>    <chr>  <chr>  <date>           <chr>        <dbl>  <dbl>          <dbl>
#>  1 RBLBA~ EQ     NA               NA           335    -9             -2.62
#>  2 BANKB~ EQ     NA               NA            94    -2.15          -2.24
#>  3 SBIN   EQ     NA               NA           312.   -6.3           -1.98
#>  4 AXISB~ EQ     NA               NA           715.  -10.5           -1.45
#>  5 INDUS~ EQ     NA               NA          1441   -20.6           -1.41
#>  6 PNB    EQ     NA               NA            61    -0.85          -1.37
#>  7 ICICI~ EQ     NA               NA           516    -6.9           -1.32
#>  8 IDFCF~ EQ     NA               NA            43    -0.5           -1.15
#>  9 FEDER~ EQ     NA               NA            86.4  -0.95          -1.09
#> 10 HDFCB~ EQ     NA               NA          1247.  -13.6           -1.08
#> 11 YESBA~ EQ     NA               NA            46.2   1.2            2.66
#> 12 KOTAK~ EQ     NA               NA          1655.  -16.2           -0.97
#> # ... with 10 more variables: prev_close <dbl>, quantity <dbl>, value <dbl>,
#> #   mkt_cap <dbl>, year_high <dbl>, year_low <dbl>, sum_val <dbl>,
#> #   sum_quantity <dbl>, fin_quantity <dbl>, sum_fin_quantity <dbl>

# retain original column names as returned by NSE
nse_preopen_nifty_bank(clean_names = FALSE)

#> # A tibble: 12 x 17
#>    symbol series xDt        caAct    iep    chn perChn   pCls trdQnty   iVal
#>    <chr>  <chr>  <date>     <chr>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1 RBLBA~ EQ     NA         NA     335    -9     -2.62  344     10557  35.4 
#>  2 BANKB~ EQ     NA         NA      94    -2.15  -2.24   96.2   89008  83.7 
#>  3 SBIN   EQ     NA         NA     312.   -6.3   -1.98  318.   191444 598.  
#>  4 AXISB~ EQ     NA         NA     715.  -10.5   -1.45  726.    21060 151.  
#>  5 INDUS~ EQ     NA         NA    1441   -20.6   -1.41 1462.     7533 109.  
#>  6 PNB    EQ     NA         NA      61    -0.85  -1.37   61.8   59892  36.5 
#>  7 ICICI~ EQ     NA         NA     516    -6.9   -1.32  523.    72395 374.  
#>  8 IDFCF~ EQ     NA         NA      43    -0.5   -1.15   43.5   57043  24.5 
#>  9 FEDER~ EQ     NA         NA      86.4  -0.95  -1.09   87.4    9282   8.02
#> 10 HDFCB~ EQ     NA         NA    1247.  -13.6   -1.08 1261.    27486 343.  
#> 11 YESBA~ EQ     NA         NA      46.2   1.2    2.66   45.0  753830 349.  
#> 12 KOTAK~ EQ     NA         NA    1655.  -16.2   -0.97 1671.     4373  72.4 
#> # ... with 7 more variables: mktCap <dbl>, yHigh <dbl>, yLow <dbl>,
#> #   sumVal <dbl>, sumQnty <dbl>, finQnty <dbl>, sumfinQnty <dbl>

# }